+++ docs/joss_paper/paper.md
@@ -14,14 +14,14 @@ authors:
 - name: Xavier Roynard
   orcid: 0000-0001-7840-2120
   affiliation: "1"
-- name: Alexandre Devaux--Rivière
-  orcid: 0009-0001-7474-944X
-  affiliation: ["1", "2"]
+# - name: Alexandre Devaux--Rivière
+#   orcid: 0009-0001-7474-944X
+#   affiliation: ["1", "2"]
 affiliations:
 - name: SafranTech, Safran Tech, Digital Sciences & Technologies, 78114 Magny-Les-Hameaux, France
   index: 1
-- name: EPITA, 14-16 Rue Voltaire, 94270 Le Kremlin-Bicêtre, France
-  index: 2
+# - name: EPITA, 14-16 Rue Voltaire, 94270 Le Kremlin-Bicêtre, France
+#   index: 2
 bibliography: paper.bib
 ---
 
