### 2025-06-12 65a284a
action paper
