+++ .github/workflows/draft-pdf.yml
@@ -13,15 +13,18 @@ jobs:
       - name: Checkout
         uses: actions/checkout@v4
       - name: Build joss draft PDF
+        id: generate-files
         uses: xuanxu/paper-action@main
         with:
           repository_url: http://github.com/PLAID-lib/plaid
           branch: action_paper
           issue_id: 8393
           journal: joss
-      - name: Upload
-        uses: actions/upload-artifact@v4
+      - name: Upload PDF
+        id: upload-pdf
+        uses: xuanxu/upload-files-action@main
         with:
-          name: paper
-          path: docs/joss_paper/paper.pdf
-
+          papers_repo: openjournals/joss-papers
+          branch_prefix: joss
+          issue_id: 8393
+          pdf_path: ${{ steps.generate-files.outputs.pdf_path}}
