### 2025-06-12 fd88db2
joss paper: add doi (PLMR and Neurips do not have DOIs)
