+++ docs/joss_paper/paper.bib
@@ -2,7 +2,8 @@
   title={Physics-Learning AI Datamodel (PLAID) datasets: a collection of physics simulations for machine learning},
   author={Casenave, Fabien and Roynard, Xavier and Staber, Brian and Akkari, Nissrine and Piat, William and Bucci, Michele Alessandro and Kabalan, Abbas and Nguyen, Xuan Minh Vuong and Saverio, Luca and Perez, Rapha{\"e}l Carpintero and others},
   journal={arXiv preprint arXiv:2505.02974},
-  year={2025}
+  year={2025},
+  doi={https://doi.org/10.48550/arXiv.2505.02974}
 }
 
 @inproceedings{poinot2018seven,
@@ -10,7 +11,8 @@
   author={Poinot, Marc and Rumsey, Christopher L},
   booktitle={2018 AIAA Aerospace Sciences Meeting},
   pages={1503},
-  year={2018}
+  year={2018},
+  doi={https://doi.org/10.2514/6.2018-1503}
 }
 
 @article{casenave2024mmgp,
@@ -34,7 +36,8 @@
   title={{O-MMGP}: {O}ptimal {M}esh {M}orphing {G}aussian {P}rocess Regression for Solving {PDEs} with non-Parametric Geometric Variations},
   author={Kabalan, Abbas and Casenave, Fabien and Bordeu, Felipe and Ehrlacher, Virginie},
   journal={arXiv preprint arXiv:2502.11632},
-  year={2025}
+  year={2025},
+  doi={https://doi.org/10.48550/arXiv.2502.11632},
 }
 
 @inproceedings{perez2024gaussian,
@@ -50,5 +53,6 @@
   title={{Learning signals defined on graphs with optimal transport and Gaussian process regression}},
   author={Carpintero Perez, Rapha{\"e}l and Da Veiga, S{\'e}bastien and Garnier, Josselin and Staber, Brian},
   journal={arXiv preprint arXiv:2410.15721},
-  year={2024}
+  year={2024},
+  doi={https://doi.org/10.48550/arXiv.2410.15721},
 }
