### 2025-06-12 b1cba27
feat(joss_paper) add a missing doi in references
