+++ docs/joss_paper/paper.bib
@@ -22,6 +22,7 @@
   volume={36},
   year={2024}
 }
+
 @article{kabalan2025elasticity,
   title={Elasticity-based morphing technique and application to reduced-order modeling},
   author={Kabalan, Abbas and Casenave, Fabien and Bordeu, Felipe and Ehrlacher, Virginie and Ern, Alexandre},
@@ -29,7 +30,8 @@
   volume={141},
   pages={115929},
   year={2025},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={https://doi.org/10.1016/j.apm.2025.115929}
 }
 
 @article{kabalan2025ommgp,
