### 2025-06-12 acf24e6
🐛 fix(joss_paper) update action, debug affiliations and add DOIs (#63)
---------

Co-authored-by: Fabien Casenave <fabien.casenave@safrangroup.com>
