### 2025-06-12 1d4e69a
fix(action joss paper) correct branch name
