+++ .github/workflows/draft-pdf.yml
@@ -17,7 +17,7 @@ jobs:
         uses: xuanxu/paper-action@main
         with:
           repository_url: http://github.com/PLAID-lib/plaid
-          branch: action_paper
+          branch: ${{ github.ref_name }}
           issue_id: 8393
           journal: joss
       - name: Upload
