### 2025-06-12 18cce28
fix(joss_paper) remove https://doi.org/ in doi entrys in paper.bib
