+++ docs/joss_paper/paper.bib
@@ -3,7 +3,7 @@
   author={Casenave, Fabien and Roynard, Xavier and Staber, Brian and Akkari, Nissrine and Piat, William and Bucci, Michele Alessandro and Kabalan, Abbas and Nguyen, Xuan Minh Vuong and Saverio, Luca and Perez, Rapha{\"e}l Carpintero and others},
   journal={arXiv preprint arXiv:2505.02974},
   year={2025},
-  doi={https://doi.org/10.48550/arXiv.2505.02974}
+  doi={10.48550/arXiv.2505.02974}
 }
 
 @inproceedings{poinot2018seven,
@@ -12,7 +12,7 @@
   booktitle={2018 AIAA Aerospace Sciences Meeting},
   pages={1503},
   year={2018},
-  doi={https://doi.org/10.2514/6.2018-1503}
+  doi={10.2514/6.2018-1503}
 }
 
 @article{casenave2024mmgp,
@@ -31,7 +31,7 @@
   pages={115929},
   year={2025},
   publisher={Elsevier},
-  doi={https://doi.org/10.1016/j.apm.2025.115929}
+  doi={10.1016/j.apm.2025.115929}
 }
 
 @article{kabalan2025ommgp,
@@ -39,7 +39,7 @@
   author={Kabalan, Abbas and Casenave, Fabien and Bordeu, Felipe and Ehrlacher, Virginie},
   journal={arXiv preprint arXiv:2502.11632},
   year={2025},
-  doi={https://doi.org/10.48550/arXiv.2502.11632},
+  doi={10.48550/arXiv.2502.11632},
 }
 
 @inproceedings{perez2024gaussian,
@@ -56,5 +56,5 @@
   author={Carpintero Perez, Rapha{\"e}l and Da Veiga, S{\'e}bastien and Garnier, Josselin and Staber, Brian},
   journal={arXiv preprint arXiv:2410.15721},
   year={2024},
-  doi={https://doi.org/10.48550/arXiv.2410.15721},
+  doi={10.48550/arXiv.2410.15721},
 }
