### 2025-06-12 #8393 @casenave
Skipped DOIs correspond to a Neurips paper and a PLMR paper, for which I could not find a DOI
