### 2025-06-12 f7b3c94
🚀 Create dependabot.yml (#67)
