+++ SECURITY.md
@@ -1,0 +1,21 @@
+# Security Policy
+
+This security policy applies to public projects under the [plaid] project on GitHub.
+
+## Security/Bugfix Versions
+
+Security and bug fixes are generally provided only for the last minor version.
+
+Security fixes are given priority and might be enough to cause a new version to be released.
+
+## Reporting a Vulnerability
+
+We encourage responsible disclosure of security vulnerabilities.
+If you find something suspicious, we encourage and appreciate your report!
+
+### Ways to report
+
+In order for the vulnerability reports to reach maintainers as soon as possible, the preferred way is to use the "Report a vulnerability" button under the "Security" tab of the GitHub [plaid] project.
+This creates a private communication channel between the reporter and the maintainers.
+
+[plaid]: https://github.com/PLAID-lib/plaid
