### 2025-06-12 d0ce4db
📄 fix(SECURITY.md) correct first sentence (#69)
Co-authored-by: Fabien Casenave <fabien.casenave@safrangroup.com>
