### 2025-06-12 4268423
🐛 fix(dependabot) remove conda package-ecosystem (#70)
Co-authored-by: Fabien Casenave <fabien.casenave@safrangroup.com>
