### 2025-06-12 50dbc21
fix(actions) restrict permission of action configuration files
