### 2025-06-12 9642a03
🚀 Potential fix for code scanning alert no. 8: Workflow does not contain permissions (#71)
* Potential fix for code scanning alert no. 8: Workflow does not contain permissions

Co-authored-by: Copilot Autofix powered by AI <62310815+github-advanced-security[bot]@users.noreply.github.com>

* fix(actions) restrict permission of action configuration files

---------

Co-authored-by: Copilot Autofix powered by AI <62310815+github-advanced-security[bot]@users.noreply.github.com>
Co-authored-by: Fabien Casenave <fabien.casenave@safrangroup.com>
