### 2025-06-12 80a1a4e
📦 fix(env) rename conda_dev_env.yml -> environment.yml (#72)
Co-authored-by: Fabien Casenave <fabien.casenave@safrangroup.com>
