+++ .readthedocs.yml
@@ -19,3 +19,25 @@ python:
   install:
     - method: pip
       path: .
+
+
+
+# version: 2
+
+# build:
+#   os: "ubuntu-22.04"
+#   tools:
+#     python: "3.11"
+#   jobs:
+#     post_install:
+#       # Install poetry
+#       # https://python-poetry.org/docs/#installing-manually
+#       - pip install poetry
+#       # Install dependencies with 'docs' dependency group
+#       # https://python-poetry.org/docs/managing-dependencies/#dependency-groups
+#       # VIRTUAL_ENV needs to be set manually for now.
+#       # See https://github.com/readthedocs/readthedocs.org/pull/11152/
+#       - VIRTUAL_ENV=$READTHEDOCS_VIRTUALENV_PATH poetry install --with docs
+
+# sphinx:
+#   configuration: docs/conf.py
