+++ pyproject.toml
@@ -1,84 +1,40 @@
-[build-system]
-requires = [
-    "setuptools>=60",
-    "setuptools-scm>=8.0"]
-build-backend = "setuptools.build_meta"
-
-[project]
+[tool.poetry]
 name = "pyplaid"
-authors = [{name = "Safran", email = "fabien.casenave@safrangroup.com"}]
+version = "0.1.5"
 description = "A package that implements a data model tailored for AI and ML in the context of physics problems"
-requires-python = ">=3.11, <3.13"
-keywords=[
-    "machine learning",
-    "physics",
-    "data-model",
-    "Safran",
-]
-classifiers = [
-    "Intended Audience :: Science/Research",
-    "License :: OSI Approved :: BSD License",
-    "Operating System :: MacOS",
-    "Operating System :: POSIX",
-    "Operating System :: Microsoft :: Windows",
-    "Programming Language :: Python :: 3.11",
-    "Programming Language :: Python :: 3.12",
-    "Topic :: Scientific/Engineering",
-    "Topic :: Scientific/Engineering :: Artificial Intelligence",
-    "Topic :: Scientific/Engineering :: Mathematics",
-]
-dependencies = [
-    "tqdm",
-    "pyyaml",
-    "pycgns",
-    "numpy",
-    "matplotlib",
-    "rich",
-    "scikit-learn",
-    "datasets >=3.0",
-    "pydantic",
+authors = ["Safran <fabien.casenave@safrangroup.com>"]
+license = "BSD-3-Clause"
+readme = "README.md"
+packages = [
+    { include = "plaid", from = "src" }
 ]
 
-dynamic = ["version"]
+[tool.poetry.dependencies]
+python = ">=3.11,<3.13"
+tqdm = "*"
+pyyaml = "*"
+pycgns-core = "*"
+numpy = "*"
+matplotlib-base = "*"
+rich = "*"
+scikit-learn = "*"
+datasets = ">=3.0"
+pydantic = "*"
 
-[tool.setuptools_scm]
-write_to = "src/plaid/_version.py"
+[tool.poetry.group.dev.dependencies]
+pre-commit = "*"
+ruff = "*"
+pytest = "*"
+pytest-cov = "*"
 
-[project.urls]
-homepage = "https://plaid-lib.github.io"
-documentation = "https://plaid-lib.readthedocs.io"
-repository = "https://github.com/PLAID-lib/plaid"
+[tool.poetry.group.docs.dependencies]
+sphinx = "*"
+myst-nb = "*"
+myst-parser = "*"
+furo = "*"
+sphinx-autoapi = "*"
+sphinx-tabs = "*"
+sphinxcontrib-bibtex = "*"
 
-[project.license]
-text="BSD 3-Clause License"
-files=["LICENSE.txt"]
-
-[project.readme]
-file="README.md"
-content-type = "text/markdown"
-
-[tool.setuptools]
-platforms = [
-    "Linux",
-    "Mac OS-X",
-    "Unix",
-]
-
-[tool.setuptools.packages.find]
-where = ["src"]
-include = ["plaid*"]
-
-[project.optional-dependencies]
-dev = [
-    "pre-commit",
-    "ruff",
-    "pytest",
-    "pytest-cov",
-    "sphinx",
-    "myst-nb",
-    "myst-parser",
-    "furo",
-    "sphinx-autoapi",
-    "sphinx-tabs",
-    "sphinxcontrib.bibtex",
-]
+[build-system]
+requires = ["poetry"]
