### 2025-06-12 dc62863
feat(pyproject) config pyproject
