+++ .github/workflows/testing.yml
@@ -31,32 +31,41 @@ jobs:
 
       - name: Create environment
         run: |
-          mamba env create -f .github/workflows/conda_envs/python${{ matrix.python-version }}.yml
-          mamba run -n plaid_python${{ matrix.python-version }} pip install -e .
+          conda create -n plaid-dev python=${{ matrix.python-version }} muscat=2.4.1 poetry
+          conda activate plaid-dev
+          poetry install
 
-      - name: Run tests on Unix
-        if: runner.os != 'Windows'
+      - name: Run tests on Linux
+        if: runner.os == 'Linux'
+        shell: bash
         run: |
-          mamba run -n plaid_python${{ matrix.python-version }} pytest tests --cov=src --cov-report=xml --cov-report=term
+          conda run -n plaid-dev pytest tests --cov=src --cov-report=xml --cov-report=term
+
+      - name: Run tests on macOS
+        if: runner.os == 'macOS'
+        shell: bash
+        run: |
+          conda run -n plaid-dev pytest tests
 
       - name: Run tests on Windows
         if: runner.os == 'Windows'
         shell: powershell
         run: |
-          mamba run -n plaid_python${{ matrix.python-version }} pytest tests
+          mamba run -n plaid-dev pytest tests
 
       - name: Run examples on Unix
         if: runner.os != 'Windows'
+        shell: bash
         run: |
           cd examples
-          mamba run -n plaid_python${{ matrix.python-version }} bash run_examples.sh
+          mamba run -n plaid-dev bash run_examples.sh
 
       - name: Run examples on Windows
         if: runner.os == 'Windows'
         shell: powershell
         run: |
           cd examples
-          mamba run -n plaid_python${{ matrix.python-version }} run_examples.bat
+          mamba run -n plaid-dev run_examples.bat
 
       - name: Upload coverage to Codecov (Linux only)
         if: runner.os == 'Linux'
