### 2025-06-12 73a214a
fix(actions) remove conda activate
