+++ .github/workflows/testing.yml
@@ -32,7 +32,6 @@ jobs:
       - name: Create environment
         run: |
           conda create -n plaid-dev python=${{ matrix.python-version }} muscat=2.4.1 poetry
-          conda activate plaid-dev
           poetry install
 
       - name: Run tests on Linux
