### 2025-06-12 6c73467
fix(actions) remove conda activate
