+++ .github/workflows/doc.yml
@@ -17,7 +17,7 @@ jobs:
       - name: Checkout code
         uses: actions/checkout@v3
 
-      - name: Set up Conda with Mamba
+      - name: Set up Conda
         uses: conda-incubator/setup-miniconda@v2
         with:
           use-mamba: true
@@ -27,7 +27,7 @@ jobs:
       - name: Create environment
         run: |
           conda create -n plaid-dev python=3.11 muscat=2.4.1 poetry
-          poetry install --with docs
+          conda run -n plaid-dev poetry install --with docs
 
       - name: Compile documentation
         run: |
