### 2025-06-12 8070107
fix(actions) simplify CI
