+++ .github/workflows/testing.yml
@@ -22,13 +22,6 @@ jobs:
       - name: Checkout code
         uses: actions/checkout@v3
 
-      - name: Set up Conda
-        uses: conda-incubator/setup-miniconda@v2
-        with:
-          use-mamba: true
-          auto-activate-base: false
-          miniforge-variant: Miniforge3
-
       - name: Create environment
         run: |
           conda create -n plaid-dev python=${{ matrix.python-version }} muscat=2.4.1 poetry
