### 2025-06-12 91337cf
feat(actions) simplify workflows
