### 2025-06-12 b77acd0
feat(actions) simplify workflows
