### 2025-06-12 3c96f9c
feat(actions) simplify workflows
