### 2025-06-12 35a817d
feat(actions) simplify workflows
