+++ .github/workflows/testing.yml
@@ -22,6 +22,11 @@ jobs:
       - name: Checkout code
         uses: actions/checkout@v3
 
+      - name: Install HDF5 dev libraries
+        run: |
+          sudo apt-get update
+          sudo apt-get install -y libhdf5-dev
+
       - name: Set up Conda
         uses: conda-incubator/setup-miniconda@v2
         with:
