### 2025-06-13 3583747
pixi: start testing
