+++ .readthedocs.yml
@@ -8,7 +8,7 @@ sphinx:
   configuration: docs/conf.py
 
 conda:
-  environment: docs/requirements.yml
+  environment: environment.yml
 
 build:
   os: "ubuntu-20.04"
