+++ .github/workflows/testing.yml
@@ -33,7 +33,7 @@ jobs:
           python_version=${{ matrix.python-version }}
           sed -i.bak "s/^  - python=.*/  - python=${python_version}/" environment.yml
 
-          - name: Modify environment.yml with matrix Python version (Windows)
+      - name: Modify environment.yml with matrix Python version (Windows)
         if: runner.os == 'Windows'
         shell: powershell
         run: |
