+++ .github/workflows/testing.yml
@@ -26,6 +26,7 @@ jobs:
         uses: conda-incubator/setup-miniconda@v2
         with:
           use-mamba: false
+          python-version: 3.11
           auto-activate-base: false
           miniforge-variant: Miniforge3
 
