+++ .github/workflows/publish-pypi.yml
@@ -73,7 +73,7 @@ jobs:
       - name: Set up Python
         uses: actions/setup-python@v4
         with:
-          python-version: "3.12"
+          python-version: "3.11"
 
       - name: Install build tools
         run: |
@@ -96,7 +96,7 @@ jobs:
 
       - uses: actions/upload-artifact@v4
         with:
-          name: wheels-3.12
+          name: wheels-3.11
           path: dist/*.whl
 
   publish:
