+++ .readthedocs.yml
@@ -11,7 +11,7 @@ conda:
   environment: environment.yml
 
 build:
-  os: "ubuntu-20.04"
+  os: "ubuntu-22.04"
   tools:
     python: "mambaforge-4.10"
 
