### 2025-06-17 5d57c1e
🚀 Simplify environments in actions and doc building (unique dev env reference) (#73)
Co-authored-by: Fabien Casenave <fabien.casenave@safrangroup.com>
