+++ .github/workflows/conda_envs/python3.12.yml
@@ -1,17 +1,0 @@
-name: plaid_python3.12
-channels:
-  - conda-forge
-  - nodefaults
-dependencies:
-  - python=3.12
-#---# io
-  - pyyaml
-#---# mesh/graph libs
-  - muscat=2.4.1
-  - pycgns-core
-#---# other
-  - scikit-learn
-  - datasets
-  - pydantic
-  - pytest
-  - pytest-cov
