+++ src/plaid/bridges/huggingface_bridge.py
@@ -1,7 +1,14 @@
 """Huggingface bridge for PLAID datasets."""
 
 import pickle
-from typing import Callable, Self
+import sys
+from typing import Callable
+
+if sys.version_info >= (3, 11):
+    from typing import Self
+else: # pragma: no cover
+    from typing import TypeVar
+    Self = TypeVar("Self")
 
 import datasets
 
