+++ src/plaid/post/bisect.py
@@ -6,6 +6,7 @@ import matplotlib as mpl
 import matplotlib.pyplot as plt
 import numpy as np
 from tqdm import tqdm
+from typing import Union
 
 from plaid.containers.dataset import Dataset
 from plaid.problem_definition import ProblemDefinition
@@ -77,10 +78,10 @@ def is_dvipng_available(verbose: bool) -> bool:
 
 
 def plot_bisect(
-    ref_dataset: Dataset | str,
-    pred_dataset: Dataset | str,
-    problem_def: ProblemDefinition | str,
-    scalar: str | int,
+    ref_dataset: Union[Dataset, str],
+    pred_dataset: Union[Dataset, str],
+    problem_def: Union[ProblemDefinition, str],
+    scalar: Union[str, int],
     save_file_name: str = "bissec_plots",
     verbose: bool = False,
 ) -> None:
