+++ .github/workflows/testing.yml
@@ -16,7 +16,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
-        python-version: [3.11, 3.12]
+        python-version: [3.9, 3.10, 3.11, 3.12]
 
     steps:
       - name: Checkout code
