+++ src/plaid/problem_definition.py
@@ -10,10 +10,12 @@
 # %% Imports
 
 import sys
+
 if sys.version_info >= (3, 11):
     from typing import Self
-else: # pragma: no cover
+else:  # pragma: no cover
     from typing import TypeVar
+
     Self = TypeVar("Self")
 
 import csv
