+++ src/plaid/containers/dataset.py
@@ -9,10 +9,12 @@
 
 # %% Imports
 import sys
+
 if sys.version_info >= (3, 11):
     from typing import Self
-else: # pragma: no cover
+else:  # pragma: no cover
     from typing import TypeVar
+
     Self = TypeVar("Self")
 
 import logging
