### 2025-06-17 7c09021
fix(action) typo in testing.yml
