### 2025-06-17 2a2d3e1
fix(action) typo in testing.yml
