+++ .github/workflows/testing.yml
@@ -34,7 +34,7 @@ jobs:
         if: runner.os != 'Windows'
         run: |
           python_version='${{ matrix.python-version }}'
-          sed -i.bak "s/^  - python=.*/  - python=${python_version}/" environment.yml
+          sed -i.bak "s/^  - python=.*/  - python=\"${python_version}\"/" environment.yml
 
       - name: Modify environment.yml with matrix Python version (Windows)
         if: runner.os == 'Windows'
