### 2025-06-17 d73718f
fix(action) typo in testing.yml
