+++ .github/workflows/testing.yml
@@ -33,14 +33,15 @@ jobs:
       - name: Modify environment.yml with matrix Python version (Linux/macOS)
         if: runner.os != 'Windows'
         run: |
-          python_version='${{ matrix.python-version }}'
-          sed -i.bak "s/^  - python=.*/  - python=\"${python_version}\"/" environment.yml
+          python_version="${{ matrix.python-version }}"
+          echo "Setting Python version to ${python_version}"
+          sed -i.bak "s/^  - python=.*/  - python=${python_version}/" environment.yml
 
       - name: Modify environment.yml with matrix Python version (Windows)
         if: runner.os == 'Windows'
         shell: powershell
         run: |
-          $python_version = '${{ matrix.python-version }}'
+          $python_version = "${{ matrix.python-version }}"
           $file = 'environment.yml'
           (Get-Content $file) -replace '^\s+- python=.*', "  - python=$python_version" | Set-Content $file
 
