### 2025-06-17 8d9aac7
fix(action) typo in testing.yml
