+++ .github/workflows/testing.yml
@@ -33,17 +33,14 @@ jobs:
       - name: Modify environment.yml with matrix Python version (Linux/macOS)
         if: runner.os != 'Windows'
         run: |
-          python_version="${{ matrix.python-version }}"
-          echo "Setting Python version to ${python_version}"
-          sed -i.bak "s/^  - python=.*/  - python=${python_version}/" environment.yml
+          sed -i.bak "s/^  - python=.*/  - python=${{ matrix.python-version }}/" environment.yml
 
       - name: Modify environment.yml with matrix Python version (Windows)
         if: runner.os == 'Windows'
         shell: powershell
         run: |
-          $python_version = "${{ matrix.python-version }}"
           $file = 'environment.yml'
-          (Get-Content $file) -replace '^\s+- python=.*', "  - python=$python_version" | Set-Content $file
+          (Get-Content $file) -replace '^\s+- python=.*', "  - python=${{ matrix.python-version }}" | Set-Content $file
 
       - name: Create environment
         run: |
