### 2025-06-17 a4cc152
feat(action) add sensitive info detection
