### 2025-06-17 6577b8c
feat(.hooks/pre-commit) configure pre-commit for confidential info check
