+++ .pre_commit_hooks/patterns.txt
@@ -1,0 +1,3 @@
+/data/
+/gpfs/
+/apps/
