+++ .pre-commit-config.yaml
@@ -21,3 +21,9 @@ repos:
         language: system
         types: [python]
         pass_filenames: false
+
+      - id: check-confidential-patterns
+        name: Confidential Pattern Check
+        entry: python .pre_commit_hooks/check_confidential_patterns.py
+        language: system
+        types: [python]
