+++ .github/workflows/detect_sensitive_info.yml
@@ -1,98 +1,0 @@
-name: Detect Sensitive Info
-
-permissions:
-  contents: read
-
-on:
-  push:
-  pull_request:
-
-jobs:
-  check-pattern:
-    runs-on: ubuntu-latest
-
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
-        with:
-          fetch-depth: 0
-
-      - name: Get commit range
-        id: range
-        run: |
-          if [ "${{ github.event_name }}" = "pull_request" ]; then
-            echo "range=${{ github.event.pull_request.base.sha }}..${{ github.event.pull_request.head.sha }}" >> $GITHUB_OUTPUT
-          else
-            echo "range=${{ github.event.before }}..${{ github.sha }}" >> $GITHUB_OUTPUT
-          fi
-
-      - name: Scan code and commits for sensitive patterns
-        run: |
-          set -e
-          range="${{ steps.range.outputs.range }}"
-          echo "Scanning commits in $range"
-
-          has_violation=0
-
-          echo '${{ secrets.SENSITIVE_PATTERNS }}' > config.json
-
-          # imported regex
-          regex=$(jq -r '.regex' config.json)
-
-          # imported list of string
-          mapfile -t patterns < <(jq -r '.strings[]' config.json)
-
-          has_violation=0
-
-          #######################
-          # check commit messages and author info
-          while IFS='|' read -r hash msg an ae cn ce; do
-            for field in "$msg" "$an" "$ae" "$cn" "$ce"; do
-              echo "$field"
-              if echo "$field" | grep -E "$regex"; then
-                echo "::error::❌ Regex match in commit $hash: '$field'"
-                has_violation=1
-              fi
-              for s in "${patterns[@]}"; do
-                if echo "$field" | grep -F "$s"; then
-                  echo "::error::❌ String '$s' found in commit $hash: '$field'"
-                  has_violation=1
-                fi
-              done
-            done
-          done < <(git log "$range" --pretty=format:'%H|%s|%an|%ae|%cn|%ce')
-
-          #######################
-          # check in added lines of code
-
-          git diff "$range" --unified=0 > diff.patch
-          current_file=""
-
-          while IFS= read -r line; do
-          if [[ $line =~ ^\+\+\+\ b/(.*) ]]; then
-            current_file="${BASH_REMATCH[1]}"
-          fi
-
-          # Process only added lines (ignore "+++" headers)
-          if [[ $line =~ ^\+[^+]{1} ]]; then
-            content="${line:1}"
-            for pattern in "${patterns[@]}"; do
-            if grep -q -E "$pattern" <<< "$content"; then
-              matches=$(grep -n -o "$pattern" "$current_file" | cut -d: -f1)
-              formatted=$(echo $matches | sed 's/ /, /g')
-              echo "❌ Pattern '$pattern' found in $current_file at lines: [$formatted]"
-              has_violation=1
-            fi
-            done
-          fi
-          done < diff.patch
-
-          #######################
-          rm diff.patch config.json
-
-          if [ "$has_violation" = 1 ]; then
-            echo "❌ Sensitive content found."
-            exit 1
-          else
-            echo "✅ No sensitive content found."
-          fi
