### 2025-06-17 9655b0f
fix(.pre-commit-hooks) remove duplicate code
