+++ .pre_commit_hooks/check_confidential_patterns.py
@@ -6,15 +6,6 @@ import subprocess
 from collections.abc import Sequence
 from pathlib import Path
 
-# Path to current script
-script_dir = Path(__file__).resolve().parent
-pattern_file = script_dir / "patterns.txt"
-
-# Get patterns
-with pattern_file.open("r") as f:
-    lines = [line.strip() for line in f if line.strip()]
-    COMPILED_PATTERN = [re.compile(line) for line in lines]
-
 
 def main(argv: Sequence[str] | None = None) -> int:
     """Function to check for confidential patterns in staged changes."""
