### 2025-06-17 d7544b9
(stats) improve OnlineStatistics and Stats classes, better docstrings, better type checking, better modularity
