+++ src/plaid/utils/stats.py
@@ -10,7 +10,7 @@
 # %% Imports
 
 import logging
-from typing import Union
+from typing import List, Union
 
 import numpy as np
 
@@ -60,7 +60,18 @@ def aggregate_stats(
 
 
 class OnlineStatistics(object):
-    """OnlineStatistics is a class for computing online statistics (e.g., min, max, mean, variance, and standard deviation) of numpy arrays."""
+    """OnlineStatistics is a class for computing online statistics of numpy arrays.
+
+    This class computes running statistics (min, max, mean, variance, std) for streaming data
+    without storing all samples in memory.
+
+    Example:
+        >>> stats = OnlineStatistics()
+        >>> stats.add_samples(np.array([[1, 2], [3, 4]]))
+        >>> stats.add_samples(np.array([[5, 6]]))
+        >>> print(stats.get_stats()['mean'])
+        [[3. 4.]]
+    """
 
     def __init__(self) -> None:
         """Initialize an empty OnlineStatistics object."""
@@ -70,6 +81,7 @@ class OnlineStatistics(object):
         self.mean: np.ndarray = None
         self.var: np.ndarray = None
         self.std: np.ndarray = None
+        self._n_features: int = None  # Add feature dimension tracking
 
     def add_samples(self, x: np.ndarray) -> None:
         """Add samples to compute statistics for.
@@ -79,23 +91,36 @@ class OnlineStatistics(object):
 
         Raises:
             ShapeError: Raised when there is an inconsistency in the shape of the input array.
+            ValueError: Raised when input contains NaN or Inf values.
         """
+        # Validate input
+        if not isinstance(x, np.ndarray):
+            raise TypeError("Input must be a numpy array")
+
+        if np.any(~np.isfinite(x)):
+            raise ValueError("Input contains NaN or Inf values")
+
+        # Handle 1D arrays
         if x.ndim == 1:
             if self.min is not None:
                 if self.min.shape[1] == 1:
-                    # n_samples x 1
                     x = x.reshape((-1, 1))
                 else:
-                    # 1 x n_features
                     x = x.reshape((1, -1))
-            else:  # pragma: no cover
-                raise ShapeError(
-                    "can't determine if input array with ndim=1, is 1 x n_features or n_samples x 1"
-                )
+            else:
+                x = x.reshape((-1, 1))  # Default to column vector
+
+        # Handle n-dimensional arrays
         elif x.ndim > 2:
-            # suppose last dim is features dim, all previous dims are space
-            # dims and are aggregated
             x = x.reshape((-1, x.shape[-1]))
+
+        # Validate feature dimension
+        if self._n_features is not None and x.shape[1] != self._n_features:
+            raise ShapeError(
+                f"Input has {x.shape[1]} features but expected {self._n_features}"
+            )
+        else:
+            self._n_features = x.shape[1]
 
         added_n_samples = len(x)
         added_min = np.min(x, axis=0, keepdims=True)
@@ -166,12 +191,19 @@ class OnlineStatistics(object):
         }
 
 
-class Stats(object):
-    """Stats is a class for aggregating and computing statistics for datasets."""
+class Stats:
+    """Class for aggregating and computing statistics across datasets.
+
+    The Stats class processes both scalar and field data from samples or datasets,
+    computing running statistics like min, max, mean, variance and standard deviation.
+
+    Attributes:
+        _stats (dict[str, OnlineStatistics]): Dictionary mapping data identifiers to their statistics
+    """
 
     def __init__(self):
         """Initialize an empty Stats object."""
-        self._stats = {}
+        self._stats: dict[str, OnlineStatistics] = {}
 
     def add_dataset(self, dset: Dataset) -> None:
[…]
