+++ .pre-commit-config.yaml
@@ -22,8 +22,8 @@ repos:
         types: [python]
         pass_filenames: false
 
-      - id: check-confidential-patterns
-        name: Confidential Pattern Check
-        entry: python .pre_commit_hooks/check_confidential_patterns.py
+      - id: check-patterns
+        name: patterns check
+        entry: python .pre_commit_hooks/check_patterns.py
         language: system
         types: [python]
