### 2025-06-19 a02d4d5
🚀 feat(*/*) Extend python versions to 3.9 and 3.10 (#76)
* feat(*/*) extend support for python 3.9 and 3.10

---------

Co-authored-by: Fabien Casenave <fabien.casenave@safrangroup.com>
