+++ .pre-commit-config.yaml
@@ -21,9 +21,3 @@ repos:
         language: system
         types: [python]
         pass_filenames: false
-
-      - id: check-patterns
-        name: patterns check
-        entry: python .pre_commit_hooks/check_patterns.py
-        language: system
-        types: [python]
