+++ src/plaid/problem_definition.py
@@ -30,6 +30,7 @@ logging.basicConfig(
     level=logging.INFO,
 )
 
+
 # %% Functions
 
 # %% Classes
