+++ .pre_commit_hooks/patterns.txt
@@ -6,5 +6,5 @@
 /gpfs/
 /apps/
 # Windows
-C:
-D:
+C:\\
+D:\\
