+++ src/plaid/problem_definition.py
@@ -30,7 +30,6 @@ logging.basicConfig(
     level=logging.INFO,
 )
 
-
 # %% Functions
 
 # %% Classes
