+++ .pre_commit_hooks/patterns.txt
@@ -6,6 +6,7 @@
 /data/
 /gpfs/
 /apps/
+/softs/
 # Windows
 C:\\
 D:\\
