### 2025-06-19 dcc6112
🚀 feat(actions) Detect pattern with pre-commit (#77)
---------

Co-authored-by: Fabien Casenave <fabien.casenave@safrangroup.com>
