+++ src/plaid/containers/dataset.py
@@ -38,7 +38,6 @@ logging.basicConfig(
     level=logging.INFO,
 )
 
-
 authorized_info_keys = {
     "legal": ["owner", "license"],
     "data_production": [
