### 2025-06-19 2cef06d
♻️  feat(actions) Remove joss paper action workflow file
