### 2025-06-23 36585b2
feat(readme/doc) add joss paper status badge
