### 2025-06-23 36ec911
📄 feat(readme/doc) add joss paper status badge (#80)
Co-authored-by: Fabien Casenave <fabien.casenave@safrangroup.com>
