### 2025-06-25 4911217
feat(tests) add scalar named 'a/b' in sample tests
