### 2025-06-25 8bf3bd2
fix(samples) add time_series summary in __str__
