### 2025-06-25 531c452
add test for time_series with name containing /
