+++ tests/containers/test_sample.py
@@ -110,6 +110,9 @@ def sample_with_tree_and_scalar_and_time_series(
     sample_with_tree.add_time_series(
         "test_time_series_1", np.arange(111, dtype=float), np.random.randn(111)
     )
+    sample_with_tree.add_time_series(
+        "a/b", np.arange(113, dtype=float), np.random.randn(113)
+    )
     return sample_with_tree
 
 
