### 2025-06-25 7d3b2a1
fix plural for 0 quantity
