+++ examples/pipeline/pipefunc_tensile2d.py
@@ -7,97 +7,96 @@ from safetensors.numpy import save_file
 from pipefunc import Pipeline, pipefunc
 from sklearn.preprocessing import StandardScaler
 
+import yaml
 
 
-# ids_train = hf_dataset.description["split"]['train_500']
-# ids_test  = hf_dataset.description["split"]['test']
+@pipefunc(output_name=("dataset", "prob_def"))
+def load_hf_from_disk(path):
+    return huggingface_dataset_to_plaid(load_from_disk(path))
 
-# sample_train_0 = Sample.model_validate(pickle.loads(hf_dataset[ids_train[0]]["sample"]))
-# sample_test_0 = Sample.model_validate(pickle.loads(hf_dataset[ids_test[0]]["sample"]))
+@pipefunc(output_name="scalar_scalers")
+def scale_scalars(dataset, prob_def, train_split_name, test_split_name):
 
+    print(">>", dataset[0].get_scalar("p1"))
 
-# Step 1: Image Loading and Preprocessing
-@pipefunc(output_name="hf_dataset")
-def load(path):
-    return load_from_disk(path)
-
-@pipefunc(output_name="dataset")
-def convert_to_plaid(hf_dataset):
-    return huggingface_dataset_to_plaid(hf_dataset)[0]
-
-@pipefunc(output_name="problem_definition")
-def generate_prob_def(hf_dataset):
-    return huggingface_description_to_problem_definition(hf_dataset.description)
-
-# # Step 2: Image Segmentation
-@pipefunc(output_name="scaler")
-def scale_scalars(dataset, problem_definition, train_split_name, test_split_name):
-
-    ids_train = problem_definition.get_split(train_split_name)
+    ids_train = prob_def.get_split(train_split_name)
     train_scalars = dataset.get_scalars_to_tabular(
         sample_ids = ids_train
     )
 
-    ids_test = problem_definition.get_split(test_split_name)
+    ids_test = prob_def.get_split(test_split_name)
     test_scalars = dataset.get_scalars_to_tabular(
         sample_ids = ids_test
     )
 
-    for sn in problem_definition.get_input_scalars_names():
+    scalar_scalers = {}
+
+    print(prob_def.get_input_scalars_names())
+
+    for sn in prob_def.get_input_scalars_names():
         scaler = StandardScaler()
         scaler.fit_transform(train_scalars[sn].reshape(-1, 1))
-        scaler.fit(test_scalars[sn].reshape(-1, 1))
+        scaler.transform(test_scalars[sn].reshape(-1, 1))
+        scalar_scalers[sn] = scaler
 
-    # print(ids_train)
-    # print(ids_test)
-    return scaler
+    for sn in prob_def.get_output_scalars_names():
+        scaler = StandardScaler()
+        scaler.fit_transform(train_scalars[sn].reshape(-1, 1))
+        scalar_scalers[sn] = scaler
+
+    for j, i in enumerate(ids_train):
+        sample = dataset[i]
+        for sn, scaler in scalar_scalers.items():
+            if sn in sample.get_scalar_names():
+                sample.add_scalar(sn, train_scalars[sn][j])
+
+    for j, i in enumerate(ids_test):
+        sample = dataset[i]
+        for sn, scaler in scalar_scalers.items():
+            if sn in sample.get_scalar_names():
+                sample.add_scalar(sn, test_scalars[sn][j])
+
+
+    print(">>", dataset[0].get_scalar("p1"))
+
+    return scalar_scalers
 
 @pipefunc(output_name="saved_path")
-def save(scaler, out_path):
+def save(scalar_scalers, out_path, dataset):
 
     os.makedirs(out_path, exist_ok=True)
 
     # Save only NumPy-compatible parameters
-    tensors = {
-        "scaler_mean": scaler.mean_,
[…]
