+++ examples/pipeline/config.yml
@@ -1,0 +1,9 @@
+load:
+  - path:
+
+scale_scalars:
+  - train_split_name: "train_500"
+  - test_split_name: "test"
+
+save:
+  - out_path: ".\\artifacts"
