### 2025-06-26 879db5f
feat(docs) replace data challenges section with plaid benchmarks
