+++ docs/source/plaid_benchmarks.rst
@@ -1,0 +1,151 @@
+PLAID Benchmarks
+================
+
+.. image:: images/plaid_benchmarks.png
+    :align: center
+    :width: 600px
+    :alt: PLAID Benchmarks
+
+
+We provide interactive benchmark hosted on Hugging Face, in which anyone can test their own SciML method.
+These benchmarks involve regression problem posed on datasets provided in PLAID format.
+Some of these datasets have been introduced in the MMGP (Mesh Morphing Gaussian Process) paper :cite:p:`casenave2023mmgp`,
+and the PLAID paper :cite:p:`casenave2025plaid`.
+A ranking is automatically updated based on a score computed on the testing set of each dataset.
+For the benchmarks to be meaningful, the outputs on the testing sets are not made public.
+
+The relative RMSE is the considered metric for comparing methods. Let :math:`\{ \mathbf{U}^i_{\rm ref} \}_{i=1}^{n_\star}`
+and :math:`\{ \mathbf{U}^i_{\rm pred} \}_{i=1}^{n_\star}` be respectively test observations and predictions of a given field of interest.
+The relative RMSE is defined as
+
+.. math::
+
+    \mathrm{RRMSE}_f(\mathbf{U}_{\rm ref}, \mathbf{U}_{\rm pred}) = \left( \frac{1}{n_\star}\sum_{i=1}^{n_\star} \frac{\frac{1}{N^i}\|\mathbf{U}^i_{\rm ref} - \mathbf{U}^i_{\rm pred}\|_2^2}{\|\mathbf{U}^i_{\rm ref}\|_\infty^2} \right)^{1/2},
+
+where :math:`N^i` is the number of nodes in the mesh :math:`i`, and :math:`\max(\mathbf{U}^i_{\rm ref})` is the maximum entry in the vector :math:`\mathbf{U}^i_{\rm ref}`. Similarly for scalar outputs:
+
+.. math::
+
+    \mathrm{RRMSE}_s(\mathbf{w}_{\rm ref}, \mathbf{w}_{\rm pred}) = \left( \frac{1}{n_\star} \sum_{i=1}^{n_\star} \frac{|w^i_{\rm ref} - w_{\rm pred}^i|^2}{|w^i_{\rm ref}|^2} \right)^{1/2}.
+
+
+Resources
+---------
+
++---------------------+-------------------------------------------+-----------------------------------------------+
+|                     |           Dataset                         |                  Benchmark                    |
++---------------------+-------------------------------------------+-----------------------------------------------+
+| **Tensile2d**       | |Tensile2d_HF| |Tensile2d_Z|              | |Tensile2d_Be|                                |
++---------------------+-------------------------------------------+-----------------------------------------------+
+| **2D_MultiScHypEl** | |2D_MultiScHypEl_HF| |2D_MultiScHypEl_Z|  | |2D_MultiScHypEl_Be|                          |
++---------------------+-------------------------------------------+-----------------------------------------------+
+| **2D_ElPlDynamics** | |2D_ElPlDynamics_HF| |2D_ElPlDynamics_Z|  | |2D_ElPlDynamics_Be|                          |
++---------------------+-------------------------------------------+-----------------------------------------------+
+| **Rotor37**         | |Rotor37_HF|   |Rotor37_Z|                | |Rotor37_Be|                                  |
[…]
