### 2025-06-27 5807458
feat(docs) add pypi download badge
