### 2025-06-27 150b55d
(stats) implement logic to check field shapes and flatten if incoherent between samples + improve tests for type checking + tests for new methods
