+++ tests/utils/test_stats.py
@@ -10,6 +10,7 @@
 import numpy as np
 import pytest
 
+from plaid.containers.sample import Sample
 from plaid.utils.stats import OnlineStatistics, Stats
 
 # %% Fixtures
@@ -50,6 +51,70 @@ def stats():
     return Stats()
 
 
+@pytest.fixture()
+def sample_data_1():
+    return np.random.randn(100, 1)
+
+
+@pytest.fixture()
+def sample_data_2():
+    return np.random.randn(50, 1)
+
+
+@pytest.fixture()
+def sample_with_scalar(sample_data_1):
+    s = Sample()
+    s.add_scalar("foo", float(sample_data_1.mean()))
+    return s
+
+
+@pytest.fixture()
+def sample_with_field(sample_data_2):
+    s = Sample()
+    # 1. Initialize the CGNS tree
+    s.init_tree()
+    # 2. Create a base and a zone
+    s.init_base(topological_dim=3, physical_dim=3)
+    s.init_zone(zone_shape=np.array([sample_data_2.shape[0], 0, 0]))
+    # 3. Set node coordinates (required for a valid zone)
+    s.set_nodes(np.zeros((sample_data_2.shape[0], 3)))
+    # 4. Add a field named "bar"
+    s.add_field(name="bar", field=sample_data_2)
+    return s
+
+
+@pytest.fixture()
+def time_series_data():
+    # 10 time steps, 1 feature
+    times = np.linspace(0, 1, 10)
+    values = np.random.randn(10)
+    return times, values
+
+
+@pytest.fixture()
+def time_series_data_of_different_size():
+    # 5 time steps, 1 feature
+    times = np.linspace(0, 1, 5)
+    values = np.random.randn(5)
+    return times, values
+
+
+@pytest.fixture()
+def sample_with_time_series(time_series_data):
+    s = Sample()
+    times, values = time_series_data
+    s.add_time_series("ts1", time_sequence=times, values=values)
+    return s
+
+
+@pytest.fixture()
+def sample_with_time_series_of_different_size(time_series_data_of_different_size):
+    s = Sample()
+    times, values = time_series_data_of_different_size
+    s.add_time_series("ts1", time_sequence=times, values=values)
+    return s
+
+
 # %% Tests
 
 
@@ -82,6 +147,38 @@ class Test_OnlineStatistics:
         online_stats.add_samples(np_samples_1)
         online_stats.get_stats()
 
+    def test_invalid_input_type(self, online_stats):
+        with pytest.raises(TypeError):
+            online_stats.add_samples([1, 2, 3])  # List instead of ndarray
+
+    def test_nan_inf_input(self, online_stats):
+        with pytest.raises(ValueError):
+            online_stats.add_samples(np.array([1, np.nan, 3]))
+        with pytest.raises(ValueError):
+            online_stats.add_samples(np.array([1, np.inf, 3]))
+
+    def test_merge_stats(self, sample_data_1, sample_data_2):
+        stats1 = OnlineStatistics()
+        stats2 = OnlineStatistics()
+        stats1.add_samples(sample_data_1)
+        stats2.add_samples(sample_data_2)
+        n_samples_before = stats1.n_samples
+        n_samples_other = stats2.n_samples
+        mean_before = stats1.mean.copy()
+        other_mean = stats2.mean.copy()
+        # do the merging
+        stats1.merge_stats(stats2)
+        # check results
+        assert stats1.n_samples == n_samples_before + stats2.n_samples
+        print(f"{n_samples_before=}, {n_samples_other=}")
+        print(f"{mean_before=}, {other_mean=}")
+        expected_mean = (
+            mean_before * n_samples_before + other_mean * n_samples_other
[…]
