+++ examples/utils/stats_example.py
@@ -33,8 +33,8 @@ from plaid.utils.stats import OnlineStatistics, Stats
 # %%
 def sprint(stats: dict):
     print("Stats:")
-    for k in stats:
-        print(" - {} -> {}".format(k, stats[k]))
+    for k,v in stats.items():
+        print(f" - {k:>9} -> {v}")
 
 
 # %% [markdown]
@@ -58,7 +58,8 @@ sprint(stats)
 # %%
 # First batch of samples
 first_batch_samples = 3.0 * np.random.randn(100, 3) + 10.0
-print(f"{first_batch_samples.shape = }")
+print()
+print(f"=== {first_batch_samples.shape = }")
 
 stats_computer.add_samples(first_batch_samples)
 stats = stats_computer.get_stats()
@@ -67,7 +68,8 @@ sprint(stats)
 
 # %%
 second_batch_samples = 10.0 * np.random.randn(1000, 3) - 1.0
-print(f"{second_batch_samples.shape = }")
+print()
+print(f"=== {second_batch_samples.shape = }")
 
 stats_computer.add_samples(second_batch_samples)
 stats = stats_computer.get_stats()
@@ -79,7 +81,8 @@ sprint(stats)
 
 # %%
 total_samples = np.concatenate((first_batch_samples, second_batch_samples), axis=0)
-print(f"{total_samples.shape = }")
+print()
+print(f"=== {total_samples.shape = }")
 
 new_stats_computer = OnlineStatistics()
 new_stats_computer.add_samples(total_samples)
@@ -110,14 +113,23 @@ print("#---# Feed Stats with samples")
 nb_samples = 11
 samples = [Sample() for _ in range(nb_samples)]
 
-spatial_shape_max = 20
+spatial_shape_max = 5
 #
 for sample in samples:
     sample.add_scalar("test_scalar", np.random.randn())
-    sample.init_base(2, 3, "test_base")
+    sample.add_scalar("test_ND_scalar", np.random.randn(3))
+    sample.init_base(2, 3,)
     zone_shape = np.array([0, 0, 0])
-    sample.init_zone(zone_shape, zone_name="test_zone")
+    sample.init_zone(zone_shape)
     sample.add_field("test_field", np.random.randn(spatial_shape_max))
+    sample.init_zone(zone_shape, zone_name="test_zone_2")
+    sample.add_field("test_field", np.random.randn(spatial_shape_max), zone_name='test_zone_2')
+    sample.init_base(2, 3, "test_base_2")
+    zone_shape = np.array([0, 0, 0])
+    sample.init_zone(zone_shape, zone_name="test_zone_1", base_name="test_base_2")
+    sample.add_field("test_field", np.random.randn(spatial_shape_max), base_name="test_base_2")
+    sample.init_zone(zone_shape, zone_name="test_zone_2", base_name="test_base_2")
+    sample.add_field("test_field", np.random.randn(spatial_shape_max), zone_name='test_zone_2', base_name="test_base_2")
 
 stats.add_samples(samples)
 
@@ -132,21 +144,38 @@ rich.print(stats.get_stats())
 # ### Feed Stats with more Samples
 
 # %%
-nb_samples = 11
-spatial_shape_max = 20
+nb_samples = 13
+# spatial_shape_max = 20
 samples = [Sample() for _ in range(nb_samples)]
 
 for sample in samples:
     sample.add_scalar("test_scalar", np.random.randn())
+    sample.init_base(2, 3,)
+    zone_shape = np.array([0, 0, 0])
+    sample.init_zone(zone_shape)
+    sample.add_field("test_field", np.random.randn(spatial_shape_max))
+    sample.init_zone(zone_shape, zone_name="test_zone_2")
[…]

[…]
