+++ examples/pipeline/config.yml
@@ -1,9 +1,9 @@
-load:
-  - path:
+load_hf_from_hub:
+  - path: "PLAID-datasets/Tensile2d"
 
 scale_scalars:
   - train_split_name: "train_500"
   - test_split_name: "test"
 
 save:
-  - out_path: ".\\artifacts"
+  - out_path: "./artifacts"
