### 2025-06-28 8387929
feat(tests/utils/stats) add tests to improve coverage
