+++ src/plaid/utils/stats.py
@@ -199,6 +199,7 @@ class OnlineStatistics(object):
             self.flatten_array()
             other = copy.deepcopy(other)
             other.flatten_array()
+            assert self.min.shape == other.min.shape, f"Shape mismatch in OnlineStatistics merging"
 
         self.min = np.min(np.concatenate((self.min, other.min), axis=0), axis=0)
         self.max = np.max(np.concatenate((self.max, other.max), axis=0), axis=0)
@@ -212,8 +213,8 @@ class OnlineStatistics(object):
 
     def flatten_array(self) -> None:
         """When a shape incoherence is detected, you should call this function."""
-        self.min = np.min(self.min, keepdims=True)
-        self.max = np.max(self.max, keepdims=True)
+        self.min = np.min(self.min, keepdims=True).reshape(-1, 1)
+        self.max = np.max(self.max, keepdims=True).reshape(-1, 1)
         self.n_points = self.n_samples * self.n_features
         assert self.mean.shape == self.var.shape
         self.n_points, self.mean, self.var = aggregate_stats(
@@ -422,7 +423,7 @@ class Stats:
                                     # set this stat as flattened
                                     self._field_is_flattened[stat_key] = True
                                     # flatten corresponding stat
-                                    if stat_key in self._stats:
+                                    if stat_key in self._stats: # TODO: ADD THIS IN TESTS
                                         self._stats[stat_key].flatten_array()
 
                             if self._field_is_flattened.get(stat_key, False):
@@ -447,7 +448,7 @@ class Stats:
                     if (
                         isinstance(list_of_arrays[i], np.ndarray)
                         and list_of_arrays[i].ndim == 1
-                    ):
+                    ): # TODO: ADD THIS IN TESTS
                         list_of_arrays[i] = list_of_arrays[i].reshape((-1, 1))
 
                 if self._field_is_flattened.get(name, False):
