### 2025-06-28 696b366
fix(docs/source/plaid_benchmarks.rst) writing
