### 2025-06-28 e534477
fix(stats) apply ruff formatting
