### 2025-06-28 4687e93
fix(pyproject.toml) upper bound in python version
