+++ examples/pipeline/ml_pipeline_nodes.py
@@ -1,0 +1,377 @@
+import os
+import json
+import joblib
+from joblib import Parallel, delayed
+import numpy as np
+from pathlib import Path
+from sklearn.base import BaseEstimator, RegressorMixin, TransformerMixin
+from sklearn.linear_model import Ridge
+from sklearn.decomposition import PCA
+
+
+class PersistentNode(BaseEstimator, RegressorMixin, TransformerMixin):
+    def __init__(self, save_path):
+        self.save_path = Path(save_path)
+        self.fitted_ = False
+        self.model = None
+
+    def save(self, obj):
+        self.save_path.parent.mkdir(parents=True, exist_ok=True)
+        joblib.dump(obj, self.save_path)
+        self.fitted_ = True
+
+    def load(self):
+        print(f"Loading existing model from {self.save_path}")
+        obj = joblib.load(self.save_path)
+        self.fitted_ = True
+        return obj
+
+    def exists(self):
+        return self.save_path.exists()
+
+    def check_fitted_or_load(self):
+        if self.fitted_:
+            return
+        if self.exists():
+            self.set_model(self.load())
+        else:
+            raise ValueError("Model not fitted and no saved model found.")
+
+    def set_model(self, obj):
+        self.model = obj
+
+    def fit(self, X, y=None):
+        if self.exists():
+            self.set_model(self.load())
+            return self
+        self._fit(X, y)
+        self.save(self.model)
+        return self
+
+    def transform(self, X):
+        self.check_fitted_or_load()
+        return self._transform(X)
+
+    def inverse_transform(self, X):
+        self.check_fitted_or_load()
+        return self._inverse_transform(X)
+
+    def predict(self, X):
+        self.check_fitted_or_load()
+        return self._predict(X)
+
+    def score(self, X, y):
+        self.check_fitted_or_load()
+        return self._score(X, y)
+
+    # Protected methods to override in subclasses
+    def _fit(self, X, y=None):
+        raise NotImplementedError
+
+    def _predict(self, X):
+        raise NotImplementedError
+
+    def _transform(self, X):
+        raise NotImplementedError
+
+    def _inverse_transform(self, X):
+        raise NotImplementedError
+
+    def _score(self, X, y):
+        raise NotImplementedError
+
+
+
+from sklearn.preprocessing import StandardScaler, MinMaxScaler
+
+available_scalers = {
+    "StandardScaler": StandardScaler,
+    "MinMaxScaler": MinMaxScaler,
+}
+
+class ScalarScalerNode(PersistentNode):
+
+    def __init__(self, name, params):
+        super().__init__(os.path.join(params['save_path'], f"{name}.joblib"))
+        self.scalar_names = params[name]['scalar_names']
+        self.model = available_scalers[params[name]['type']]()
+
+    def get_scalars(self, dataset):
+        return dataset.get_scalars_to_tabular(
+            scalar_names = self.scalar_names,
+            as_nparray = True
+        )
+
+    def set_scalars(self, dataset, scalars):
+        for i in range(len(dataset)):
+            for j, sn in enumerate(self.scalar_names):
+                dataset[i].add_scalar(sn, scalars[i, j])
+
+    def _fit(self, dataset, y=None):
+        scalars = self.get_scalars(dataset)
+        self.model.fit(scalars)
+
+    def _transform(self, dataset):
+        scalars = self.get_scalars(dataset)
+        scaled_scalars = self.model.transform(scalars)
+        self.set_scalars(dataset, scaled_scalars)
+
+        return dataset
+
+    def _inverse_transform(self, dataset):
+        scaled_scalars = self.get_scalars(dataset)
[…]
