### 2025-06-29 64db621
wip: a cleaned version of scikit-learn pipelines is provided in examples/pipelines
