+++ examples/pipelines/ml_pipeline_nodes.py
@@ -1,0 +1,249 @@
+import os
+import json
+import joblib
+from joblib import Parallel, delayed
+import numpy as np
+from pathlib import Path
+from sklearn.base import BaseEstimator, RegressorMixin, TransformerMixin
+from sklearn.linear_model import Ridge
+from sklearn.decomposition import PCA
+from plaid.containers.dataset import Dataset
+import copy
+
+from sklearn.preprocessing import StandardScaler, MinMaxScaler
+
+available_scalers = {
+    "StandardScaler": StandardScaler,
+    "MinMaxScaler": MinMaxScaler,
+}
+
+class ScalarScalerNode(BaseEstimator, TransformerMixin):
+
+    def __init__(self, type, scalar_names):
+        self.type = type
+        self.scalar_names = scalar_names
+
+        assert type in available_scalers.keys(), "Scaler "+type+" not available"
+
+        self.model = None
+
+
+    def get_scalars(self, dataset):
+        if isinstance(dataset, list):
+            dataset = Dataset.from_list_of_samples(dataset)
+        return dataset.get_scalars_to_tabular(
+            scalar_names = self.scalar_names,
+            as_nparray = True
+        )
+
+    def set_scalars(self, dataset, scalars):
+        for i in range(len(dataset)):
+            for j, sn in enumerate(self.scalar_names):
+                dataset[i].add_scalar(sn, scalars[i, j])
+
+    def fit(self, dataset, y=None):
+        self.model = available_scalers[self.type]()
+
+        scalars = self.get_scalars(dataset)
+        self.model.fit(scalars)
+        self.fitted_ = True
+        return self
+
+    def transform(self, dataset):
+        scalars = self.get_scalars(dataset)
+        scaled_scalars = self.model.transform(scalars)
+        dataset_ = copy.deepcopy(dataset)
+        self.set_scalars(dataset_, scaled_scalars)
+        return dataset_
+
+    def inverse_transform(self, dataset):
+        scaled_scalars = self.get_scalars(dataset)
+        scalars = self.model.inverse_transform(scaled_scalars)
+        dataset_ = copy.deepcopy(dataset)
+        self.set_scalars(dataset_, scalars)
+        return dataset_
+
+
+class PCAEmbeddingNode(BaseEstimator, RegressorMixin, TransformerMixin):
+
+    def __init__(self, field_name = None, n_components = None, zone_name = None, base_name = None, time = None, location = "Vertex"):
+
+        self.zone_name = zone_name
+        self.base_name = base_name
+        self.time      = time
+        self.location  = location
+
+        self.field_name   = field_name
+        self.n_components = n_components
+
+        self.model = None
+
+    def get_all_fields(self, dataset):
+        all_fields = []
+        for sample in dataset:
+            if self.field_name == "nodes":
+                field = sample.get_nodes(self.zone_name, self.base_name, self.time).flatten()
+            else:
+                field = sample.get_field(self.field_name, self.zone_name, self.base_name, self.location, self.time)
+            all_fields.append(field)
+        return np.array(all_fields)
+
+    def set_reduced_fields(self, dataset, reduced_fields):
+        for i in range(len(dataset)):
+            for j in range(self.n_components):
+                dataset[i].add_scalar(f"reduced_{self.field_name}_{j}", reduced_fields[i, j])
+
+    def get_reduced_fields(self, dataset):
+        if isinstance(dataset, list):
+            dataset = Dataset.from_list_of_samples(dataset)
+        return dataset.get_scalars_to_tabular(
+            scalar_names = [f"reduced_{self.field_name}_{j}" for j in range(self.n_components)],
+            as_nparray = True
+        )
+
+    def set_fields(self, dataset, fields):
+        for i in range(len(dataset)):
[…]
