+++ src/plaid/containers/dataset.py
@@ -698,6 +698,20 @@ class Dataset(object):
         shutil.rmtree(savedir)
 
     @classmethod
+    def from_list_of_samples(cls, list_of_samples: list[Sample]) -> Self:
+        """Initialise a dataset from a list of samples.
+
+        Args:
+            list_of_samples (list[Sample]): The list of samples.
+
+        Returns:
+            Self: The intialized dataset (Dataset).
+        """
+        instance = cls()
+        instance.add_samples(list_of_samples)
+        return instance
+
+    @classmethod
     def load_from_file(
         cls, fname: Union[str, Path], verbose: bool = False, processes_number: int = 0
     ) -> Self:
@@ -1025,17 +1039,19 @@ class Dataset(object):
         """
         return len(self._samples)
 
-    def __getitem__(self, id: int) -> Sample:
+    def __getitem__(
+        self, id: Union[int, slice, list[int], np.ndarray]
+    ) -> Union[Sample, Self]:
         """Retrieve a specific sample by its ID int this dataset.
 
         Args:
-            id (int): The ID of the sample to retrieve.
+            id (Union[int, slice, list[int], np.ndarray]): The ID(s) of the sample to retrieve.
 
         Raises:
             IndexError: If the provided ID is out of bounds or does not exist in the dataset.
 
         Returns:
-            Sample: The sample with the specified ID.
+            Union[Sample, Dataset]: The sample with the specified ID or a dataset in the specified IDs.
 
         Example:
             .. code-block:: python
@@ -1047,12 +1063,21 @@ class Dataset(object):
         Seealso:
             This function can also be called using `__call__()`.
         """
-        if id in self._samples:
-            return self._samples[id]
+        if isinstance(id, slice) or isinstance(id, list) or isinstance(id, np.ndarray):
+            if isinstance(id, slice):
+                id = list(range(*id.indices(len(self))))
+            dataset = Dataset()
+            for i in id:
+                dataset.add_sample(self[i], i)
+            dataset.set_infos(self.get_infos())
+            return dataset
         else:
-            raise IndexError(
-                f"sample with {id=} not set -> use 'Dataset.add_sample' or 'Dataset.add_samples'"
-            )
+            if id in self._samples:
+                return self._samples[id]
+            else:
+                raise IndexError(
+                    f"sample with {id=} not set -> use 'Dataset.add_sample' or 'Dataset.add_samples'"
+                )
 
     __call__ = __getitem__
 

[…]
