### 2025-06-29 c89f2f3
(stats) fix flattening of time_series
