+++ tests/utils/test_stats.py
@@ -79,6 +79,16 @@ def sample_with_field(np_samples_6):
 
 
 @pytest.fixture()
+def field_data():
+    return np.random.randn(101)
+
+
+@pytest.fixture()
+def field_data_of_different_size():
+    return np.random.randn(51)
+
+
+@pytest.fixture()
 def time_series_data():
     # 10 time steps, 1 feature
     times = np.linspace(0, 1, 10)
@@ -95,18 +105,26 @@ def time_series_data_of_different_size():
 
 
 @pytest.fixture()
-def sample_with_time_series(time_series_data):
+def sample_with_time_series(time_series_data, field_data):
     s = Sample()
     times, values = time_series_data
     s.add_time_series("ts1", time_sequence=times, values=values)
+    s.init_base(1, 1)
+    s.init_zone(np.array([0, 0, 0]))
+    s.add_field(name="field1", field=field_data)
     return s
 
 
 @pytest.fixture()
-def sample_with_time_series_of_different_size(time_series_data_of_different_size):
+def sample_with_time_series_of_different_size(
+    time_series_data_of_different_size, field_data_of_different_size
+):
     s = Sample()
     times, values = time_series_data_of_different_size
     s.add_time_series("ts1", time_sequence=times, values=values)
+    s.init_base(1, 1)
+    s.init_zone(np.array([0, 0, 0]))
+    s.add_field(name="field1", field=field_data_of_different_size)
     return s
 
 
@@ -143,7 +161,20 @@ class Test_OnlineStatistics:
 
     def test_get_stats(self, online_stats, np_samples_1):
         online_stats.add_samples(np_samples_1)
-        online_stats.get_stats()
+        stats_dict = online_stats.get_stats()
+        # Check that all expected statistics keys are present
+        expected_keys = {
+            "mean",
+            "min",
+            "max",
+            "var",
+            "std",
+            "n_samples",
+            "n_points",
+            "n_features",
+        }
+        for key in expected_keys:
+            assert key in stats_dict, f"Missing key: {key}"
 
     def test_invalid_input_type(self, online_stats):
         with pytest.raises(TypeError):
@@ -195,7 +226,35 @@ class Test_Stats:
 
     def test_get_stats(self, stats, samples):
         stats.add_samples(samples)
-        stats.get_stats()
+        stats_dict = stats.get_stats()
+        sample: Sample = samples[0]
+        feature_names = sample.get_scalar_names()
+        for base_name in sample.get_base_names():
+            for zone_name in sample.get_zone_names(base_name=base_name):
+                for field_name in sample.get_field_names(
+                    zone_name=zone_name, base_name=base_name
+                ):
+                    feature_names.append(f"{base_name}/{zone_name}/{field_name}")
+        feature_names.extend(sample.get_time_series_names())
+        for feat_name in feature_names:
+            assert feat_name in stats_dict, (
+                f"Missing {feat_name=}, in {stats_dict.keys()}"
+            )
+            # Check that all expected statistics keys are present
+            expected_keys = {
+                "mean",
+                "min",
+                "max",
+                "var",
+                "std",
+                "n_samples",
+                "n_points",
+                "n_features",
+            }
+            for key in expected_keys:
+                assert key in stats_dict[feat_name], (
+                    f"   Missing {key=}, in {feat_name=}, in {stats_dict[feat_name].keys()}"
+                )
 
     def test_invalid_input(self, stats):
         with pytest.raises(TypeError):
@@ -226,45 +285,136 @@ class Test_Stats:
         stats.clear_statistics()
[…]
