+++ src/plaid/containers/dataset.py
@@ -1064,11 +1064,12 @@ class Dataset(object):
             This function can also be called using `__call__()`.
         """
         if isinstance(id, slice) or isinstance(id, list) or isinstance(id, np.ndarray):
+            print(">>>", type(id))
             if isinstance(id, slice):
                 id = list(range(*id.indices(len(self))))
             dataset = Dataset()
             for i in id:
-                dataset.add_sample(self[i], i)
+                dataset.add_sample(self[int(i)], int(i))
             dataset.set_infos(self.get_infos())
             return dataset
         else:
