### 2025-06-30 5ec1d01
feat(tests) reduce samples list in autotests to speedup pytest runs
