### 2025-06-30 02cf15d
feat(tests) improve coverage of huggingface_bridge
