### 2025-06-30 c449a92
feat(huggingface_bridge) improve coverage
