### 2025-06-30 aadbc61
feat(examples/pipelines) remove wip files
