+++ examples/pipelines/experiments/persistent_node.py
@@ -1,448 +1,0 @@
-from sklearn.base import BaseEstimator, RegressorMixin
-import joblib
-import json
-from pathlib import Path
-import hashlib
-import numpy as np
-
-
-class PersistentNode(BaseEstimator):
-    def __init__(self, save_path="models/unnamed.joblib"):
-        self.save_path = save_path
-        self._path = Path(save_path)
-        self._meta_path = self._path.with_suffix('.meta.json')
-        self.fitted_ = False
-        self.model = None
-
-    def _save_file(self, obj, input_hash=None):
-        self._path.parent.mkdir(parents=True, exist_ok=True)
-        joblib.dump(obj, self._path)
-        if input_hash is not None:
-            self._meta_path.write_text(json.dumps({'hash': input_hash}))
-        self.fitted_ = True
-
-    def save_model_with_hash(self, obj, data):
-        input_hash = self.hash_input(data)
-        self._save_file(obj, input_hash=input_hash)
-
-    def load(self):
-        obj = joblib.load(self._path)
-        self.fitted_ = True
-        return obj
-
-    def exists(self):
-        return self._path.exists() and self._meta_path.exists()
-
-    def get_stored_hash(self):
-        if self._meta_path.exists():
-            return json.loads(self._meta_path.read_text()).get('hash')
-        return None
-
-    def hash_input(self, data):
-        m = hashlib.sha256()
-        if isinstance(data, tuple):
-            for arr in data:
-                m.update(np.ascontiguousarray(arr).data)
-        else:
-            m.update(np.ascontiguousarray(data).data)
-        return m.hexdigest()
-
-    def check_fitted_or_load(self):
-        if self.fitted_:
-            return
-        if self.exists():
-            loaded = self.load()
-            self.set_model(loaded)
-        else:
-            raise ValueError("Model not fitted or not saved.")
-
-    def set_model(self, obj):
-        self.model = obj
-
-    def get_model(self):
-        return self.model
-
-    def load_if_cached(self, data):
-        input_hash = self.hash_input(data)
-        if self.exists() and self.get_stored_hash() == input_hash:
-            self.set_model(self.load())
-            return True
-        return False
-
-    def get_params(self, deep=True):
-        params = super().get_params(deep=deep)
-        if deep and hasattr(self.model, "get_params"):
-            for k, v in self.model.get_params(deep=True).items():
-                params[f"model__{k}"] = v
-        return params
-
-    def set_params(self, **params):
-        model_params = {k[7:]: v for k, v in params.items() if k.startswith("model__")}
-        node_params = {k: v for k, v in params.items() if not k.startswith("model__")}
-        if node_params:
-            super().set_params(**node_params)
-        if model_params and self.model is not None:
-            self.model.set_params(**model_params)
-        return self
-
-
-
-from sklearn.base import BaseEstimator, RegressorMixin
-from sklearn.model_selection import GridSearchCV
-from sklearn.datasets import make_regression
-from sklearn.metrics import mean_squared_error
-from pipefunc import pipefunc, Pipeline
-import numpy as np
-import inspect
-from typing import Dict, Any, Optional
-
-class PipefuncSklearnWrapper(BaseEstimator, RegressorMixin):
-    """
-    Scikit-learn compatible wrapper for pipefunc Pipeline objects.
-
-    This wrapper allows pipefunc pipelines to be used with scikit-learn tools
-    like GridSearchCV, cross_val_score, etc.
-    """
-
-    def __init__(self, pipeline: Pipeline, prediction_output: str = "predictions", **kwargs):
-        """
-        Initialize the wrapper.
-
-        Parameters:
-        -----------
-        pipeline : pipefunc.Pipeline
-            The pipefunc pipeline to wrap
-        prediction_output : str
[…]
