+++ examples/pipelines/with_persistence/ml_pipeline_nodes.py
@@ -1,0 +1,424 @@
+import os
+import json
+import joblib
+from joblib import Parallel, delayed
+import numpy as np
+from pathlib import Path
+from sklearn.base import BaseEstimator, RegressorMixin, TransformerMixin
+from sklearn.linear_model import Ridge
+from sklearn.decomposition import PCA
+
+
+# def flatten_dict(d, prefix=''):
+#     items = {}
+#     for k, v in d.items():
+#         new_key = f'{prefix}__{k}' if prefix else k
+#         if isinstance(v, dict):
+#             items.update(flatten_dict(v, new_key))
+#         else:
+#             items[f'param__{new_key}'] = v
+#     return items
+
+# def unflatten_dict(flat):
+#     nested = {}
+#     for flat_key, value in flat.items():
+#         keys = flat_key.split('__')
+#         d = nested
+#         for k in keys[:-1]:
+#             d = d.setdefault(k, {})
+#         d[keys[-1]] = value
+#     return nested
+
+
+class PersistentNode(BaseEstimator, RegressorMixin, TransformerMixin):
+    def __init__(self, name, global_params, params):
+        self.name = name
+        self.global_params = global_params
+        self.params = params
+        self.save_path = Path(os.path.join(global_params['save_path'], f"{name}.joblib"))
+        self.fitted_ = False
+        self.model = None
+
+    # def get_params(self, deep=True):
+    #     return flatten_dict(self.params[self.name])
+
+    # def set_params(self, **kwargs):
+    #     if 'name' in kwargs:
+    #         self.name = kwargs.pop('name')
+    #     if 'params' in kwargs:
+    #         self.params.update(kwargs.pop('params'))
+
+    #     # Extract param__ keys and merge into self.params
+    #     flat = {}
+    #     for k, v in kwargs.items():
+    #         if k.startswith('param__'):
+    #             flat_key = k[len('param__'):]
+    #             flat[flat_key] = v
+    #     nested_update = unflatten_dict(flat)
+    #     self.params = self._deep_merge(self.params, nested_update)
+    #     return self
+
+    def save(self, obj):
+        self.save_path.parent.mkdir(parents=True, exist_ok=True)
+        joblib.dump(obj, self.save_path)
+        self.fitted_ = True
+
+    def load(self):
+        print(f"Loading existing model from {self.save_path}")
+        obj = joblib.load(self.save_path)
+        self.fitted_ = True
+        return obj
+
+    def exists(self):
+        return self.save_path.exists()
+
+    def check_fitted_or_load(self):
+        if self.fitted_:
+            return
+        if self.exists():
+            self.set_model(self.load())
+        else:
+            raise ValueError("Model not fitted and no saved model found.")
+
+    def set_model(self, obj):
+        self.model = obj
+
+    def fit(self, X, y=None):
+        if self.exists():
+            self.set_model(self.load())
+            return self
+        self._fit(X, y)
+        self.save(self.model)
+        return self
+
+    def transform(self, X):
+        self.check_fitted_or_load()
+        return self._transform(X)
+
+    def inverse_transform(self, X):
+        self.check_fitted_or_load()
+        return self._inverse_transform(X)
+
+    def predict(self, X):
+        self.check_fitted_or_load()
+        return self._predict(X)
+
+    def score(self, X, y):
+        self.check_fitted_or_load()
+        return self._score(X, y)
+
+    # Protected methods to override in subclasses
+    def _fit(self, X, y=None):
+        raise NotImplementedError
+
+    def _predict(self, X):
+        raise NotImplementedError
+
+    def _transform(self, X):
+        raise NotImplementedError
+
[…]
