### 2025-06-30 16c7268
fix(huggingface_bridge) coverage: ignore line not reported with multiprocessing
