### 2025-06-30 9f23488
fix(examples/pipelines) remove comment
