+++ examples/pipelines/sklearn_pipeline.py
@@ -1,21 +1,18 @@
-from datasets import load_dataset
-from plaid.containers.sample import Sample
-from plaid.bridges.huggingface_bridge import huggingface_dataset_to_plaid, huggingface_description_to_problem_definition
-import os, pickle
-import numpy as np
-from safetensors.numpy import save_file
-from sklearn.base import BaseEstimator, RegressorMixin
-from sklearn.model_selection import GridSearchCV
+# -*- coding: utf-8 -*-
+#
+# This file is subject to the terms and conditions defined in
+# file 'LICENSE.txt', which is part of this source code package.
+#
+#
 
-from sklearn.pipeline import Pipeline
-from sklearn.preprocessing import StandardScaler
-
-from pathlib import Path
-import joblib
-from copy import copy
-
+import os
 import yaml
 import time
+
+from datasets import load_dataset
+from plaid.bridges.huggingface_bridge import huggingface_dataset_to_plaid, huggingface_description_to_problem_definition
+from sklearn.model_selection import GridSearchCV
+from sklearn.pipeline import Pipeline
 
 from ml_pipeline_nodes import ScalarScalerNode, GPRegressorNode, PCAEmbeddingNode
 
