### 2025-06-30 618ef5e
Add types package and constants module
