### 2025-06-30 7881cda
(stats) fix wrong shape of stats + improve test
