### 2025-06-30 4cc2e01
TypeAlias for python3.9
