### 2025-06-30 fc8db2a
int/float arrays
