+++ src/plaid/types/common.py
@@ -18,8 +18,8 @@ import numpy as np
 from numpy.typing import NDArray
 
 # A generic float array type (float32 or float64)
-ArrayDType = Union[np.float32, np.float64]
+ArrayDType = Union[np.int32, np.int64, np.float32, np.float64]
 Array: TypeAlias = NDArray[ArrayDType]
 
 # Types used in indexing operations
-IndexType = Union[list[int], Array]
+IndexType = Union[list[int], NDArray[Union[np.int32, np.int64]]]
