+++ src/plaid/constants.py
@@ -1,5 +1,12 @@
 """Common constants used across the PLAID library."""
 
+# -*- coding: utf-8 -*-
+#
+# This file is subject to the terms and conditions defined in
+# file 'LICENSE.txt', which is part of this source code package.
+#
+#
+
 from typing import Literal
 
 AUTHORIZED_TASKS = Literal["regression", "classification"]
