### 2025-07-01 06526bc
📄 feat(docs) replace data challenges section with plaid benchmarks (#98)
* feat(docs) replace data challenges section with plaid benchmarks

* feat(docs) add pypi download badge

* fix(docs/source/plaid_benchmarks.rst) writing

---------

Co-authored-by: Fabien Casenave <fabien.casenave@safrangroup.com>
