### 2025-07-01 53f00cb
fix(pipeline) configure parallel conversion in example and remove time arg in PCAEmbeddingNode
