### 2025-07-01 df0cb1f
feat(examples/pipelines) simplify arguments in pipeline nodes, by giving a global dict and specifying only arguments to be optimized by GridSearchCV (n_components of PCA for the moment)
