### 2025-07-01 f0e081c
feat(examples/pipelines) add other pipeline definition with all arguments specified in config.yml
