### 2025-07-01 06f9967
constants docstrings
