### 2025-07-01 298de71
(stats) fix internal array dim check
