### 2025-07-01 67215eb
:recycle: Add types package and constants module (#102)
