+++ .github/workflows/testing.yml
@@ -5,7 +5,11 @@ permissions:
 
 on:
   push:
+    branches: 
+      - '**'
   pull_request:
+    branches: 
+      - '**'
   schedule:
     - cron: '0 0 * * *'  # Every day at 00:00 UTC
 
@@ -13,6 +17,7 @@ jobs:
   test:
     name: Run Tests and Examples
     runs-on: ${{ matrix.os }}
+    if: github.event_name != 'pull_request' || github.event.pull_request.head.repo.full_name != github.event.pull_request.base.repo.full_name
     strategy:
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
