### 2025-07-01 7ec655b
📦 add python313 support (#99)
* feat(pyproject.toml/testing action) add python 3.13 support

* fix(pyproject.toml) upper bound in python version
