### 2025-07-07 13d7819
(containers) minor fix docstrings
