### 2025-07-07 07c115b
feat(sample.py) add an attribute _extra_data
