### 2025-07-07 e00877c
feat(dataset.py) get_scalars_to_tabular create array with dtype depending on data
