### 2025-07-07 bd5600f
feat(dataset.py) improve __getitem__ to work with slices
