### 2025-07-07 776b4e8
feat(dataset.py) add method extract_dataset to extract a dataset with samples containing only requiered scalars/fields/timeseries(/nodes ?)
