### 2025-07-07 52ff991
feat(dataset.py) add method merge_samples to merge scalars/fields/trees/time_series of of samples if they have same ids
