### 2025-07-07 70962b2
feat(dataset.py) add methods to work with tabular fields the same way as for scalars
