### 2025-07-12 2dc9f54
(sklearn wrapper) add classes to wrap any sklearn block to use and return a PLAID dataset
