+++ src/plaid/wrappers/sklearn.py
@@ -1,0 +1,385 @@
+# -*- coding: utf-8 -*-
+#
+# This file is subject to the terms and conditions defined in
+# file 'LICENSE.txt', which is part of this source code package.
+#
+#
+"""This module provides wrappers for scikit-learn estimators and transformers so they can be used seamlessly in scikit-learn Pipelines
+with PLAID objects. The wrapped blocks (e.g. PCA, GaussianProcessRegressor, StandardScaler, etc.) take a `plaid.containers.Dataset` as input,
+and return a `plaid.containers.Dataset` as output. This allows you to build
+scikit-learn Pipelines where all blocks operate on PLAID objects, enabling end-to-end workflows with domain-specific data structures.
+
+Example usage:
+
+    from sklearn.pipeline import Pipeline
+    from plaid.wrappers.sklearn import WrappedSklearnTransform, WrappedSklearnRegressor
+    from sklearn.decomposition import PCA
+    from sklearn.gaussian_process import GaussianProcessRegressor
+    from plaid.containers.dataset import Dataset
+
+    # Define your PLAID dataset
+    dataset = Dataset(...)
+
+    # Build a pipeline with wrapped sklearn blocks
+    pipe = Pipeline([
+        ("pca", WrappedSklearnTransform(PCA(n_components=2))),
+        ("reg", WrappedSklearnRegressor(GaussianProcessRegressor()))
+    ])
+
+    # Fit the pipeline (all steps receive and return Dataset objects)
+    pipe.fit(dataset)
+    # Predict
+    y_pred = pipe.predict(dataset)
+
+All wrapped blocks must accept and return PLAID Dataset objects.
+
+Some inspiration come from [TensorDict](https://pytorch.org/tensordict/stable/reference/generated/tensordict.nn.TensorDictModule.html#tensordict.nn.TensorDictModule).
+
+This module defines the following classes:
+`PlaidWrapper`: Base class for scikit-learn estimators and transformers to operate on PLAID objects.
+├── `WrappedSklearnTransform`: Wrapper for scikit-learn Transformer blocks.
+└── `WrappedSklearnPredictor`: Wrapper for scikit-learn Predictor blocks.
+    ├── `WrappedSklearnClassifier`: Wrapper for scikit-learn Classifier blocks.
+    └── `WrappedSklearnRegressor`: Wrapper for scikit-learn Regressor blocks.
+"""
+
+# %% Imports
+import sys
+
+if sys.version_info >= (3, 11):
+    from typing import Self
+else:  # pragma: no cover
+    from typing import TypeVar
+
+    Self = TypeVar("Self")
+
+import logging
+from typing import Union
+
+import numpy as np
+from sklearn.base import (
+    BaseEstimator,
+    BiclusterMixin,
+    ClassifierMixin,
+    ClusterMixin,
+    DensityMixin,
+    MetaEstimatorMixin,
+    MultiOutputMixin,
+    OutlierMixin,
+    RegressorMixin,
+    TransformerMixin,
+)
+
+from plaid.containers.dataset import Dataset
+
+logger = logging.getLogger(__name__)
+logging.basicConfig(
+    format="[%(asctime)s:%(levelname)s:%(filename)s:%(funcName)s(%(lineno)d)]:%(message)s",
+    level=logging.INFO,
+)
+
+# %% Classes
+
+SklearnBlock = Union[
+    BaseEstimator,
+    TransformerMixin,
+    RegressorMixin,
+    ClassifierMixin,
+    ClusterMixin,
+    BiclusterMixin,
+    DensityMixin,
+    OutlierMixin,
+    MultiOutputMixin,
+]
+"""Union type for all scikit-learn blocks that can be used in a Pipeline."""
+
+
[…]
