### 2025-07-12 a660b51
(dataset) fix __getitem__ with slices
