### 2025-07-16 7057937
(sample.py) add check that names does not contain /
