+++ src/plaid/containers/sample.py
@@ -119,6 +119,10 @@ TimeSeriesType = tuple[TimeSequenceType, FieldType]
 """A TimeSeriesType is a tuple[TimeSequenceType,FieldType]
 """
 
+def check_names(names:list[str]):
+    for name in names:
+        if '/' in name:
+            raise ValueError(f"feature_names containing `/` are not allowed, but {name=}, you should first replace any occurence of `/` with something else, for example: `name.replace('/','  ')`")
 
 def read_index(pyTree: list, dim: list[int]):
     """Read Index Array or Index Range from CGNS.
@@ -872,6 +876,8 @@ class Sample(BaseModel):
         Returns:
             CGNSNode: The created Base node.
         """
+        check_names([base_name])
+
         time = self.get_time_assignment(time)
 
         if base_name is None:
@@ -1014,6 +1020,8 @@ class Sample(BaseModel):
         Returns:
             CGLNode: The newly initialized zone node within the CGNS tree.
         """
+        check_names([zone_name])
+
         # init_tree will look for default time
         self.init_tree(time)
         # get_base will look for default base_name and time
@@ -1201,6 +1209,7 @@ class Sample(BaseModel):
             name (str): The name of the scalar value.
             value (ScalarType): The scalar value to add or update in the dictionary.
         """
+        check_names([name])
         if self._scalars is None:
             self._scalars = {name: value}
         else:
@@ -1274,6 +1283,7 @@ class Sample(BaseModel):
         Raises:
             TypeError: Raised if the length of `time_sequence` is not equal to the length of `values`.
         """
+        check_names([name])
         assert len(time_sequence) == len(values), (
             "time sequence and values do not have the same size"
         )
@@ -1633,6 +1643,7 @@ class Sample(BaseModel):
         Raises:
             KeyError: Raised if the specified zone does not exist in the given base.
         """
+        check_names([name])
         # init_tree will look for default time
         self.init_tree(time)
         # get_zone will look for default zone_name, base_name and time
