### 2025-07-16 0d1b97a
feat(dataset.py) add methods to handle time_series
