+++ src/plaid/containers/dataset.py
@@ -353,6 +353,28 @@ class Dataset(object):
         return scalars_names
 
     # -------------------------------------------------------------------------#
+    def get_time_series_names(self, ids: list[int] = None) -> list[str]:
+        """Return union of time series names in all samples with id in ids.
+
+        Args:
+            ids (list[int], optional): Select time_series depending on sample id. If None, take all samples. Defaults to None.
+
+        Returns:
+            list[str]: List of all time_series names
+        """
+        if ids is not None and len(set(ids)) != len(ids):
+            logger.warning("Provided ids are not unique")
+
+        time_series_names = []
+        for sample in self.get_samples(ids, as_list=True):
+            s_names = sample.get_time_series_names()
+            for s_name in s_names:
+                if s_name not in time_series_names:
+                    time_series_names.append(s_name)
+        time_series_names.sort()
+        return time_series_names
+
+    # -------------------------------------------------------------------------#
     def get_field_names(
         self, ids: list[int] = None, zone_name: str = None, base_name: str = None
     ) -> list[str]:
@@ -461,7 +483,135 @@ class Dataset(object):
             named_tabular[s_name] = res
 
         if as_nparray:
-            named_tabular = np.array(list(named_tabular.values())).T
+            named_tabular = np.array(
+                [named_tabular[s_name] for s_name in scalar_names]
+            ).T
+        return named_tabular
+
+    # -------------------------------------------------------------------------#
+    def add_tabular_time_series(
+        self, tabular: np.ndarray, names: list[str] = None
+    ) -> None:
+        """Add tabular time_series data to the summary.
+
+        Args:
+            tabular (np.ndarray): A 2D NumPy array containing tabular time_series data, with shape (n_samples, max_n_timestamps, n_features, 2). Last dimension should contain (timestamp, value) pairs. Timestamps should be in strictly increasing order, and padded with decreasing values so that all features have same length.
+            names (list[str], optional): A list of column names for the tabular data. Defaults to None.
+
+        Raises:
+            ShapeError: Raised if the input tabular array does not have the correct shape (2D).
+            ShapeError: Raised if the number of columns in the tabular data does not match the number of names provided.
+
+        Note:
+            If no names are provided, it will automatically create names based on the pattern 'X{number}'
+        """
+        nb_samples = len(tabular)
+
+        if tabular.ndim != 4:
+            raise ShapeError(f"{tabular.ndim=}!=4, should be == 4")
+        if names is None:
+            names = [f"X{i}" for i in range(tabular.shape[2])]
+        if tabular.shape[2] != len(names):
+            raise ShapeError(
+                f"tabular's 3rd dimension should have same size as there are names, but {tabular.shape[2]=} and {len(names)=}"
+            )
+
+        # ---# For efficiency, first add values to storage
+        name_to_ids = {}
+        for i_col, name in enumerate(names):
+            name_to_ids[name] = tabular[:, :, i_col]
+
+        # ---# Then add data in sample
+        for i_samp in range(nb_samples):
+            sample = Sample()
+            for i_name, name in enumerate(names):
+                timestamps = name_to_ids[name][i_samp, :, 0]
+                values = name_to_ids[name][i_samp, :, 1]
+                # detect first decreasing timestamp
+                diff_ts = timestamps[1:] - timestamps[:-1]
+                decreasing_ts_ids = (diff_ts <= 0).nonzero()[0]
+                if len(decreasing_ts_ids) > 0:
+                    first_decreasing_ts_id = decreasing_ts_ids[0]
+                else:
+                    first_decreasing_ts_id = len(timestamps) - 1
+                # add only values and timestamps up to the first decreasing timestamp
[…]
