+++ src/plaid/wrappers/sklearn.py
@@ -54,6 +54,7 @@ else:  # pragma: no cover
     Self = TypeVar("Self")
 
 import logging
+from copy import copy
 from typing import Union
 
 import numpy as np
@@ -106,55 +107,29 @@ class PlaidWrapper(BaseEstimator, MetaEstimatorMixin):
         self,
         sklearn_block: SklearnBlock,
         fit_only_ones: bool = True,
-        in_keys: Union[list[str], str] = [],
-        out_keys: Union[list[str], str] = [],
+        in_features: Union[list[str], str] = [],
+        out_features: Union[list[str], str] = [],
     ):
         """Wrap a scikit-learn estimator or transformer.
 
         Args:
             sklearn_block (SklearnBlock): Any scikit-learn transform or predictor (e.g. PCA, StandardScaler, GPRegressor).
             fit_only_ones (bool, optional): If True, the model will only be fitted once. Defaults to True.
-            in_keys (Union[list[str],str], optional):
+            in_features (Union[list[str],str], optional):
                 Names of scalars and/or fields to take as input.
                 Scalar (resp. field) names should be given as 'scalar::<scalar-name>' (resp. 'field::<field-name>').
                 Use 'all' to use all available scalars and fields, or 'scalar::all'/'field::all' for all scalars/fields.
                 Defaults to [].
-            out_keys (Union[list[str],str], optional): Names of scalars and/or fields to take as output, using the same convention as for `in_keys`. Defaults to [].
+            out_features (Union[list[str],str], optional): Names of scalars and/or fields to take as output, using the same convention as for `in_features`. Defaults to [].
                 Additionally, if 'same', 'scalar::same' or 'field::same' is given, it will use as output the same names as for input.
         """
+        # TODO: check https://scikit-learn.org/stable/developers/develop.html#instantiation
         self.sklearn_block = sklearn_block
         self.fit_only_ones = fit_only_ones
-        self.in_keys = in_keys
-        self.out_keys = out_keys
+        self.in_features = copy(in_features)
+        self.out_features = copy(out_features)
 
-        # ---# Scalars
-        if in_keys == "all" or "scalar::all" in in_keys:
-            self.input_scalars = "all"
-        else:
-            self.input_scalars = [s[8:] for s in in_keys if s[:8] == "scalar::"]
-        #
-        if out_keys == "same" or "scalar::same" in out_keys:
-            self.output_scalars = self.input_scalars
-        else:
-            self.output_scalars = [s[8:] for s in out_keys if s[:8] == "scalar::"]
-
-        # ---# Fields
-        if in_keys == "all" or "field::all" in in_keys:
-            self.input_fields = "all"
-        else:
-            self.input_fields = [s[7:] for s in in_keys if s[:7] == "field::"]
-        #
-        if out_keys == "same" or "field::same" in out_keys:
-            self.output_fields = self.input_fields
-        else:
-            self.output_fields = [s[7:] for s in out_keys if s[:7] == "field::"]
-
-        print(f"{self.input_scalars=}")
-        print(f"{self.input_fields=}")
-        print(f"{self.output_scalars=}")
-        print(f"{self.output_fields=}")
-
-    def fit(self, dataset: Dataset, *args, **kwargs):
+    def fit(self, dataset: Dataset, **kwargs):
         """Fit the wrapped scikit-learn model on a PLAID dataset.
 
         Args:
@@ -166,11 +141,104 @@ class PlaidWrapper(BaseEstimator, MetaEstimatorMixin):
         if self.fit_only_ones and self.__sklearn_is_fitted__():
             return self
 
+        self._determine_input_output_names(dataset)
+
[…]
