### 2025-07-17 17d4f00
feat(dataset.py) add __iter__ fonction to class
