### 2025-07-17 ef43e4d
🐛 feat(dataset.py) add __iter__ fonction to class to correct bugged for loops over datasets (#109)
* feat(dataset.py) add __iter__ fonction to class

* feat(test_datasets) improve coverage

* feat(test_datasets) improve coverage
