### 2025-07-18 f786a6e
(sample.py) fix check_names
