+++ src/plaid/containers/sample.py
@@ -119,10 +119,14 @@ TimeSeriesType = tuple[TimeSequenceType, FieldType]
 """A TimeSeriesType is a tuple[TimeSequenceType,FieldType]
 """
 
-def check_names(names:list[str]):
+
+def _check_names(names: list[str]):
     for name in names:
-        if '/' in name:
-            raise ValueError(f"feature_names containing `/` are not allowed, but {name=}, you should first replace any occurence of `/` with something else, for example: `name.replace('/','  ')`")
+        if (name is not None) and ("/" in name):
+            raise ValueError(
+                f"feature_names containing `/` are not allowed, but {name=}, you should first replace any occurence of `/` with something else, for example: `name.replace('/','__')`"
+            )
+
 
 def read_index(pyTree: list, dim: list[int]):
     """Read Index Array or Index Range from CGNS.
@@ -876,7 +880,7 @@ class Sample(BaseModel):
         Returns:
             CGNSNode: The created Base node.
         """
-        check_names([base_name])
+        _check_names([base_name])
 
         time = self.get_time_assignment(time)
 
@@ -1020,7 +1024,7 @@ class Sample(BaseModel):
         Returns:
             CGLNode: The newly initialized zone node within the CGNS tree.
         """
-        check_names([zone_name])
+        _check_names([zone_name])
 
         # init_tree will look for default time
         self.init_tree(time)
@@ -1209,7 +1213,7 @@ class Sample(BaseModel):
             name (str): The name of the scalar value.
             value (ScalarType): The scalar value to add or update in the dictionary.
         """
-        check_names([name])
+        _check_names([name])
         if self._scalars is None:
             self._scalars = {name: value}
         else:
@@ -1283,7 +1287,7 @@ class Sample(BaseModel):
         Raises:
             TypeError: Raised if the length of `time_sequence` is not equal to the length of `values`.
         """
-        check_names([name])
+        _check_names([name])
         assert len(time_sequence) == len(values), (
             "time sequence and values do not have the same size"
         )
@@ -1643,7 +1647,7 @@ class Sample(BaseModel):
         Raises:
             KeyError: Raised if the specified zone does not exist in the given base.
         """
-        check_names([name])
+        _check_names([name])
         # init_tree will look for default time
         self.init_tree(time)
         # get_zone will look for default zone_name, base_name and time
