### 2025-07-18 e4973fa
(sample.py) fix check_names for str argument
