### 2025-07-18 2b58648
(sample.py) fix check_names tests
