+++ tests/containers/test_dataset.py
@@ -74,6 +74,13 @@ def dataset_with_samples(dataset, samples, infos):
 
 
 @pytest.fixture()
+def dataset_with_samples_with_tree(empty_dataset, samples_with_tree, infos):
+    empty_dataset.add_samples(samples_with_tree)
+    empty_dataset.set_infos(infos)
+    return empty_dataset
+
+
+@pytest.fixture()
 def other_dataset_with_samples(other_samples):
     other_dataset = Dataset()
     other_dataset.add_samples(other_samples)
@@ -362,6 +369,128 @@ class Test_Dataset:
         dataset.get_scalars_to_tabular(scalar_names=["test", "test"])
 
     # -------------------------------------------------------------------------#
+    def test_get_feature_from_string_identifier(
+        self, dataset_with_samples, dataset_with_samples_with_tree
+    ):
+        dataset_with_samples.get_feature_from_string_identifier("scalar::test_scalar")
+
+        dataset_with_samples.get_feature_from_string_identifier(
+            "time_series::test_time_series_1"
+        )
+
+        dataset_with_samples.get_feature_from_string_identifier(
+            "field::test_field_same_size"
+        )
+        dataset_with_samples.get_feature_from_string_identifier(
+            "field::test_field_same_size/TestBaseName"
+        )
+        dataset_with_samples.get_feature_from_string_identifier(
+            "field::test_field_same_size/TestBaseName/TestZoneName"
+        )
+        dataset_with_samples.get_feature_from_string_identifier(
+            "field::test_field_same_size/TestBaseName/TestZoneName/Vertex"
+        )
+        dataset_with_samples.get_feature_from_string_identifier(
+            "field::test_field_same_size/TestBaseName/TestZoneName/Vertex/0"
+        )
+
+        dataset_with_samples_with_tree.get_feature_from_string_identifier("nodes::")
+        dataset_with_samples_with_tree.get_feature_from_string_identifier(
+            "nodes::Base_2_2"
+        )
+        dataset_with_samples_with_tree.get_feature_from_string_identifier(
+            "nodes::Base_2_2/Zone"
+        )
+        dataset_with_samples_with_tree.get_feature_from_string_identifier(
+            "nodes::Base_2_2/Zone/0"
+        )
+
+    def test_get_feature_from_identifier(
+        self, dataset_with_samples, dataset_with_samples_with_tree
+    ):
+        dataset_with_samples.get_feature_from_identifier(
+            {"type": "scalar", "name": "test_scalar"}
+        )
+        dataset_with_samples.get_feature_from_identifier(
+            {"type": "time_series", "name": "test_time_series_1"}
+        )
+
+        dataset_with_samples.get_feature_from_identifier(
+            {"type": "field", "name": "test_field_same_size"}
+        )
+        dataset_with_samples.get_feature_from_identifier(
+            {
+                "type": "field",
+                "name": "test_field_same_size",
+                "base_name": "TestBaseName",
+            }
+        )
+        dataset_with_samples.get_feature_from_identifier(
+            {
+                "type": "field",
+                "name": "test_field_same_size",
+                "base_name": "TestBaseName",
+                "zone_name": "TestZoneName",
+            }
+        )
+        dataset_with_samples.get_feature_from_identifier(
+            {
+                "type": "field",
+                "name": "test_field_same_size",
+                "base_name": "TestBaseName",
+                "zone_name": "TestZoneName",
+                "location": "Vertex",
+            }
+        )
+        dataset_with_samples.get_feature_from_identifier(
+            {
+                "type": "field",
+                "name": "test_field_same_size",
+                "base_name": "TestBaseName",
+                "zone_name": "TestZoneName",
+                "location": "Vertex",
+                "time": 0.0,
[…]
