### 2025-07-18 a600847
feat(test_dataset.py) improve coverage
