### 2025-07-18 f01422c
🎉 feat(sample.py) add get_feature_from_string_identifier and get_feature_from_identifier (#110)
* feat(sample.py) add get_feature_from_string_identifier and get_feature_from_identifier that provide unified way to retrieve a feature from a unique identifier

* feat(dataset.py) add get_feature_from_identifier and get_feature_from_string_identifier

* feat(test_dataset.py) improve coverage

---------

Co-authored-by: Fabien Casenave <fabien.casenave@safrangroup.com>
