+++ tests/containers/test_dataset.py
@@ -74,6 +74,13 @@ def dataset_with_samples(dataset, samples, infos):
 
 
 @pytest.fixture()
+def dataset_with_samples_with_tree(empty_dataset, samples_with_tree, infos):
+    empty_dataset.add_samples(samples_with_tree)
+    empty_dataset.set_infos(infos)
+    return empty_dataset
+
+
+@pytest.fixture()
 def other_dataset_with_samples(other_samples):
     other_dataset = Dataset()
     other_dataset.add_samples(other_samples)
@@ -316,6 +323,16 @@ class Test_Dataset:
         dataset_with_samples.get_scalar_names()
         dataset_with_samples.get_scalar_names([0, 0])
 
+    def test_get_time_series_names(self, dataset_with_samples, nb_samples):
+        dataset_with_samples.get_time_series_names()
+        dataset_with_samples.get_time_series_names(
+            np.random.randint(2, nb_samples, size=2)
+        )
+
+    def test_get_time_series_names_same_ids(self, dataset_with_samples):
+        dataset_with_samples.get_time_series_names()
+        dataset_with_samples.get_time_series_names([0, 0])
+
     def test_get_field_names(self, dataset_with_samples, nb_samples):
         dataset_with_samples.get_field_names()
         dataset_with_samples.get_field_names(np.random.randint(2, nb_samples, size=2))
@@ -360,6 +377,128 @@ class Test_Dataset:
         )
         dataset.get_scalars_to_tabular(sample_ids=[0, 0])
         dataset.get_scalars_to_tabular(scalar_names=["test", "test"])
+
+    # -------------------------------------------------------------------------#
+    def test_get_feature_from_string_identifier(
+        self, dataset_with_samples, dataset_with_samples_with_tree
+    ):
+        dataset_with_samples.get_feature_from_string_identifier("scalar::test_scalar")
+
+        dataset_with_samples.get_feature_from_string_identifier(
+            "time_series::test_time_series_1"
+        )
+
+        dataset_with_samples.get_feature_from_string_identifier(
+            "field::test_field_same_size"
+        )
+        dataset_with_samples.get_feature_from_string_identifier(
+            "field::test_field_same_size/TestBaseName"
+        )
+        dataset_with_samples.get_feature_from_string_identifier(
+            "field::test_field_same_size/TestBaseName/TestZoneName"
+        )
+        dataset_with_samples.get_feature_from_string_identifier(
+            "field::test_field_same_size/TestBaseName/TestZoneName/Vertex"
+        )
+        dataset_with_samples.get_feature_from_string_identifier(
+            "field::test_field_same_size/TestBaseName/TestZoneName/Vertex/0"
+        )
+
+        dataset_with_samples_with_tree.get_feature_from_string_identifier("nodes::")
+        dataset_with_samples_with_tree.get_feature_from_string_identifier(
+            "nodes::Base_2_2"
+        )
+        dataset_with_samples_with_tree.get_feature_from_string_identifier(
+            "nodes::Base_2_2/Zone"
+        )
+        dataset_with_samples_with_tree.get_feature_from_string_identifier(
+            "nodes::Base_2_2/Zone/0"
+        )
+
+    def test_get_feature_from_identifier(
+        self, dataset_with_samples, dataset_with_samples_with_tree
+    ):
+        dataset_with_samples.get_feature_from_identifier(
+            {"type": "scalar", "name": "test_scalar"}
+        )
+        dataset_with_samples.get_feature_from_identifier(
+            {"type": "time_series", "name": "test_time_series_1"}
+        )
+
+        dataset_with_samples.get_feature_from_identifier(
+            {"type": "field", "name": "test_field_same_size"}
+        )
[…]
