### 2025-07-18 403261b
add optuna notebook
