### 2025-07-19 fbf658e
feat(sample.py) add function to update a sample from features or create a new sample by extracting features
