### 2025-07-19 f24ae53
fix(sample.py) typing correction for python3.9
