### 2025-07-19 b5b843f
🎉 Dataset and sample utilities for updating and extracting from features  (#112)
* temp

* feat(sample.py) add function to update a sample from features or create a new sample by extracting features

* fix(sample.py) typing correction for python3.9
