+++ src/plaid/containers/sample.py
@@ -31,7 +31,12 @@ import numpy as np
 from CGNS.PAT.cgnsutils import __CHILDREN__, __NAME__
 from pydantic import BaseModel, model_serializer
 
-from plaid.constants import CGNS_ELEMENT_NAMES
+from plaid.constants import (
+    AUTHORIZED_FEATURE_INFOS,
+    AUTHORIZED_FEATURE_TYPES,
+    AUTHORIZED_FIELD_LOCATIONS,
+    CGNS_ELEMENT_NAMES,
+)
 from plaid.types import (
     CGNSNode,
     CGNSTree,
@@ -131,6 +136,48 @@ def read_index_range(pyTree: list, dim: list[int]):
     return np.array(res, dtype=int).ravel()
 
 
+def _get_feature_type_and_details_from_identifier(
+    feature_identifier: dict[str, Union[str, float]],
+) -> tuple[str, dict[str, Union[str, float]]]:
+    """Extract and validate the feature type and its associated metadata from a feature identifier.
+
+    This utility function ensures that the `feature_identifier` dictionary contains a valid
+    "type" key (e.g., "scalar", "time_series", "field", "node") and returns the type along
+    with the remaining identifier keys, which are specific to the feature type.
+
+    Args:
+        feature_identifier (dict): A dictionary with a "type" key, and
+            other keys (some optional) depending on the feature type. For example:
+                - {"type": "scalar", "name": "Mach"}
+                - {"type": "time_series", "name": "AOA"}
+                - {"type": "field", "name": "pressure"}
+
+    Returns:
+        tuple[str, dict]: A tuple `(feature_type, feature_details)` where:
+            - `feature_type` is the value of the "type" key (e.g., "scalar").
+            - `feature_details` is a dictionary of the remaining keys.
+
+    Raises:
+        AssertionError:
+            - If "type" is missing.
+            - If the type is not in `AUTHORIZED_FEATURE_TYPES`.
+            - If any unexpected keys are present for the given type.
+    """
+    assert "type" in feature_identifier, (
+        "feature type not specified in feature_identifier"
+    )
+    feature_type = feature_identifier["type"]
+    feature_details = {k: v for k, v in feature_identifier.items() if k != "type"}
+
+    assert feature_type in AUTHORIZED_FEATURE_TYPES, "feature type not known"
+
+    assert all(
+        key in AUTHORIZED_FEATURE_INFOS[feature_type] for key in feature_details
+    ), "Unexpected key(s) in feature_identifier"
+
+    return feature_type, feature_details
+
+
 class Sample(BaseModel):
     """Represents a single sample. It contains data and information related to a single observation or measurement within a dataset."""
 
@@ -195,6 +242,18 @@ class Sample(BaseModel):
             "active_zone": None,
             "active_time": None,
         }
+
+    def copy(self) -> Self:
+        """Create a deep copy of the sample.
+
+        Returns:
+            A new `Sample` instance with all internal data (scalars, time series, fields, meshes, etc.)
+            deeply copied to ensure full isolation from the original.
+
+        Note:
+            This operation may be memory-intensive for large samples.
+        """
+        return copy.deepcopy(self)
 
     # -------------------------------------------------------------------------#
     def set_default_base(self, base_name: str, time: float = None) -> None:
@@ -1389,11 +1448,25 @@ class Sample(BaseModel):
                 f"there is no base/zone <{base_name}/{zone_name}>, you should first create one with `Sample.init_zone({zone_name=},{base_name=})`"
             )
 
+        # Check if GridCoordinates_t node exists
+        gc_nodes = [
[…]
