+++ src/plaid/containers/sample.py
@@ -37,6 +37,7 @@ from plaid.constants import (
     AUTHORIZED_FIELD_LOCATIONS,
     CGNS_ELEMENT_NAMES,
 )
+from plaid.containers.utils import get_feature_type_and_details_from_identifier
 from plaid.types import (
     CGNSNode,
     CGNSTree,
@@ -134,48 +135,6 @@ def read_index_range(pyTree: list, dim: list[int]):
             res.extend(np.arange(begin, end + 1).ravel())
 
     return np.array(res, dtype=int).ravel()
-
-
-def _get_feature_type_and_details_from_identifier(
-    feature_identifier: dict[str, Union[str, float]],
-) -> tuple[str, dict[str, Union[str, float]]]:
-    """Extract and validate the feature type and its associated metadata from a feature identifier.
-
-    This utility function ensures that the `feature_identifier` dictionary contains a valid
-    "type" key (e.g., "scalar", "time_series", "field", "node") and returns the type along
-    with the remaining identifier keys, which are specific to the feature type.
-
-    Args:
-        feature_identifier (dict): A dictionary with a "type" key, and
-            other keys (some optional) depending on the feature type. For example:
-                - {"type": "scalar", "name": "Mach"}
-                - {"type": "time_series", "name": "AOA"}
-                - {"type": "field", "name": "pressure"}
-
-    Returns:
-        tuple[str, dict]: A tuple `(feature_type, feature_details)` where:
-            - `feature_type` is the value of the "type" key (e.g., "scalar").
-            - `feature_details` is a dictionary of the remaining keys.
-
-    Raises:
-        AssertionError:
-            - If "type" is missing.
-            - If the type is not in `AUTHORIZED_FEATURE_TYPES`.
-            - If any unexpected keys are present for the given type.
-    """
-    assert "type" in feature_identifier, (
-        "feature type not specified in feature_identifier"
-    )
-    feature_type = feature_identifier["type"]
-    feature_details = {k: v for k, v in feature_identifier.items() if k != "type"}
-
-    assert feature_type in AUTHORIZED_FEATURE_TYPES, "feature type not known"
-
-    assert all(
-        key in AUTHORIZED_FEATURE_INFOS[feature_type] for key in feature_details
-    ), "Unexpected key(s) in feature_identifier"
-
-    return feature_type, feature_details
 
 
 class Sample(BaseModel):
@@ -1859,7 +1818,7 @@ class Sample(BaseModel):
         Returns:
             FeatureType: The corresponding feature instance retrieved via the appropriate accessor.
         """
-        feature_type, feature_details = _get_feature_type_and_details_from_identifier(
+        feature_type, feature_details = get_feature_type_and_details_from_identifier(
             feature_identifier
         )
 
@@ -1892,7 +1851,7 @@ class Sample(BaseModel):
         Raises:
             AssertionError: If types are inconsistent or identifiers contain unexpected keys.
         """
-        feature_type, feature_details = _get_feature_type_and_details_from_identifier(
+        feature_type, feature_details = get_feature_type_and_details_from_identifier(
             feature_identifier
         )
 
