### 2025-07-19 dbb9449
fix(sample) correct coverage
