### 2025-07-20 01c93ca
feat(dataset) implement get_tabular_from_identifier
