+++ examples/pipelines/config.yml
@@ -4,19 +4,6 @@ global:
   train_split_name: "train"
   test_split_name: "test"
 
-test_0:
-  features:
-    - type: scalar
-      name: angle_in
-    - type: scalar
-      name: mach_out
-
-test_1:
-  features:
-    - type: field
-      name: mach
-      base_name: Base_2_2
-
 input_scalar_scaler:
   scaler_type: "MinMaxScaler"
   feature_names:

[…]
