+++ examples/pipelines/ml_pipeline_nodes_2.py
@@ -67,267 +67,28 @@ class PlaidSklearnBlockWrapper(BaseEstimator, MetaEstimatorMixin):
         else:
             self.out_features_identifiers = params['in_features_identifiers']
 
-
-    def _extract_X_y_from_plaid(self, dataset_in: Dataset, dataset_out: Dataset = None):
-
-        X = dataset_in.get_tabular_from_identifier(self.in_features_identifiers)
-
-        if dataset_out is None:
-            return X, None
-        else:
-            y = dataset_out.get_tabular_from_identifier(self.out_features_identifiers)
-            return X, y
-
+        # self.in_dim_features = None
+        # self.out_dim_features = None
 
     def fit(self, dataset_in: Dataset, dataset_out: Dataset = None):
 
-        X, y = self._extract_X_y_from_plaid(dataset_in, dataset_out)
+        X = dataset_in.get_tabular_from_identifier(self.in_features_identifiers)
+        assert X.shape[1] == len(self.in_features_identifiers), "number of features not consistent between generated tabular and in_features_identifiers"
+        # self.in_dim_features = X.shape[2]
+        X = np.squeeze(X)
+
+        if dataset_out is None:
+            y = None
+        else:
+            y = dataset_out.get_tabular_from_identifier(self.out_features_identifiers)
+            assert y.shape[1] == len(self.out_features_identifiers), "number of features not consistent between generated tabular and out_features_identifiers"
+            # self.out_dim_features = y.shape[2]
+            y = np.squeeze(y)
 
         self.sklearn_block.fit(X, y)
 
         self._is_fitted = True
         return self
-
-    # def _determine_input_output_names(self, dataset: Dataset):
-    #     """Determine the input/output names based on the in_features and out_features."""
-    #     self.in_features = (
-    #         self.in_features
-    #         if isinstance(self.in_features, list)
-    #         else [self.in_features]
-    #     )
-    #     self.out_features = (
-    #         self.out_features
-    #         if isinstance(self.out_features, list)
-    #         else [self.out_features]
-    #     )
-    #     self._determine_scalar_names(dataset)
-    #     self._determine_time_series_names(dataset)
-    #     self._determine_field_names(dataset)
-
-    # def _determine_scalar_names(self, dataset: Dataset):
-    #     """Determine the input/output scalar names based on the in_features and out_features."""
-    #     # Input scalars
-    #     if ("all" in self.in_features) or ("scalar::all" in self.in_features):
-    #         self.input_scalars = dataset.get_scalar_names()
-    #     else:
-    #         self.input_scalars = [
-    #             s[8:] for s in self.in_features if s[:8] == "scalar::"
-    #         ]
-
-    #     # Output scalars
-    #     if ("all" in self.out_features) or ("scalar::all" in self.out_features):
-    #         assert "same" not in self.out_features
-    #         assert "scalar::same" not in self.out_features
-    #         self.output_scalars = dataset.get_scalar_names()
-    #     elif ("same" in self.out_features) or ("scalar::same" in self.out_features):
-    #         self.output_scalars = self.input_scalars
-    #     else:
-    #         self.output_scalars = [
-    #             s[8:] for s in self.out_features if s[:8] == "scalar::"
-    #         ]
-
-    # def _determine_time_series_names(self, dataset: Dataset):
-    #     """Determine the input/output time_series names based on the in_features and out_features."""
-    #     # Input time_series
-    #     if ("all" in self.in_features) or ("time_series::all" in self.in_features):
[…]
