### 2025-07-20 5867ad0
feat(example/pipelines) working generic example in optuna_pipeline2.py
