### 2025-07-21 fd2d27f
feat(dataset/sample) implement merge features functionalities
