+++ tests/containers/test_dataset.py
@@ -824,6 +824,24 @@ class Test_Dataset:
         with pytest.raises(ValueError):
             dataset_with_samples.merge_dataset(3)
 
+    def test_merge_features(self, dataset_with_samples, other_dataset_with_samples):
+        feat_id = dataset_with_samples.get_all_features_identifiers()
+        feat_id = [
+            fid for fid in feat_id if fid["type"] not in ["scalar", "time_series"]
+        ]
+        dataset_1 = dataset_with_samples.from_features_identifier(feat_id)
+        feat_id = other_dataset_with_samples.get_all_features_identifiers()
+        feat_id = [fid for fid in feat_id if fid["type"] not in ["field", "node"]]
+        dataset_2 = other_dataset_with_samples.from_features_identifier(feat_id)
+        dataset_merge_1 = dataset_1.merge_features(dataset_2, in_place=False)
+        dataset_merge_2 = dataset_2.merge_features(dataset_1, in_place=False)
+        assert (
+            dataset_merge_1[0].get_all_features_identifiers()
+            == dataset_merge_2[0].get_all_features_identifiers()
+        )
+        dataset_2.merge_features(dataset_1, in_place=True)
+        dataset_1.merge_features(dataset_2, in_place=True)
+
     # -------------------------------------------------------------------------#
 
     def test_from_list_of_samples(self, samples):
