### 2025-07-21 6dd0de9
improve ml_pipeline_nodes_2.py
