+++ examples/pipelines/ml_pipeline_nodes_2.py
@@ -5,6 +5,12 @@
 #
 #
 
+# TODO:
+# - better handling of params (in the GPRegressor here, params of the GP are not reachable with pipeline.get_params())
+# - save state / load state -> normalement ici, tout hérite de BaseEstimator, donc facile ?
+# - move this in src and provide tests, examples and move the notebook to the doc
+
+
 import numpy as np
 from sklearn.base import (
     BaseEstimator,
@@ -19,14 +25,19 @@ from sklearn.base import (
     TransformerMixin,
 )
 from sklearn.utils.metaestimators import _BaseComposition
+from sklearn.utils._repr_html.estimator import _VisualBlock
+from sklearn.utils._tags import get_tags
 from sklearn.compose import ColumnTransformer
 from sklearn.decomposition import PCA
 from plaid.containers.dataset import Dataset
 from plaid.containers.utils import check_features_type_homogeneity, merge_dataset_by_features
 import copy
 from typing import Union, Callable, Any
+from sklearn.base import clone
+from sklearn.utils.validation import check_is_fitted
 
 from sklearn.preprocessing import StandardScaler, MinMaxScaler
+
 
 SklearnBlock = Union[
     BaseEstimator,
@@ -51,50 +62,22 @@ available_scalers = {
 class PlaidSklearnBlockWrapper(BaseEstimator, MetaEstimatorMixin):
     def __init__(
         self,
-        sklearn_block: SklearnBlock,
-        params:dict
+        sklearn_block: SklearnBlock = None,
+        sklearn_block_constructor: Callable = None,
+        params: dict = None,
     ):
-        # TODO: check https://scikit-learn.org/stable/developers/develop.html#instantiation
+        if sklearn_block is None and sklearn_block_constructor is None:
+            raise ValueError("You must provide either 'sklearn_block' or 'sklearn_block_constructor'.")
+        if sklearn_block is not None and sklearn_block_constructor is not None:
+            raise ValueError("Only one of 'sklearn_block' or 'sklearn_block_constructor' should be provided.")
+        if sklearn_block is not None:
+            assert isinstance(sklearn_block, SklearnBlock), "sklearn_block must by a scikit_learn block"
+        if sklearn_block_constructor is not None:
+            assert isinstance(sklearn_block_constructor, Callable), "sklearn_block_constructor must by a Callable, returning a scikit_learn block"
+
         self.sklearn_block = sklearn_block
-        self.params = params
-
-        self.in_features_identifiers = params['in_features_identifiers']
-        check_features_type_homogeneity(self.in_features_identifiers)
-
-        if "out_features_identifiers" in params:
-            self.out_features_identifiers = params['out_features_identifiers']
-            check_features_type_homogeneity(self.out_features_identifiers)
-        else:
-            self.out_features_identifiers = params['in_features_identifiers']
-
-    def __sklearn_is_fitted__(self):
-        """Check if the wrapped scikit-learn model is fitted.
-
-        Returns:
-            bool: True if the model is fitted, False otherwise.
-        """
-        return hasattr(self, "_is_fitted") and self._is_fitted
-
-    def __repr__(self):
-        """String representation of the wrapper, showing the underlying sklearn block."""
-        return f"{self.__class__.__name__}({self.sklearn_block.__repr__()})"
-
-    def __str__(self):
-        """String representation of the wrapper, showing the underlying sklearn block."""
-        return f"{self.__class__.__name__}({self.sklearn_block.__str__()})"
-
-
-class PlaidSklearnBlockConstructorWrapper(BaseEstimator, MetaEstimatorMixin):
-    def __init__(
-        self,
-        sklearn_block_constructor: Callable,
-        params:dict
-    ):
[…]
