+++ examples/pipelines/ml_pipeline_nodes_3.py
@@ -6,6 +6,10 @@
 #
 
 # TODO:
+# - recoder score avec la logique
+#       def score(self, X, y):
+#       y_pred = self.predict(X)
+#       return -mean_squared_error(y, y_pred)
 # - better handling of params (in the GPRegressor here, params of the GP are not reachable with pipeline.get_params())
 # - save state / load state -> normalement ici, tout hérite de BaseEstimator, donc facile ?
 # - move this in src and provide tests, examples and move the notebook to the doc
@@ -62,22 +66,21 @@ available_scalers = {
 class PlaidSklearnBlockWrapper(BaseEstimator, MetaEstimatorMixin):
     def __init__(
         self,
-        sklearn_block: SklearnBlock = None,
-        sklearn_block_constructor: Callable = None,
-        params: dict = None,
+        sklearn_block: SklearnBlock,
+        params:dict
     ):
-        if sklearn_block is None and sklearn_block_constructor is None:
-            raise ValueError("You must provide either 'sklearn_block' or 'sklearn_block_constructor'.")
-        if sklearn_block is not None and sklearn_block_constructor is not None:
-            raise ValueError("Only one of 'sklearn_block' or 'sklearn_block_constructor' should be provided.")
-        if sklearn_block is not None:
-            assert isinstance(sklearn_block, SklearnBlock), "sklearn_block must by a scikit_learn block"
-        if sklearn_block_constructor is not None:
-            assert isinstance(sklearn_block_constructor, Callable), "sklearn_block_constructor must by a Callable, returning a scikit_learn block"
+        assert isinstance(sklearn_block, SklearnBlock), "sklearn_block must be a scikit_learn block"
+        assert isinstance(params, dict), "params must be a dict containing the scikit_learn block parameters to set"
+        assert "in_features_identifiers" in params, "Must provide in_features_identifiers in sklearn_params of the node"
 
         self.sklearn_block = sklearn_block
-        self.sklearn_block_constructor = sklearn_block_constructor
-        self.params = params or {}
+        self.params = params
+
+        # print("kwargs =", kwargs)
+        # 1./0.
+        # print(self.in_features_identifiers)
+        # print(self.in_features_identifiersAA)
+        # AttributeError
 
         self.in_features_identifiers = params['in_features_identifiers']
         check_features_type_homogeneity(self.in_features_identifiers)
@@ -88,10 +91,52 @@ class PlaidSklearnBlockWrapper(BaseEstimator, MetaEstimatorMixin):
         else:
             self.out_features_identifiers = params['in_features_identifiers']
 
-    def fit_sklearn_block(self, X, y):
-        if self.sklearn_block is None:
-            self.sklearn_block = self.sklearn_block_constructor(X, y, self.params)
-        self.sklearn_block_ = clone(self.sklearn_block).fit(X, y)
+        if 'sklearn_params' not in params:
+            params['sklearn_params'] = {}
+
+        self.sklearn_block.set_params(**params['sklearn_params'])
+
+    # def _get_param_names(self):
+    #     return self.sklearn_block._get_param_names()
+
+    # def get_params(self, deep=True):
+    #     return self.sklearn_block.get_params()
+
+    # def set_params(self, **params):
+    #     self.params.update(params)
+    #     self.sklearn_block.set_params(**params)
+    #     return self
+
+    # def get_params(self, deep=True):
+    #     out = {
+    #         "sklearn_block": self.sklearn_block,
+    #     }
+    #     out.update(self.params['sklearn_params'])
+    #     return out
+
[…]
