+++ examples/pipelines/tests_5.py
@@ -1,0 +1,118 @@
+# from ml_pipeline_nodes_5 import PlaidTransformedTargetRegressor, WrappedPlaidSklearnTransformer, WrappedPlaidSklearnRegressor, PlaidColumnTransformer, PlaidSklearnBlockWrapper_
+from ml_pipeline_nodes_5 import WrappedPlaidSklearnTransformer, PlaidColumnTransformer, WrappedPlaidSklearnRegressor, PlaidTransformedTargetRegressor
+from sklearn.preprocessing import StandardScaler, MinMaxScaler
+from sklearn.decomposition import PCA
+import yaml
+from sklearn.base import clone
+from sklearn.pipeline import Pipeline
+
+with open("config_5.yml") as f:
+    config = yaml.safe_load(f)
+
+
+
+conf1 = config['pca_nodes']
+
+
+
+wrapped_transf = WrappedPlaidSklearnTransformer(PCA(**config['pca_nodes']['sklearn_block_params']), **config['pca_nodes']['plaid_params'])
+print("wrapped_transf.get_params()['sklearn_block__n_components'] (3) =", wrapped_transf.get_params()['sklearn_block__n_components'])
+assert wrapped_transf.get_params()['sklearn_block__n_components'] == config['pca_nodes']['sklearn_block_params']['n_components']
+
+wrapped_transf_clone = clone(wrapped_transf)
+
+wrapped_transf_clone.set_params(**{'sklearn_block__n_components':5})
+print("wrapped_transf_clone.get_params()['sklearn_block__n_components'] (5) =", wrapped_transf_clone.get_params()['sklearn_block__n_components'])
+# print("wrapped_transf_clone.sklearn_block.get_params() =", wrapped_transf_clone.sklearn_block.get_params())
+
+print("==")
+
+preprocessor = PlaidColumnTransformer([
+    ('input_scalar_scaler', WrappedPlaidSklearnTransformer(MinMaxScaler(), **config['input_scalar_scaler']['plaid_params'])),
+    ('pca_nodes', WrappedPlaidSklearnTransformer(PCA(**config['pca_nodes']['sklearn_block_params']), **config['pca_nodes']['plaid_params'])),
+], remainder_feature_id = config['pca_mach']['plaid_params']['in_features_identifiers'])
+print("preprocessor.get_params()['pca_nodes__sklearn_block__n_components'] (3) =", preprocessor.get_params()['pca_nodes__sklearn_block__n_components'])
+assert preprocessor.get_params()['pca_nodes__sklearn_block__n_components'] == config['pca_nodes']['sklearn_block_params']['n_components']
+
+preprocessor_clone = clone(preprocessor)
+preprocessor_clone.set_params(**{'pca_nodes__sklearn_block__n_components':5})
+print("preprocessor_clone.get_params()['pca_nodes__sklearn_block__n_components'] (5) =", preprocessor_clone.get_params()['pca_nodes__sklearn_block__n_components'])
+# print("wrapped_transf_clone.get_params() =", wrapped_transf_clone.get_params())
+
+
+print("==")
+
+from sklearn.gaussian_process import GaussianProcessRegressor
+from sklearn.gaussian_process.kernels import Matern, WhiteKernel, ConstantKernel
+from sklearn.multioutput import MultiOutputRegressor
+
+
+available_kernel_classes = {
+    "Matern":Matern
+}
+
+
+kernel = ConstantKernel() * Matern(length_scale_bounds=(1e-8, 1e8), nu = 2.5) + WhiteKernel(noise_level_bounds=(1e-8, 1e8))
+
+gpr = GaussianProcessRegressor(
+    kernel=kernel,
+    optimizer='fmin_l_bfgs_b',
[…]
