+++ examples/pipelines/ml_pipeline_nodes_5.py
@@ -1,0 +1,572 @@
+# -*- coding: utf-8 -*-
+#
+# This file is subject to the terms and conditions defined in
+# file 'LICENSE.txt', which is part of this source code package.
+#
+#
+
+# TODO:
+# - better handling of params (in the GPRegressor here, params of the GP are not reachable with pipeline.get_params())
+# - test restrict_to_features = True dans transformer (limit memory consumption)
+# - save state / load state -> normalement ici, tout hérite de BaseEstimator, donc facile ?
+# - move this in src and provide tests, examples and move the notebook to the doc
+
+
+import numpy as np
+from sklearn.base import (
+    BaseEstimator,
+    BiclusterMixin,
+    ClassifierMixin,
+    ClusterMixin,
+    DensityMixin,
+    MetaEstimatorMixin,
+    MultiOutputMixin,
+    OutlierMixin,
+    RegressorMixin,
+    TransformerMixin,
+)
+from sklearn.utils.metaestimators import _BaseComposition
+from sklearn.utils._repr_html.estimator import _VisualBlock
+from sklearn.utils._tags import get_tags
+from sklearn.compose import ColumnTransformer
+from sklearn.decomposition import PCA
+from plaid.containers.dataset import Dataset
+from plaid.containers.utils import check_features_type_homogeneity
+import copy
+from typing import Union, Callable, Any
+from sklearn.base import clone
+from sklearn.utils.validation import check_is_fitted
+
+from sklearn.preprocessing import StandardScaler, MinMaxScaler
+
+
+SklearnBlock = Union[
+    BaseEstimator,
+    TransformerMixin,
+    RegressorMixin,
+    ClassifierMixin,
+    ClusterMixin,
+    BiclusterMixin,
+    DensityMixin,
+    OutlierMixin,
+    MultiOutputMixin,
+]
+"""Union type for all scikit-learn blocks that can be used in a Pipeline."""
+
+
+available_scalers = {
+    "StandardScaler": StandardScaler,
+    "MinMaxScaler": MinMaxScaler,
+}
+
+
+
+# class PlaidSklearnBlockWrapper_(BaseEstimator):
+#     # split in 2 classes, with a block and a block_constructor
+#     def __init__(
+#         self,
+#         sklearn_block: SklearnBlock = None,
+#         **kwargs: dict = None,
+#     ):
+#         self.sklearn_block = sklearn_block
+#         self.params = params
+
+#     def _prepare_fit(self):
+
+#         check_features_type_homogeneity(self.params['in_features_identifiers'])
+#         self.in_features_identifiers_ = copy.deepcopy(self.params['in_features_identifiers'])
+
+#         if "out_features_identifiers" in self.params:
+#             check_features_type_homogeneity(self.params['out_features_identifiers'])
+#             self.out_features_identifiers_ = self.params['out_features_identifiers']
+#         else:
+#             self.out_features_identifiers_ = copy.deepcopy(self.params['in_features_identifiers'])
+
+#     def _fit_sklearn_block(self, X, y):
+
+#         if self.sklearn_block is None:
+#             self.sklearn_block = self.sklearn_block_constructor(X, y, self.params)
+#         self.sklearn_block_ = clone(self.sklearn_block).fit(X, y)
+
+#     # def get_params(self, deep=True):
+#     #     out = {}
+#     #     if deep and isinstance(self.sklearn_block, SklearnBlock):
+#     #         for key, value in self.sklearn_block.get_params(deep=True).items():
+#     #             out[key] = value
+#     #     return out
+
+#     # def set_params(self, **params):
+#     #     print(self.sklearn_block)
+#     #     self.sklearn_block.set_params(params)
[…]
