+++ examples/pipelines/config_5.yml
@@ -1,0 +1,98 @@
+global:
+  dataset_path: "PLAID-datasets/VKI-LS59"
+  save_path: "./artifacts"
+  train_split_name: "train"
+  test_split_name: "test"
+
+
+input_scalar_scaler:
+  sklearn_block: "MinMaxScaler"
+  plaid_params:
+    in_features_identifiers:
+      - type: scalar
+        name: angle_in
+      - type: scalar
+        name: mach_out
+    out_features_identifiers:
+      - type: scalar
+        name: angle_in
+      - type: scalar
+        name: mach_out
+
+output_scalar_scaler:
+  sklearn_block: "MinMaxScaler"
+  plaid_params:
+    in_features_identifiers:
+      - type: scalar
+        name: Q
+      - type: scalar
+        name: power
+      - type: scalar
+        name: Pr
+      - type: scalar
+        name: Tr
+      - type: scalar
+        name: eth_is
+      - type: scalar
+        name: angle_out
+    out_features_identifiers:
+      - type: scalar
+        name: Q
+      - type: scalar
+        name: power
+      - type: scalar
+        name: Pr
+      - type: scalar
+        name: Tr
+      - type: scalar
+        name: eth_is
+      - type: scalar
+        name: angle_out
+
+pca_nodes:
+  sklearn_block: "PCA"
+  sklearn_block_params:
+    n_components: 2
+  plaid_params:
+    in_features_identifiers:
+      - type: nodes
+        base_name: Base_2_2
+    out_features_identifiers:
+      - type: scalar
+        name: reduced_nodes_*
+
+pca_mach:
+  sklearn_block: "PCA"
+  sklearn_block_params:
+    n_components: 2
+  plaid_params:
+    in_features_identifiers:
+      - type: field
+        name: mach
+        base_name: Base_2_2
+    out_features_identifiers:
+      - type: scalar
+        name: reduced_mach_*
+
+regressor_mach:
+  sklearn_block: "MultiOutputRegressor(GaussianProcessRegressor(ConstantKernel()*Matern()+WhiteKernel()))"
+  sklearn_block_params:
+    options:
+      kernel: Matern
+      kernel_options:
+        nu: 2.5
+      optim: fmin_l_bfgs_b
+      num_restarts: 1
+      anisotropic: True
+      random_state: 42
+  plaid_params:
+    in_features_identifiers:
+      - type: scalar
+        name: angle_in
+      - type: scalar
+        name: mach_out
+      - type: scalar
+        name: reduced_nodes_*
+    out_features_identifiers:
+      - type: scalar
+        name: reduced_mach_*
