### 2025-07-23 f3268d1
extract dataset to lower memory consumption
