+++ examples/pipelines/config_5.yml
@@ -19,36 +19,6 @@ input_scalar_scaler:
       - type: scalar
         name: mach_out
 
-output_scalar_scaler:
-  sklearn_block: "MinMaxScaler"
-  plaid_params:
-    in_features_identifiers:
-      - type: scalar
-        name: Q
-      - type: scalar
-        name: power
-      - type: scalar
-        name: Pr
-      - type: scalar
-        name: Tr
-      - type: scalar
-        name: eth_is
-      - type: scalar
-        name: angle_out
-    out_features_identifiers:
-      - type: scalar
-        name: Q
-      - type: scalar
-        name: power
-      - type: scalar
-        name: Pr
-      - type: scalar
-        name: Tr
-      - type: scalar
-        name: eth_is
-      - type: scalar
-        name: angle_out
-
 pca_nodes:
   sklearn_block: "PCA"
   sklearn_block_params:
@@ -96,3 +66,35 @@ regressor_mach:
     out_features_identifiers:
       - type: scalar
         name: reduced_mach_*
+
+
+
+output_scalar_scaler:
+  sklearn_block: "MinMaxScaler"
+  plaid_params:
+    in_features_identifiers:
+      - type: scalar
+        name: Q
+      - type: scalar
+        name: power
+      - type: scalar
+        name: Pr
+      - type: scalar
+        name: Tr
+      - type: scalar
+        name: eth_is
+      - type: scalar
+        name: angle_out
+    out_features_identifiers:
+      - type: scalar
+        name: Q
+      - type: scalar
+        name: power
+      - type: scalar
+        name: Pr
+      - type: scalar
+        name: Tr
+      - type: scalar
+        name: eth_is
+      - type: scalar
+        name: angle_out
