### 2025-07-23 031e308
pipelines: add save and load example
