+++ examples/pipelines/ml_pipeline_nodes_5.py
@@ -6,7 +6,6 @@
 #
 
 # TODO:
-# - test restrict_to_features = True dans transformer (limit memory consumption)
 # - save state / load state -> normalement ici, tout hérite de BaseEstimator, donc facile ?
 # - move this in src and provide tests, examples and move the notebook to the doc
 
@@ -196,149 +195,6 @@ class WrappedPlaidSklearnRegressor(RegressorMixin, BaseEstimator):
     def inverse_transform(self, dataset):
         return dataset
 
-##########################################################################
-
-
-# class PlaidSklearnBlockWrapper(BaseEstimator):
-#     # split in 2 classes, with a block and a block_constructor
-#     def __init__(
-#         self,
-#         sklearn_block: SklearnBlock = None,
-#         sklearn_block_constructor: Callable = None,
-#         params: dict = None,
-#     ):
-#         self.sklearn_block = sklearn_block
-#         self.sklearn_block_constructor = sklearn_block_constructor
-#         self.params = params
-
-
-#     def _prepare_fit(self):
-
-#         if self.sklearn_block is None and self.sklearn_block_constructor is None:
-#             raise ValueError("You must provide either 'sklearn_block' or 'sklearn_block_constructor'.")
-#         if self.sklearn_block is not None and self.sklearn_block_constructor is not None:
-#             raise ValueError("Only one of 'sklearn_block' or 'sklearn_block_constructor' should be provided.")
-#         if self.sklearn_block is not None:
-#             assert isinstance(self.sklearn_block, SklearnBlock), "sklearn_block must by a scikit_learn block"
-#         if self.sklearn_block_constructor is not None:
-#             assert isinstance(self.sklearn_block_constructor, Callable), "sklearn_block_constructor must by a Callable, returning a scikit_learn block"
-
-#         self.in_features_identifiers = self.params['in_features_identifiers']
-#         check_features_type_homogeneity(self.in_features_identifiers)
-
-#         if "out_features_identifiers" in self.params:
-#             self.out_features_identifiers = self.params['out_features_identifiers']
-#             check_features_type_homogeneity(self.out_features_identifiers)
-#         else:
-#             self.out_features_identifiers = self.params['in_features_identifiers']
-
-#     def _fit_sklearn_block(self, X, y):
-
-#         if self.sklearn_block is None:
-#             self.sklearn_block = self.sklearn_block_constructor(X, y, self.params)
-#         self.sklearn_block_ = clone(self.sklearn_block).fit(X, y)
-
-
-# class WrappedPlaidSklearnTransformer(TransformerMixin, PlaidSklearnBlockWrapper):
-#     """Wrapper for scikit-learn Transformer blocks to operate on PLAID objects in a Pipeline.
-
-#     This class allows you to use any sklearn Transformer (e.g. PCA, StandardScaler) in a Pipeline where all steps
-#     accept and return PLAID Dataset objects. The transform and inverse_transform methods take a Dataset and return a new Dataset.
-#     """
-
-#     def fit(self, dataset: Dataset, y=None):
-
-#         self._prepare_fit()
-
-#         X = dataset.get_tabular_from_homogeneous_identifiers(self.in_features_identifiers)
-#         assert X.shape[1] == len(self.in_features_identifiers), "number of features not consistent between generated tabular and in_features_identifiers"
-#         X = np.squeeze(X)
-
-#         print(">>> =", self.sklearn_block.get_params())
-
-#         self._fit_sklearn_block(X, y)
-
-#         return self
-
-
-#     def transform(self, dataset: Dataset):
-#         """Transform the dataset using the wrapped sklearn transformer.
-
-#         Args:
[…]
