### 2025-07-23 086fe56
update temp example notebooks and start implementing pipelines in src and tests
