+++ examples/pipelines/ml_pipeline_nodes_2.py
@@ -276,56 +276,3 @@ class PlaidColumnTransformer(ColumnTransformer):
 
     def fit_transform(self, X, y=None):
         return self.fit(X, y).transform(X)
-
-
-from sklearn.gaussian_process import GaussianProcessRegressor
-class LazyGPR(BaseEstimator, RegressorMixin):
-    def __init__(self, kernel_factory=None, gpr_params=None):
-        """
-        kernel_factory: Callable[[int], sklearn.gaussian_process.kernels.Kernel]
-            Function taking n_features and returning a kernel instance.
-        gpr_params: dict
-            Parameters passed to GaussianProcessRegressor constructor.
-        """
-        self.kernel_factory = kernel_factory
-        self.gpr_params = gpr_params or {}
-        self.gpr_ = None
-
-    def fit(self, X, y):
-        if self.kernel_factory is None:
-            raise ValueError("kernel_factory must be provided")
-
-        n_features = X.shape[1]
-        kernel = self.kernel_factory(n_features)
-
-        self.gpr_ = GaussianProcessRegressor(kernel=kernel, **self.gpr_params)
-        self.gpr_.fit(X, y)
-        return self
-
-    def predict(self, X, return_std=False, return_cov=False):
-        return self.gpr_.predict(X, return_std=return_std, return_cov=return_cov)
-
-    def get_params(self, deep=True):
-        # Flatten gpr_params with prefix for better integration with nested wrappers
-        params = {"kernel_factory": self.kernel_factory}
-        for k, v in self.gpr_params.items():
-            params[f"gpr_params__{k}"] = v
-        return params
-
-    def set_params(self, **params):
-        gpr_params = {}
-        for key, value in params.items():
-            if key.startswith("gpr_params__"):
-                gpr_params[key[len("gpr_params__"):]] = value
-            elif key == "kernel_factory":
-                self.kernel_factory = value
-            else:
-                # Unknown param, ignore or raise error
-                pass
-        if gpr_params:
-            self.gpr_params.update(gpr_params)
-        return self
-
-    @property
-    def kernel_(self):
-        return self.gpr_.kernel_ if self.gpr_ is not None else None
