+++ src/plaid/pipelines/sklearn_block_wrappers.py
@@ -1,0 +1,265 @@
+"""Wrapped scikit-learn transformers and regressors for PLAID Dataset compatibility.
+
+Provides adapters to use scikit-learn estimators within the PLAID feature/block system:
+- WrappedPlaidSklearnTransformer: wraps a TransformerMixin
+- WrappedPlaidSklearnRegressor: wraps a RegressorMixin
+"""
+
+# -*- coding: utf-8 -*-
+#
+# This file is subject to the terms and conditions defined in
+# file 'LICENSE.txt', which is part of this source code package.
+#
+#
+
+import copy
+from typing import Union
+
+from sklearn.base import (
+    BaseEstimator,
+    BiclusterMixin,
+    ClassifierMixin,
+    ClusterMixin,
+    DensityMixin,
+    MultiOutputMixin,
+    OutlierMixin,
+    RegressorMixin,
+    TransformerMixin,
+    clone,
+)
+from sklearn.utils.validation import check_is_fitted
+
+from plaid.containers.dataset import Dataset
+from plaid.containers.utils import check_features_type_homogeneity
+
+SklearnBlock = Union[
+    BaseEstimator,
+    TransformerMixin,
+    RegressorMixin,
+    ClassifierMixin,
+    ClusterMixin,
+    BiclusterMixin,
+    DensityMixin,
+    OutlierMixin,
+    MultiOutputMixin,
+]
+
+
+def get_tabular_from_homogeneous_identifiers(
+    dataset: Dataset, features_identifiers: list[dict]
+):
+    """Returns a 2D array from a Dataset and a feature id.
+
+    The function calls `dataset.get_tabular_from_homogeneous_identifiers(...)`, then removes
+    either the second or third dimension if it has size 1, so that the output is 2D.
+
+    Args:
+        dataset: A Dataset object exposing `get_tabular_from_homogeneous_identifiers`.
+        features_identifiers: a list of input feature identifiers.
+
+    Returns:
+        A NumPy array of shape (n_samples, n_features),.
+
+    Raises:
+        AssertionError: If the number of features in the output does not match the identifiers.
+        ValueError: If both the second and third dimensions have size greater than 1.
+    """
+    X = dataset.get_tabular_from_homogeneous_identifiers(features_identifiers)
+    # X is of size (nb_sample, nb_features, dim_features), either nb_features or dim_features should be 1 to be compatible with scikit-learn blocks
+    if X.shape[1] == 1:
+        X = X[:, 0, :]
+    elif X.shape[2] == 1:
+        X = X[:, :, 0]
+    else:
+        raise ValueError(
+            "X (generate by dataset.get_tabular_from_homogeneous_identifiers) is expected to have its second or third dimension equal to 1"
+        )
+
+    return X
+
+
+class WrappedPlaidSklearnTransformer(TransformerMixin, BaseEstimator):
+    """Adapter for using a scikit-learn transformer on PLAID Datasets.
+
+    Transforms tabular data extracted from homogeneous feature identifiers,
+    and returns results as a `Dataset`. Supports forward and inverse transforms.
+
+    Args:
+        sklearn_block: A scikit-learn Transformer implementing fit/transform APIs.
+        in_features_identifiers: List of feature identifiers to extract input data from.
+        out_features_identifiers: List of feature identifiers used for outputs. If None,
+            defaults to `in_features_identifiers`.
+    """
+
+    def __init__(
+        self,
+        sklearn_block: SklearnBlock = None,
+        in_features_identifiers: list[dict] = None,
+        out_features_identifiers: list[dict] = None,
+    ):
+        self.sklearn_block = sklearn_block
+        self.in_features_identifiers = in_features_identifiers
[…]
