+++ tests/pipelines/test_plaid_blocks.py
@@ -1,0 +1,14 @@
+from plaid.pipelines.plaid_blocks import (
+    PlaidColumnTransformer,
+    PlaidTransformedTargetRegressor,
+)
+
+
+class Test_PlaidTransformedTargetRegressor:
+    def test___init__(self):
+        PlaidTransformedTargetRegressor()
+
+
+class Test_PlaidColumnTransformer:
+    def test___init__(self):
+        PlaidColumnTransformer()
