+++ tests/containers/test_sample.py
@@ -1209,19 +1209,19 @@ class Test_Sample:
     def test_get_all_features_identifiers(
         self, sample_with_tree_and_scalar_and_time_series
     ):
-        feat_id = (
+        feat_ids = (
             sample_with_tree_and_scalar_and_time_series.get_all_features_identifiers()
         )
-        assert len(feat_id) == 10
-        assert {"type": "scalar", "name": "r"} in feat_id
-        assert {"type": "scalar", "name": "test_scalar_1"} in feat_id
-        assert {"type": "time_series", "name": "test_time_series_1"} in feat_id
+        assert len(feat_ids) == 10
+        assert {"type": "scalar", "name": "r"} in feat_ids
+        assert {"type": "scalar", "name": "test_scalar_1"} in feat_ids
+        assert {"type": "time_series", "name": "test_time_series_1"} in feat_ids
         assert {
             "type": "nodes",
             "base_name": "Base_2_2",
             "zone_name": "Zone",
             "time": 0.0,
-        } in feat_id
+        } in feat_ids
         assert {
             "type": "field",
             "name": "big_node_field",
@@ -1229,7 +1229,7 @@ class Test_Sample:
             "zone_name": "Zone",
             "location": "Vertex",
             "time": 0.0,
-        } in feat_id
+        } in feat_ids
         assert {
             "type": "field",
             "name": "test_node_field_1",
@@ -1237,7 +1237,7 @@ class Test_Sample:
             "zone_name": "Zone",
             "location": "Vertex",
             "time": 0.0,
-        } in feat_id
+        } in feat_ids
         assert {
             "type": "field",
             "name": "OriginalIds",
@@ -1245,7 +1245,7 @@ class Test_Sample:
             "zone_name": "Zone",
             "location": "Vertex",
             "time": 0.0,
-        } in feat_id
+        } in feat_ids
         assert {
             "type": "field",
             "name": "OriginalIds",
@@ -1253,7 +1253,7 @@ class Test_Sample:
             "zone_name": "Zone",
             "location": "FaceCenter",
             "time": 0.0,
-        } in feat_id
+        } in feat_ids
         assert {
             "type": "field",
             "name": "test_elem_field_1",
@@ -1261,7 +1261,7 @@ class Test_Sample:
             "zone_name": "Zone",
             "location": "CellCenter",
             "time": 0.0,
-        } in feat_id
+        } in feat_ids
         assert {
             "type": "field",
             "name": "OriginalIds",
@@ -1269,7 +1269,45 @@ class Test_Sample:
             "zone_name": "Zone",
             "location": "CellCenter",
             "time": 0.0,
-        } in feat_id
+        } in feat_ids
+
+    def test_get_all_features_identifiers_by_type(
+        self, sample_with_tree_and_scalar_and_time_series
+    ):
+        feat_ids = sample_with_tree_and_scalar_and_time_series.get_all_features_identifiers_by_type(
+            "scalar"
+        )
+        assert len(feat_ids) == 2
+        assert {"type": "scalar", "name": "r"} in feat_ids
+        assert {"type": "scalar", "name": "test_scalar_1"} in feat_ids
+
+        feat_ids = sample_with_tree_and_scalar_and_time_series.get_all_features_identifiers_by_type(
+            "time_series"
+        )
+        assert {"type": "time_series", "name": "test_time_series_1"} in feat_ids
+
+        feat_ids = sample_with_tree_and_scalar_and_time_series.get_all_features_identifiers_by_type(
+            "nodes"
+        )
+        assert {
+            "type": "nodes",
+            "base_name": "Base_2_2",
+            "zone_name": "Zone",
+            "time": 0.0,
+        } in feat_ids
+
[…]
