### 2025-07-24 18aadb2
🎉 feat(cgns_helper): add 'summarize_cgns_tree' and usage example
