### 2025-07-24 241c7cf
📄 Fix error in function docstring
