+++ pyproject.toml
@@ -40,6 +40,7 @@ dependencies = [
     "scikit-learn",
     "datasets >=3.0",
     "pydantic",
+    "optuna",
 ]
 
 dynamic = ["version"]
