### 2025-07-24 d2c06aa
rework: finish unit tests
