### 2025-07-24 ff70047
♻️ remove unnecessary check in 'summarize_cgns_tree'
