### 2025-07-24 c7efa23
add notebooks in docs, add example and clean
