+++ examples/pipelines/ml_pipeline_nodes_5.py
@@ -1,301 +1,0 @@
-# -*- coding: utf-8 -*-
-#
-# This file is subject to the terms and conditions defined in
-# file 'LICENSE.txt', which is part of this source code package.
-#
-#
-
-# TODO:
-# - move this in src and provide tests, examples and move the notebook to the doc
-
-
-import numpy as np
-from sklearn.base import (
-    BaseEstimator,
-    BiclusterMixin,
-    ClassifierMixin,
-    ClusterMixin,
-    DensityMixin,
-    MetaEstimatorMixin,
-    MultiOutputMixin,
-    OutlierMixin,
-    RegressorMixin,
-    TransformerMixin,
-)
-from sklearn.utils.metaestimators import _BaseComposition
-from sklearn.utils._repr_html.estimator import _VisualBlock
-from sklearn.utils._tags import get_tags
-from sklearn.compose import ColumnTransformer
-from sklearn.decomposition import PCA
-from plaid.containers.dataset import Dataset
-from plaid.containers.utils import check_features_type_homogeneity
-import copy
-from typing import Union, Callable, Any
-from sklearn.base import clone
-from sklearn.utils.validation import check_is_fitted
-
-from sklearn.preprocessing import StandardScaler, MinMaxScaler
-
-
-SklearnBlock = Union[
-    BaseEstimator,
-    TransformerMixin,
-    RegressorMixin,
-    ClassifierMixin,
-    ClusterMixin,
-    BiclusterMixin,
-    DensityMixin,
-    OutlierMixin,
-    MultiOutputMixin,
-]
-"""Union type for all scikit-learn blocks that can be used in a Pipeline."""
-
-
-available_scalers = {
-    "StandardScaler": StandardScaler,
-    "MinMaxScaler": MinMaxScaler,
-}
-
-
-class WrappedPlaidSklearnTransformer(TransformerMixin, BaseEstimator):
-    """Wrapper for scikit-learn Transformer blocks to operate on PLAID objects in a Pipeline.
-
-    This class allows you to use any sklearn Transformer (e.g. PCA, StandardScaler) in a Pipeline where all steps
-    accept and return PLAID Dataset objects. The transform and inverse_transform methods take a Dataset and return a new Dataset.
-    """
-
-    def __init__(
-        self,
-        sklearn_block: SklearnBlock = None,
-        in_features_identifiers: dict = None,
-        out_features_identifiers: dict = None,
-    ):
-        self.sklearn_block = sklearn_block
-        self.in_features_identifiers  = in_features_identifiers
-        self.out_features_identifiers = out_features_identifiers
-
-    def fit(self, dataset: Dataset, y=None):
-
-        self.in_features_identifiers_  = copy.deepcopy(self.in_features_identifiers)
-        self.out_features_identifiers_ = copy.deepcopy(self.out_features_identifiers)
-
-        check_features_type_homogeneity(self.in_features_identifiers_)
-        if self.out_features_identifiers_:
-            check_features_type_homogeneity(self.out_features_identifiers_)
-        else:
-            self.out_features_identifiers_ = self.in_features_identifiers_
-
-        X = dataset.get_tabular_from_homogeneous_identifiers(self.in_features_identifiers_)
-        assert X.shape[1] == len(self.in_features_identifiers_), "number of features not consistent between generated tabular and in_features_identifiers"
-        X = np.squeeze(X)
-
-        self.sklearn_block_ = clone(self.sklearn_block).fit(X, y)
-
-        return self
-
-
-    def transform(self, dataset: Dataset):
-        """Transform the dataset using the wrapped sklearn transformer.
-
-        Args:
-            dataset (Dataset): The dataset to transform.
-
-        Returns:
-            Dataset: The transformed PLAID dataset.
-        """
-        check_is_fitted(self, "sklearn_block_")
-
[…]
