### 2025-07-25 34ee6bf
📄 add missing element in 'summarize_cgns_tree' docstring
