### 2025-07-25 bd22d93
🎉 add 'summarize_cgns_tree' (#114)
* 🎉 feat(cgns_helper): add 'summarize_cgns_tree' and usage example

* ♻️ tests(cgns_helper): add tests to 'summarize_cgns_tree'

* 📄 Fix error in function docstring

* 📄 add documentation, fix wrong element count, add node count for 'summarize_cgns_tree'

* ♻️ remove unnecessary check in 'summarize_cgns_tree'

* 📄 add missing element in 'summarize_cgns_tree' docstring
