### 2025-07-25 d96117c
remove pipelines to examples script
