### 2025-07-25 235a793
pipeline notebook update
