### 2025-07-25 76cfd71
update pipeline docs and example
