+++ examples/pipelines/pipeline.py
@@ -1,0 +1,289 @@
+#!/usr/bin/env python
+# coding: utf-8
+
+# # Pipeline Examples
+#
+# This notebook presents how the provided pipeline blocks can be used in a PCA-GP algorithm. Pipeline block directly link PLAID datasets, and hyperpamater tuning is available using scikit-learn's GridSearchCV or Optuna.
+#
+# We start by a few imports:
+
+# In[ ]:
+
+
+import os
+import yaml
+from pathlib import Path
+import optuna
+import numpy as np
+
+from sklearn.base import clone
+from sklearn.pipeline import Pipeline
+
+from sklearn.decomposition import PCA
+from sklearn.preprocessing import MinMaxScaler
+from sklearn.gaussian_process import GaussianProcessRegressor
+from sklearn.gaussian_process.kernels import Matern
+from sklearn.multioutput import MultiOutputRegressor
+
+from sklearn.model_selection import KFold, GridSearchCV
+
+import warnings
+from tqdm import TqdmWarning
+warnings.filterwarnings('ignore', module='sklearn')
+warnings.filterwarnings('ignore', category=TqdmWarning)
+
+from datasets import load_dataset, get_dataset_infos
+from plaid.bridges.huggingface_bridge import huggingface_dataset_to_plaid, huggingface_description_to_problem_definition
+from plaid.pipelines.sklearn_block_wrappers import WrappedPlaidSklearnTransformer, WrappedPlaidSklearnRegressor
+from plaid.pipelines.plaid_blocks import PlaidTransformedTargetRegressor, PlaidColumnTransformer
+
+nb_cpus = os.cpu_count()
+n_processes = max(1, int(nb_cpus / 4))
+
+
+# We load the `VKI-LS59` dataset from Hugging Face, and restrict ourselves to the first 24 samples of the training set.
+
+# In[ ]:
+
+
+hf_dataset = load_dataset("PLAID-datasets/VKI-LS59", split="all_samples[:24]")
+dataset_train, _ = huggingface_dataset_to_plaid(hf_dataset, processes_number = n_processes, verbose = False)
+
+del hf_dataset
+
+
+# We print the summary of dataset_train:
+
+# In[ ]:
+
+
+print(dataset_train)
+
+
+# There are 24 samples, with 8 scalars and 8 fields, which is consistent with the VKI-LS59` dataset. To contain memory consumption, we restrict the dataset to the features required for this example. Far praticity, in_features_identifiers and out_features_identifiers of each pipeline block are defined in a ``.yml`` file. In this example, we try to predict the ``mach`` based on two input scalars ``angle_in`` and ``mach_out``, and the mesh node coordinates.
+
+# In[ ]:
+
+
+try:
+    filename = Path(__file__).parent.parent.parent / "docs" / "source" / "notebooks" / "config_pipeline.yml"
+except NameError:
+    filename = "config_pipeline.yml"
+
+with open(filename, 'r') as f:
+    config = yaml.safe_load(f)
+
+all_feature_id = config['input_scalar_scaler']['in_features_identifiers'] +\
+    config['pca_nodes']['in_features_identifiers'] +\
+    config['pca_mach']['in_features_identifiers']
+
+
+dataset_train = dataset_train.from_features_identifier(all_feature_id)
+print(dataset_train)
+
+
+# Hence, we keep only 2 scalars and 1 field of interest.
+#
+# We now define a preprocessor: a `MinMaxScaler` of the 2 input scalars and a `PCA` on the nodes coordinates of the meshes:
+
+# In[ ]:
+
+
+preprocessor = PlaidColumnTransformer([
+    ('input_scalar_scaler', WrappedPlaidSklearnTransformer(MinMaxScaler(), **config['input_scalar_scaler'])),
[…]
