+++ docs/conf.py
@@ -160,6 +160,8 @@ autoapi_python_class_content = "both"  # default is 'class'
 #     #     - Classes
 #     #     - Functions
 #     #     - Methods
+jupyter_execute_notebooks = "force"
+execution_timeout = 300
 
 numfig = True
 
