+++ examples/run_examples.sh
@@ -1,9 +1,17 @@
 #!/bin/bash
-for file in *.py utils/*.py containers/*.py post/*.py pipelines/*.py
-    do echo "--------------------------------------------------------------------------------------"
-        echo "#---# run python $file"
-        python $file
-        if [ $? -ne 0 ]; then
-            exit 1
-        fi
-    done
+
+if [[ "$(uname)" == "Linux" ]]; then
+    FILES="*.py utils/*.py containers/*.py post/*.py pipelines/*.py"
+else
+    FILES="*.py utils/*.py containers/*.py post/*.py"
+fi
+
+for file in $FILES
+do
+    echo "--------------------------------------------------------------------------------------"
+    echo "#---# run python $file"
+    python "$file"
+    if [ $? -ne 0 ]; then
+        exit 1
+    fi
+done
