### 2025-07-26 7edb644
minor modifs
