+++ examples/pipelines/pipeline.py
@@ -2,10 +2,23 @@
 # coding: utf-8
 
 # # Pipeline Examples
-#
-# This notebook presents how the provided pipeline blocks can be used in a PCA-GP algorithm. Pipeline block directly link PLAID datasets, and hyperpamater tuning is available using scikit-learn's GridSearchCV or Optuna.
-#
-# We start by a few imports:
+# 
+# ## PCA-GP Regression on Flow Field Data
+# 
+# This notebook demonstrates the end-to-end process of building a machine learning pipeline for predicting high-dimensional flow fields from scalar input parameters using the PLAID dataset. The pipeline leverages PLAID’s scikit-learn-compatible blocks and standard dataset structures to enable modular, reusable workflows.
+# 
+# Key steps covered in this notebook:
+# 
+# - **Loading and preparing the PLAID dataset** using Hugging Face integration and PLAID’s dataset classes  
+# - **Standardizing features** with PLAID-wrapped scikit-learn transformers for scalars and fields  
+# - **Dimensionality reduction** of flow fields via Principal Component Analysis (PCA) to reduce output complexity  
+# - **Regression modeling** of PCA coefficients from scalar inputs using Gaussian Process regression  
+# - **Pipeline assembly** combining transformations and regressors into a single scikit-learn-compatible workflow  
+# - **Hyperparameter tuning** using Optuna and scikit-learn’s `GridSearchCV`
+# - **Model evaluation** using cross-validation and appropriate metrics  
+# - **Best practices** for working with PLAID datasets and pipelines in a reproducible and modular manner
+
+# ### 📦 Imports
 
 # In[ ]:
 
@@ -15,6 +28,13 @@ import yaml
 from pathlib import Path
 import optuna
 import numpy as np
+
+import warnings
+warnings.filterwarnings('ignore', module='sklearn')
+warnings.filterwarnings("ignore", module="tqdm")
+
+from datasets.utils.logging import disable_progress_bar
+from datasets import load_dataset
 
 from sklearn.base import clone
 from sklearn.pipeline import Pipeline
@@ -27,21 +47,17 @@ from sklearn.multioutput import MultiOutputRegressor
 
 from sklearn.model_selection import KFold, GridSearchCV
 
-import warnings
-from tqdm import TqdmWarning
-warnings.filterwarnings('ignore', module='sklearn')
-warnings.filterwarnings('ignore', category=TqdmWarning)
-
-from datasets import load_dataset, get_dataset_infos
 from plaid.bridges.huggingface_bridge import huggingface_dataset_to_plaid, huggingface_description_to_problem_definition
 from plaid.pipelines.sklearn_block_wrappers import WrappedPlaidSklearnTransformer, WrappedPlaidSklearnRegressor
 from plaid.pipelines.plaid_blocks import PlaidTransformedTargetRegressor, PlaidColumnTransformer
 
-nb_cpus = os.cpu_count()
-n_processes = max(1, int(nb_cpus / 4))
+disable_progress_bar()
+n_processes = min(max(1, os.cpu_count()), 6)
 
 
-# We load the `VKI-LS59` dataset from Hugging Face, and restrict ourselves to the first 24 samples of the training set.
+# ## 📥 Load Dataset
+# 
+# We load the `VKI-LS59` dataset from Hugging Face and restrict ourselves to the first 24 samples of the training set.
 
 # In[ ]:
 
@@ -49,10 +65,8 @@ n_processes = max(1, int(nb_cpus / 4))
 hf_dataset = load_dataset("PLAID-datasets/VKI-LS59", split="all_samples[:24]")
 dataset_train, _ = huggingface_dataset_to_plaid(hf_dataset, processes_number = n_processes, verbose = False)
 
-del hf_dataset
 
-
-# We print the summary of dataset_train:
[…]
