### 2025-07-26 153164c
attempt to remove warning from RTD executed notebook
