### 2025-07-26 394b829
attempt to remove warning from RTD executed notebook
