+++ docs/conf.py
@@ -220,8 +220,6 @@ github_url = "https://github.com/PLAID-lib/plaid"
 
 # -----------------------------------------------------------------------------#
 
-os.environ["PYTHONWARNINGS"] = "ignore:::tqdm.*"
-
 def skip_logger_attribute(app, what, name, obj, skip, options):
     if what == "data" and "logger" in name:
         print(f"WILL SKIP: {what=}, {name=}")
