### 2025-07-26 2742da3
feat(pipelines) greatly simplify block API, and impose all transformer to set in_features_identifiers_ and out_features_identifiers_ at fit
