+++ src/plaid/pipelines/plaid_blocks.py
@@ -20,6 +20,7 @@ from sklearn.compose import ColumnTransformer
 from sklearn.utils.validation import check_is_fitted
 
 from plaid.containers.dataset import Dataset
+from plaid.containers.utils import has_duplicates_feature_ids
 
 
 class PlaidColumnTransformer(ColumnTransformer):
@@ -33,27 +34,22 @@ class PlaidColumnTransformer(ColumnTransformer):
     Args:
         plaid_transformers: A list of tuples
             (name, transformer), where each `transformer` is a TransformerMixin.
-        remainder_feature_ids: List of feature identifiers to pass through
-            without transformation.
+
+    Notes:
+        At fit, it is checked that `plaid_transformers` share no in_features_identifiers and no out_features_identifiers.
     """
 
-    def __init__(
-        self,
-        plaid_transformers: list[tuple[str, TransformerMixin]] = None,
-        remainder_feature_ids: list[dict] = None,
-    ):
+    def __init__(self, plaid_transformers: list[tuple[str, TransformerMixin]] = None):
         self.plaid_transformers = plaid_transformers
-        self.remainder_feature_ids = remainder_feature_ids
 
         if plaid_transformers:
-            transformers_with_feat_ids = [
-                (name, transformer, transformer.get_params()["in_features_identifiers"])
-                for name, transformer in plaid_transformers
+            transformers_dummy = [
+                (name, transformer, "_") for name, transformer in plaid_transformers
             ]
         else:
-            transformers_with_feat_ids = None
+            transformers_dummy = None
 
-        super().__init__(transformers_with_feat_ids)
+        super().__init__(transformers_dummy)
 
     def fit(self, dataset, _y=None):
         """Fits all transformers on their corresponding feature subsets.
@@ -68,14 +64,39 @@ class PlaidColumnTransformer(ColumnTransformer):
         if isinstance(dataset, list):
             dataset = Dataset.from_list_of_samples(dataset)
 
-        self.plaid_transformers_ = copy.deepcopy(self.plaid_transformers)
-        self.remainder_feature_ids_ = copy.deepcopy(self.remainder_feature_ids)
+        self.in_features_identifiers_ = []
+        for _, transformer in self.plaid_transformers:
+            self.in_features_identifiers_ += copy.deepcopy(
+                transformer.in_features_identifiers
+            )
+
+        assert not has_duplicates_feature_ids(self.in_features_identifiers_), (
+            "Identical in_features_identifiers found among provided transformer: not compatible with PlaidColumnTransformer."
+        )
+
+        self.plaid_transformers_ = [
+            (copy.deepcopy(name), clone(transformer))
+            for name, transformer in self.plaid_transformers
+        ]
 
         self.transformers_ = []
-        for name, transformer, feat_ids in self.transformers:
-            sub_dataset = dataset.from_features_identifier(feat_ids)
+        for name, transformer in self.plaid_transformers_:
+            sub_dataset = dataset.from_features_identifier(
+                transformer.in_features_identifiers
+            )
             transformer_ = clone(transformer).fit(sub_dataset)
-            self.transformers_.append((name, transformer_, feat_ids))
+            self.transformers_.append((name, transformer_, "_"))
+
+        self.out_features_identifiers_ = []
+        for _, transformer, _ in self.transformers_:
+            self.out_features_identifiers_ += copy.deepcopy(
+                transformer.out_features_identifiers_
+            )
+
+        assert not has_duplicates_feature_ids(self.out_features_identifiers_), (
+            "Identical out_features_identifiers found among provided transformer: not compatible with PlaidColumnTransformer."
+        )
+
         return self
 
     def transform(self, dataset):
[…]
