+++ examples/pipelines/pipeline.py
@@ -3,7 +3,7 @@
 
 # # Pipeline Examples
 # 
-# ## PCA-GP Regression on Flow Field Data
+# ## PCA-GP on Field Data
 # 
 # This notebook demonstrates the end-to-end process of building a machine learning pipeline for predicting high-dimensional flow fields from scalar input parameters using the PLAID dataset. The pipeline leverages PLAID’s scikit-learn-compatible blocks and standard dataset structures to enable modular, reusable workflows.
 # 
@@ -20,18 +20,19 @@
 
 # ### 📦 Imports
 
-# In[ ]:
+# In[1]:
 
-
-import os
-import yaml
-from pathlib import Path
-import optuna
-import numpy as np
 
 import warnings
 warnings.filterwarnings('ignore', module='sklearn')
-warnings.filterwarnings("ignore", module="tqdm")
+warnings.filterwarnings("ignore", message=".*IProgress not found.*")
+
+import os
+from pathlib import Path
+
+import yaml
+import numpy as np
+import optuna
 
 from datasets.utils.logging import disable_progress_bar
 from datasets import load_dataset
@@ -59,7 +60,7 @@ n_processes = min(max(1, os.cpu_count()), 6)
 # 
 # We load the `VKI-LS59` dataset from Hugging Face and restrict ourselves to the first 24 samples of the training set.
 
-# In[ ]:
+# In[2]:
 
 
 hf_dataset = load_dataset("PLAID-datasets/VKI-LS59", split="all_samples[:24]")
@@ -68,7 +69,7 @@ dataset_train, _ = huggingface_dataset_to_plaid(hf_dataset, processes_number = n
 
 # We print the summary of dataset_train, which contains 24 samples, with 8 scalars and 8 fields, which is consistent with the VKI-LS59` dataset:
 
-# In[ ]:
+# In[3]:
 
 
 print(dataset_train)
@@ -88,7 +89,7 @@ print(dataset_train)
 #       name: reduced_nodes_*
 # ```
 
-# In[ ]:
+# In[4]:
 
 
 try:
@@ -105,7 +106,7 @@ all_feature_id = config['input_scalar_scaler']['in_features_identifiers'] +\
 
 # In this example, we aim to predict the ``mach`` based on two input scalars ``angle_in`` and ``mach_out``, and the mesh node coordinates. To contain memory consumption, we restrict the dataset to the features required for this example:
 
-# In[ ]:
+# In[5]:
 
 
 dataset_train = dataset_train.from_features_identifier(all_feature_id)
@@ -118,21 +119,23 @@ print(dataset_train)
 # 
 # We now define a preprocessor: a `MinMaxScaler` of the 2 input scalars and a `PCA` on the nodes coordinates of the meshes:
 
-# In[ ]:
+# In[6]:
 
 
-preprocessor = PlaidColumnTransformer([
-    ('input_scalar_scaler', WrappedPlaidSklearnTransformer(MinMaxScaler(), **config['input_scalar_scaler'])),
-    ('pca_nodes', WrappedPlaidSklearnTransformer(PCA(), **config['pca_nodes'])),
-], remainder_feature_ids = config['pca_mach']['in_features_identifiers'])
+preprocessor = PlaidColumnTransformer(
+    [
+        ('input_scalar_scaler', WrappedPlaidSklearnTransformer(MinMaxScaler(), **config['input_scalar_scaler'])),
+        ('pca_nodes', WrappedPlaidSklearnTransformer(PCA(), **config['pca_nodes'])),
+    ]
+)
 preprocessor
 
 
[…]
