### 2025-07-26 9d12a1f
improve notebook
