+++ examples/pipelines/pipeline.py
@@ -2,7 +2,7 @@
 # coding: utf-8
 
 # # Pipeline Examples
-# 
+
 # ## PCA-GP on Field Data
 # 
 # This notebook demonstrates the end-to-end process of building a machine learning pipeline for predicting high-dimensional flow fields from scalar input parameters using the PLAID dataset. The pipeline leverages PLAID’s scikit-learn-compatible blocks and standard dataset structures to enable modular, reusable workflows.
@@ -20,7 +20,7 @@
 
 # ### 📦 Imports
 
-# In[1]:
+# In[ ]:
 
 
 import warnings
@@ -60,16 +60,16 @@ n_processes = min(max(1, os.cpu_count()), 6)
 # 
 # We load the `VKI-LS59` dataset from Hugging Face and restrict ourselves to the first 24 samples of the training set.
 
-# In[2]:
+# In[ ]:
 
 
 hf_dataset = load_dataset("PLAID-datasets/VKI-LS59", split="all_samples[:24]")
 dataset_train, _ = huggingface_dataset_to_plaid(hf_dataset, processes_number = n_processes, verbose = False)
 
 
-# We print the summary of dataset_train, which contains 24 samples, with 8 scalars and 8 fields, which is consistent with the VKI-LS59` dataset:
+# We print the summary of dataset_train, which contains 24 samples, with 8 scalars and 8 fields, which is consistent with the `VKI-LS59` dataset:
 
-# In[3]:
+# In[ ]:
 
 
 print(dataset_train)
@@ -89,7 +89,7 @@ print(dataset_train)
 #       name: reduced_nodes_*
 # ```
 
-# In[4]:
+# In[ ]:
 
 
 try:
@@ -104,22 +104,24 @@ all_feature_id = config['input_scalar_scaler']['in_features_identifiers'] +\
     config['pca_nodes']['in_features_identifiers'] + config['pca_mach']['in_features_identifiers']
 
 
-# In this example, we aim to predict the ``mach`` based on two input scalars ``angle_in`` and ``mach_out``, and the mesh node coordinates. To contain memory consumption, we restrict the dataset to the features required for this example:
+# In this example, we aim to predict the ``mach`` field based on two input scalars ``angle_in`` and ``mach_out``, and the mesh node coordinates. To contain memory consumption, we restrict the dataset to the features required for this example:
 
-# In[5]:
+# In[ ]:
 
 
 dataset_train = dataset_train.from_features_identifier(all_feature_id)
-print(dataset_train)
+print("dataset_train:", dataset_train)
+print("scalar names =", dataset_train.get_scalar_names())
+print("field names =", dataset_train.get_field_names())
 
 
 # We notive that only the 2 scalars and the field of interest are kept after restriction.
 
-# 1. Preprocessor
+# ### 1. Preprocessor
 # 
 # We now define a preprocessor: a `MinMaxScaler` of the 2 input scalars and a `PCA` on the nodes coordinates of the meshes:
 
-# In[6]:
+# In[ ]:
 
 
 preprocessor = PlaidColumnTransformer(
@@ -135,29 +137,29 @@ preprocessor
 # 
 # To verify this behavior, we apply the `preprocessor` to `dataset_train`:
 
-# In[7]:
+# In[ ]:
 
 
 preprocessed_dataset = preprocessor.fit_transform(dataset_train)
-print(preprocessed_dataset)
+print("preprocessed_dataset:", preprocessed_dataset)
 print("scalar names =", preprocessed_dataset.get_scalar_names())
 print("field names =", preprocessed_dataset.get_field_names())
 
 
[…]
