### 2025-07-26 6ec3889
fix(dataset.py) typo in __repr__: no need to add an 's' for the plural of 'timeseries'
