### 2025-07-29 abf0c94
📄 docs(dataset): add __iter__ examples for Dataset (#118)
