### 2025-07-30 a2aa272
minor modifs
