### 2025-07-30 ad4e77b
save wip notebook pipeline MMGP
