### 2025-07-31 2f014d3
feat(notebooks) add MMGP pipeline
