+++ docs/source/notebooks/config_pipeline2.yml
@@ -1,3 +1,13 @@
+mmgp_nodes_transf:
+  in_features_identifiers:
+    - type: nodes
+
+mmgp_u1_transf:
+  in_features_identifiers:
+    - type: field
+      name: U1
+
+
 input_scalar_scaler:
   in_features_identifiers:
     - type: scalar
@@ -16,7 +26,6 @@ input_scalar_scaler:
 pca_nodes:
   in_features_identifiers:
     - type: nodes
-      time: 0.
   out_features_identifiers:
     - type: scalar
       name: reduced_nodes_*
