+++ docs/source/notebooks/MMGP_node.py
@@ -1,1202 +1,0 @@
-
-import copy
-from typing import Union
-
-from sklearn.base import (
-    BaseEstimator,
-    TransformerMixin,
-    clone,
-)
-from sklearn.utils.validation import check_is_fitted
-
-from plaid.containers.dataset import Dataset
-from plaid.containers.sample import Sample
-from plaid.containers.utils import get_feature_type_and_details_from_identifier, check_features_type_homogeneity
-
-import Muscat.Containers.ElementsDescription as ED
-import Muscat.Containers.Filters.FilterObjects as FilterObjects
-import Muscat.Containers.MeshModificationTools as MMT
-import networkx
-import numpy as np
-from Muscat.Containers.Mesh import Mesh
-from scipy import sparse
-
-
-import Muscat.Containers.Filters.FilterObjects as FilterObjects
-import Muscat.Containers.MeshInspectionTools as MIT
-import networkx
-import numpy as np
-from Muscat.Containers import MeshFieldOperations as MFO
-from Muscat.Containers.Mesh import Mesh
-from Muscat.FE import FETools as FT
-from Muscat.FE.Fields import FEField as FF
-from Muscat.Bridges.CGNSBridge import CGNSToMesh, MeshToCGNS
-import time
-from typing import Tuple, Callable
-
-
-
-
-class MMGPPreparer(TransformerMixin, BaseEstimator):
-    """Adapter for using a scikit-learn transformer on PLAID Datasets.
-
-    Transforms tabular data extracted from homogeneous feature identifiers,
-    and returns results as a `Dataset`. Supports forward and inverse transforms.
-
-    Args:
-        sklearn_block: A scikit-learn Transformer implementing fit/transform APIs.
-        in_features_identifiers: List of feature identifiers to extract input data from.
-        out_features_identifiers: List of feature identifiers used for outputs. If None,
-            defaults to `in_features_identifiers`.
-    """
-    # TODO: check if restrict_to_features=True can be used to reduce further memory consumption
-    def __init__(
-        self,
-        common_mesh_id: int = None
-    ):
-        self.common_mesh_id = common_mesh_id
-
-    def morph(self, mesh):
-        mesh_renumb, _, n_boundary = renumber_mesh_for_parametrization(
-            mesh, in_place=False)
-        mesh_renumb.elemFields = mesh_renumb.nodeFields = {}
-        morphed_mesh, _ = floater_mesh_parametrization(
-            mesh_renumb, n_boundary)
-        return morphed_mesh
-
-    def fit(self, dataset: Dataset, _y=None):
-        """Fits the underlying scikit-learn transformer on selected input features.
-
-        Args:
-            dataset: A `Dataset` object containing the features to transform.
-            _y: Ignored.
-
-        Returns:
-            self: The fitted transformer.
-        """
-
-        self.common_mesh_id_ = 0 if self.common_mesh_id is None else self.common_mesh_id
-
-        self.morphed_common_mesh_ = self.morph(CGNSToMesh(dataset[self.common_mesh_id_].get_mesh()))
-        self.morphed_common_tree_ = MeshToCGNS(self.morphed_common_mesh_, exportOriginalIDs=False, tagsAsFields=False)
-
-        return self
-
-    def transform(self, dataset: Dataset):
-        """Applies the fitted transformer to the selected input features.
-
-        Args:
-            dataset: A `Dataset` object to transform.
-
-        Returns:
-            Dataset: Transformed features wrapped as a new `Dataset`.
-        """
-        check_is_fitted(self, "morphed_common_mesh_")
-
-        # n_dim = self.morphed_common_mesh_.GetPointsDimensionality()
[…]
