+++ benchmarks/Rotor37/MMGP/run.py
@@ -1,0 +1,271 @@
+import pickle
+from typing import List, Tuple
+
+import numpy as np
+from datasets import load_dataset
+from GPy.kern import RBF, Matern32, Matern52
+from GPy.models import GPRegression
+from sklearn.base import BaseEstimator, RegressorMixin, clone
+from sklearn.compose import ColumnTransformer, TransformedTargetRegressor
+from sklearn.decomposition import PCA
+from sklearn.pipeline import Pipeline
+from sklearn.preprocessing import MinMaxScaler, StandardScaler
+
+from plaid.containers.sample import Sample
+
+dataset = load_dataset("PLAID-datasets/Rotor37", split="all_samples")
+
+ids_train = dataset.description["split"]["train_1000"]
+ids_test = dataset.description["split"]["test"]
+
+out_fields_names = ["Density", "Pressure", "Temperature"]
+out_scalars_names = ["Massflow", "Compression_ratio", "Efficiency"]
+
+
+def convert_data(
+    ids: List[int],
+) -> Tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray, np.ndarray, np.ndarray]:
+    """Converts a list of sample IDs into structured numpy arrays containing input and output data.
+
+    Parameters:
+    ----------
+    ids : List[int]
+        List of sample indices to retrieve from the dataset.
+
+    Returns:
+    -------
+    nodes : np.ndarray
+        Flattened array of node features for each sample.
+    X_scalars : np.ndarray
+        Array containing input scalar values (Omega, Pressure) for each sample.
+    Y_scalars : np.ndarray
+        Array containing output scalar values (Massflow, Compression Ratio, Efficiency).
+    Y_density : np.ndarray
+        Array containing field values for Density across samples.
+    Y_pressure : np.ndarray
+        Array containing field values for Pressure across samples.
+    Y_temperature : np.ndarray
+        Array containing field values for Temperature across samples.
+    """
+    X_scalars = []
+    Y_scalars = []
+    Y_density = []
+    Y_pressure = []
+    Y_temperature = []
+    nodes = []
+
+    for id in ids:
+        sample = Sample.model_validate(pickle.loads(dataset[id]["sample"]))
+
+        nodes.append(sample.get_nodes())
+        omega = sample.get_scalar("Omega")
+        pressure = sample.get_scalar("P")
+
+        density = sample.get_field("Density")
+        pressure_field = sample.get_field("Pressure")
+        temperature = sample.get_field("Temperature")
+
+        massflow = sample.get_scalar("Massflow")
+        compression_ratio = sample.get_scalar("Compression_ratio")
+        efficiency = sample.get_scalar("Efficiency")
+
+        X_scalars.append(np.array([omega, pressure]))
+        Y_scalars.append(np.array([massflow, compression_ratio, efficiency]))
+        Y_density.append(density)
+        Y_pressure.append(pressure_field)
+        Y_temperature.append(temperature)
+
+    # Convert lists to numpy arrays
+    nodes = np.stack(nodes).reshape(len(nodes), -1)
+    X_scalars = np.stack(X_scalars)
+    Y_scalars = np.stack(Y_scalars)
+    Y_density = np.stack(Y_density)
+    Y_pressure = np.stack(Y_pressure)
+    Y_temperature = np.stack(Y_temperature)
+
+    return nodes, X_scalars, Y_scalars, Y_density, Y_pressure, Y_temperature
+
+
+class GPyRegressor(BaseEstimator, RegressorMixin):
+    """Custom Gaussian Process Regressor using GPy library.
+
+    Args:
+        normalizer (bool): Whether to normalize the output.
+        kernel (str): Type of kernel to use in the Gaussian Process.
[…]
