+++ benchmarks/Rotor37/README.md
@@ -1,0 +1,1 @@
+Scripts for generating the benchmarking entries of https://huggingface.co/spaces/PLAIDcompetitions/Rotor37Benchmark, where all instructions for participating are detailed.
