+++ benchmarks/2D_ElPlDynamics/README.md
@@ -1,0 +1,1 @@
+Scripts for generating the benchmarking entries of https://huggingface.co/spaces/PLAIDcompetitions/2DElastoPlastoDynamics, where all instructions for participating are detailed.
