+++ benchmarks/2D_profile/FNO/train.py
@@ -150,4 +150,4 @@ dataset[ids_test]._save_to_dir_(predicted_data_dir)
 
 
 print("duration train =", time.time()-start)
-# GPUA30, 1117 seconds
+# GPUA30, 8887 seconds
