### 2025-08-02 fd99d3b
benchmarks: typo Readme
