### 2025-08-03 3135f20
benchmarks: add README
