### 2025-08-03 b7533dd
benchmark: improve readme
