### 2025-08-03 af2e734
benchmark: improve readme
