### 2025-08-03 275af90
benchmarks: add dependencies
