### 2025-08-04 63c9b20
reorganize benchmark folder by method instead of by dataset
