+++ benchmarks/README.md
@@ -24,7 +24,7 @@ We believe this additional work fully addresses points W1 and W2 raised in the i
 |-------------------|-----|------|------------|-------|-------|-------|
 | `Tensile2d`       |  0.0673 |  **0.0026**  |   0.0058     |  0.0154   |  0.021  |  *0.0038*  |
 | `2D_MultiScHypEl` | 0.0437  |  ❌  |   *0.0341*     |  **0.0232**   |   0.0439  |  0.0573  |
-| `2D_ElPlDynamics` | 🕑  |  ❌  |   🕑     |  🕑    |  **0.0158**  |  🕑  |
+| `2D_ElPlDynamics` | 🕑  |  ❌  |   🕑     |  🕑    |  **0.0215**  |  🕑  |
 | `Rotor37`         | 0.0074  |  **0.0014**  |   0.0029     |  0.0033   |   0.0313  |  *0.0017*  |
 | `2D_profile`      |  0.0593 |  0.0365  |   *0.0319*     |  0.0425   |  0.0972  |  **0.0307**  |
 | `VKI-LS59`        | 0.0684  |  0.0312  |   0.0493     |  *0.0267*   |   0.0581  |  **0.0124**  |
