+++ benchmarks/2D_MultiScHypEl/README.md
@@ -1,3 +1,0 @@
-Scripts for generating the benchmarking entries of https://huggingface.co/spaces/PLAIDcompetitions/2DMultiscaleHyperelasticityBenchmark, where all instructions for participating are detailed.
-
-MMGP is not applicable on this dataset (due to topology variations)
