### 2025-08-04 7f25661
🎉 Add code FNO in benchmarks/FNO/2D_ElPlDynamics
