+++ benchmarks/FNO/2D_ElPlDynamics/build_pred.py
@@ -1,0 +1,348 @@
+from Muscat.Bridges.CGNSBridge import CGNSToMesh
+import numpy as np
+from Muscat.MeshTools.MeshFieldOperations import GetFieldTransferOp
+from Muscat.FE.Fields.FEField import FEField
+from Muscat.FE.Spaces.FESpaces import LagrangeSpaceGeo
+from Muscat.FE.DofNumbering import ComputeDofNumbering
+from Muscat.Bridges.CGNSBridge import CGNSToMesh
+import Muscat.MeshContainers.ElementsDescription as ED
+from Muscat.MeshTools.ConstantRectilinearMeshTools import CreateConstantRectilinearMesh
+from Muscat.MeshTools.MeshTetrahedrization import Tetrahedrization
+
+from scipy.sparse import coo_matrix
+import copy
+import torch
+from physicsnemo.models.fno.fno import FNO
+import torch.nn as nn
+from Muscat.MeshTools.MeshTools import ComputeSignedDistance
+from plaid.containers.dataset import Dataset as Plaid_Dataset
+import pickle
+
+
+def renormalize(values):
+    """
+    Function to normalize input of the FNO
+    """
+    input_values = copy.deepcopy(values)
+    input_values[:, [2]] = torch.sigmoid(input_values[:, [2]]/0.1)
+    xhi = input_values[:, [2]]
+    input_values[:, [0, 1]] /= 20
+    return input_values, xhi
+
+
+def denormalize(values):
+    """
+    Inverse function of renormalise except the mask that we do not need
+    """
+    input_values = copy.deepcopy(values)
+    # not denormalising mask
+    # input_values[:,[2]] = 10*torch.log(input_values[:,[2]]/(1-input_values[:,[2]]))
+    xhi = input_values[:, [2]]
+    input_values[:, [0, 1]] *= 20
+    return input_values, xhi
+
+
+def preproccess_sample(sample):
+    """Function to project the sample to a rectilinear grid using Muscat
+    This function only return the initial timestep as it is the only information needed
+    in the autoregressive method (when testing we simply apply recusively the model)
+
+    Parameters
+    ----------
+    sample : plaid.Sample
+        plaid sample to be projected to the rectilinear grid
+
+    Returns
+    -------
+    input_fields: torch.Tensor
+        2D fields (shape: (1,3,301,151)) ready to be used as a 2D image
+    xhi: torch.Tensor
+        2D mask (shape: (1,1,301,151)) ready to be used as a 2D image
+    """
+    old_mesh = CGNSToMesh(sample.get_mesh(time=0))
+    # Switching to 2D mesh instead of 3D shell, need C order for cython
+    old_mesh.nodes = (old_mesh.nodes[:, [0, 1]]).copy(order='C')
+    size = 150
+    ref_mesh = Tetrahedrization(CreateConstantRectilinearMesh(
+        [size*2+1, size+1], [0, 0], [100/size, 100/size]))
+
+    indexes = np.zeros(ref_mesh.GetNumberOfNodes(), int)
+    values = np.zeros((ref_mesh.GetNumberOfNodes(), 15))
+    for connect in ref_mesh.GetElementsOfType(ED.Triangle_3).connectivity:
+        values[connect, indexes[connect]] = 1
+        indexes[connect] += 1
+    average = 1/indexes
+    data, i, j = np.zeros(0), np.zeros(0).astype(int), np.zeros(0).astype(int)
+    for elem_index, connect in enumerate(ref_mesh.GetElementsOfType(ED.Triangle_3).connectivity):
+        data = np.concatenate((data, average[connect]), axis=0)
+        i = np.concatenate((i, connect), axis=0)
[…]
