+++ benchmarks/FNO/2D_ElPlDynamics/convert_to_rectilinear_grid.py
@@ -1,0 +1,153 @@
+from plaid.containers.dataset import Dataset
+from plaid.containers.sample import Sample
+from Muscat.Bridges.CGNSBridge import CGNSToMesh
+import numpy as np
+from Muscat.MeshTools.MeshFieldOperations import GetFieldTransferOp
+from Muscat.FE.Fields.FEField import FEField
+from Muscat.FE.Spaces.FESpaces import LagrangeSpaceGeo
+from Muscat.FE.DofNumbering import ComputeDofNumbering
+from Muscat.Bridges.CGNSBridge import MeshToCGNS, CGNSToMesh
+import Muscat.MeshContainers.ElementsDescription as ED
+from Muscat.MeshTools.ConstantRectilinearMeshTools import CreateConstantRectilinearMesh
+from Muscat.MeshTools.MeshTetrahedrization import Tetrahedrization
+from Muscat.MeshTools.MeshModificationTools import DeleteElements
+from Muscat.MeshTools.MeshTools import ComputeSignedDistance
+from scipy.sparse import coo_matrix
+import copy
+import os
+
+
+datapath = "/path/to/plaid/dataset"
+output_data_dir = "/path/where/to/register/new_dataset"
+dataset = Dataset()
+for sample_index in range(1000):
+    dataset._load_from_dir_(
+        savedir=datapath,
+        verbose=True,
+        processes_number=1,
+        ids=[sample_index])
+    sample = dataset.get_samples([sample_index])[sample_index]
+    old_mesh = CGNSToMesh(sample.get_mesh(time=0))
+
+    old_mesh.nodes = (old_mesh.nodes[:, [0, 1]]).copy(order='C')
+    size = 150
+    ref_mesh = Tetrahedrization(CreateConstantRectilinearMesh(
+        [size * 2 + 1, size + 1], [0, 0], [100 / size, 100 / size]))
+
+    indexes = np.zeros(ref_mesh.GetNumberOfNodes(), int)
+    values = np.zeros((ref_mesh.GetNumberOfNodes(), 15))
+    for connect in ref_mesh.GetElementsOfType(ED.Triangle_3).connectivity:
+        values[connect, indexes[connect]] = 1
+        indexes[connect] += 1
+    average = 1 / indexes
+    data, i, j = np.zeros(0), np.zeros(
+        0).astype(int), np.zeros(0).astype(int)
+    for elem_index, connect in enumerate(
+            ref_mesh.GetElementsOfType(ED.Triangle_3).connectivity):
+        data = np.concatenate((data, average[connect]), axis=0)
+        i = np.concatenate((i, connect), axis=0)
+        j = np.concatenate((j, [elem_index] * 3), axis=0)
+
+    operator_elem_to_node = coo_matrix((data, (i, j)))
+
+    # Compute Field Transfer operator for node fields
+    space = LagrangeSpaceGeo
+    numbering = ComputeDofNumbering(old_mesh, space, fromConnectivity=True)
+    displacement_field = FEField("FakeField", old_mesh, space, numbering)
+    op, _, _ = GetFieldTransferOp(
+        displacement_field, ref_mesh.nodes, method="Interp/Clamp")
+
+    # Compute elem to node operator
+    triangle_centers = np.mean(
+        ref_mesh.nodes[ref_mesh.GetElementsOfType(ED.Triangle_3).connectivity], axis=-2)
+    # entities contains the index of the element for each triangle_center
+    _, _, entities = GetFieldTransferOp(
+        displacement_field, triangle_centers, method="Interp/Clamp")
+    data = np.ones_like(entities.squeeze())
+    i = old_mesh.GetElementsOfType(ED.Triangle_3).GetNumberOfElements()
[…]
