### 2025-08-04 a3ef0ce
update benchmarks/FNO/2D_ElPlDynamics
