### 2025-08-04 43c8d45
Fixed the Tensile2D Vi-Transformer score
