+++ benchmarks/MMGP/2D_profile/train_and_predict.py
@@ -1,0 +1,222 @@
+import numpy as np
+import copy
+import pickle
+import time
+from sklearn.preprocessing import StandardScaler, MinMaxScaler
+
+from Muscat.IO.CGNSReader import ReadCGNS
+from Muscat.FieldTransferKokkos.FieldTransfer import FieldTransferKokkos
+from Muscat.Containers.Filters.FilterObjects import ElementFilter
+from Muscat.FE.Fields.FEField import FEField
+from Muscat.FE.FETools import PrepareFEComputation
+from Muscat.Containers.Filters import FilterObjects as FO
+from Muscat.IO import XdmfReader as XR
+from scipy.sparse import identity
+
+from model import GPyRegressor
+
+from utils_2dprofile import POD
+import os
+
+plaid_location = # path to update to input plaid dataset
+
+
+
+start=time.time()
+
+# plaid_location_coarse = # path to update to plaid dataset containing a sample with the coarse common mesh
+# reference_mesh_index=0
+# reference_mesh_path = os.path.join(plaid_location_coarse, "dataset/samples/sample_00000000"+str(reference_mesh_index)+"/meshes/mesh_000000000.cgns")
+# reference_mesh = ReadCGNS(fileName=reference_mesh_path)
+reference_mesh = XR.ReadXdmf("coarse_common_mesh.xdmf")
+
+reference_mesh_path = os.path.join(plaid_location, "dataset/samples/sample_00000000"+str(reference_mesh_index)+"/meshes/mesh_000000000.cgns")
+reference_mesh_fin = ReadCGNS(fileName=reference_mesh_path)
+
+space, numbering,_,_ = PrepareFEComputation(reference_mesh,numberOfComponents=1)
+fefield = FEField("", mesh=reference_mesh, space=space, numbering=numbering[0])
+op_ref_mesh=FieldTransferKokkos(inputField= fefield, targetPoints= reference_mesh_fin.nodes, method="Interp/Clamp" , elementFilter= ElementFilter(dimensionality=2) )[0]
+
+
+
+node_fields=['Mach', 'Pressure', 'Velocity-x', 'Velocity-y']
+n_train=300
+nNodes_referenceMesh=reference_mesh_fin.nodes.shape[0]
+
+flattend_data_morphing=np.zeros((n_train,2*reference_mesh.nodes.shape[0]))
+
+
+
+# transfer data on reference mesh by pull back
+
+data={}
+
+for field_name in node_fields:
+    data[field_name] = np.zeros((n_train,nNodes_referenceMesh))
+
+morphing_displacement_fields=np.zeros((n_train,nNodes_referenceMesh,2))
+
+
+for sample in range(n_train):
+    print(sample)
+    Tmesh_index=str(sample).zfill(3)
+
+    file = open("displacement_field/displacement_field"+Tmesh_index+".pkl", 'rb')
+    morphing_displacement_field_coarse = pickle.load(file)
+    file.close()
+
+    morphing_displacement_fields[sample] = op_ref_mesh.dot(morphing_displacement_field_coarse)
+
+    temp_mesh=copy.deepcopy(reference_mesh_fin)
+
+    temp_mesh.nodes +=  morphing_displacement_fields[sample]
+
+
+    Tmesh_path = os.path.join(plaid_location, "dataset/samples/sample_00000000"+str(Tmesh_index)+"/meshes/mesh_000000000.cgns")
+    Tmesh = ReadCGNS(fileName=Tmesh_path)
+
+    space, numbering,_,_ = PrepareFEComputation(Tmesh,numberOfComponents=1)
[…]
