+++ benchmarks/MMGP/2D_profile/create_coarse_common_mesh.py
@@ -1,9 +1,6 @@
-from plaid.containers.sample import Sample
 from Muscat.IO import XdmfWriter as XW
-from Muscat.Containers.MmgRemeshing import MmgRemeshing
+from Muscat.MeshTools.Remesh import Remesh
 from Muscat.Containers import AnisotropicMetricComputation as AMC
-# import OpenPisco.Unstructured.MmgMesher as MmgMesher
-import Muscat.Helpers.ParserHelper as PH
 import numpy as np
 from Muscat.Containers.MeshModificationTools import ComputeSkin
 
@@ -17,8 +14,7 @@ from Muscat.IO.CGNSReader import ReadCGNS
 import os
 
 
-plaid_location = # path to update to input plaid dataset
-
+plaid_location = "/data/ssa/units/stn/mads/flex/datasets/C1/2D_profile/large_dataset/clipped/plaid" # path to update to input plaid dataset
 
 
 reference_mesh_index = 0
@@ -26,12 +22,13 @@ reference_mesh_index = 0
 common_mesh_path = os.path.join(plaid_location, "dataset/samples/sample_00000000"+str(reference_mesh_index)+"/meshes/mesh_000000000.cgns")
 mesh = ReadCGNS(fileName=common_mesh_path)
 
+
 mesh.nodes = np.ascontiguousarray(mesh.nodes[:,:2])
 
 m = mesh.nodeFields['Mach']
 p = mesh.nodeFields['Pressure']
-Ux = mesh.nodeFields['Velocity'][:,0]
-Uy = mesh.nodeFields['Velocity'][:,1]
+Ux = mesh.nodeFields['Velocity-x']
+Uy = mesh.nodeFields['Velocity-y']
 
 
 m_scaled = (1.0 / np.std(m)) * (m - np.mean(m))
@@ -40,23 +37,21 @@ norm_U = np.sqrt(Ux**2 + Uy**2)
 norm_U_scaled = (1.0 / np.std(norm_U)) * (norm_U - np.mean(norm_U))
 psi = 1.*p_scaled + 1.*m_scaled + 0.*norm_U_scaled
 
-metric = AMC.ComputeMetric(mesh, psi, err = 0.03, hmin = 0.005, hmax = 1000, gradL2 = False)
+metric = AMC.ComputeMetric(mesh, psi, err = 0.03, gradL2 = False)
+# metric = AMC.ComputeMetric(mesh, psi, err = 0.03, hmin = 0.005, hmax = 1000, gradL2 = False)
 
-obj = MmgRemeshing(mesh = mesh, solution = None, metric = metric)
-remeshed_mesh = obj.Runner(options = {}, temp_name = "temp", binary = True)
+remeshed_mesh = Remesh(mesh = mesh, solution = None, metric = metric)
 
-
-
-ComputeSkin(remeshed_mesh, inPlace=True)
+remeshed_mesh = ComputeSkin(remeshed_mesh, inPlace=True)
 
 nf_skin = FO.NodeFilter(eTag =["Skin"])
 indices_skin = nf_skin.GetNodesIndices(remeshed_mesh)
-# remeshed_mesh.GetNodalTag("Skin").AddToTag(indices_skin)
 
 nf1 = FO.NodeFilter(eTag =["Skin"], zone=[lambda xyz: (xyz[:, 0] - 1.5)])
 nf2 = FO.NodeFilter(eTag =["Skin"], zone=[lambda xyz: (xyz[:, 1] - 0.5)])
[…]
