### 2025-08-04 6598055
benchmarks: various minor modifs
