+++ benchmarks/FNO/Tensile2d/train_and_predict.py
@@ -15,6 +15,8 @@ plaid_location = # path to update
 prepared_data_dir = # path to update
 predicted_data_dir = # path to update
 
+
+
 pb_defpath=os.path.join(plaid_location, "problem_definition")
 
 dataset = Dataset()
@@ -23,7 +25,7 @@ dataset._load_from_dir_(os.path.join(prepared_data_dir, "dataset"), verbose=True
 problem = ProblemDefinition()
 problem._load_from_dir_(pb_defpath)
 
-ids_train = problem.get_split('train_1000')
+ids_train = problem.get_split('train_500')
 ids_test  = problem.get_split('test')
 
 
@@ -31,42 +33,42 @@ n_train = len(ids_train)
 n_test  = len(ids_test)
 
 
-in_scalars_names = ['Omega', 'P']
-out_fields_names = ['Density', 'Pressure', 'Temperature']
-out_scalars_names = ['Massflow', 'Compression_ratio', 'Efficiency']
+in_scalars_names = ["P","p1","p2","p3","p4","p5"]
+out_fields_names = ["U1","U2","sig11","sig22","sig12"]
+out_scalars_names = ["max_von_mises","max_U2_top",'max_sig22_top']
 
-
-size = 64
+size = 201
 
 
 # TRAIN
-inputs = np.empty((n_train, len(in_scalars_names)+1, size, size, size))
+
+inputs = np.empty((n_train, len(in_scalars_names)+1, size, size))
 for i, id_sample in enumerate(ids_train):
     for in_chan in range(len(in_scalars_names)+1):
-        inputs[i, in_chan, :, :, :] = dataset[id_sample].get_field("Signed_Distance").reshape((size, size, size))
+        inputs[i, in_chan, :, :] = dataset[id_sample].get_field("Signed_Distance").reshape((size, size))
         for k, sn in enumerate(in_scalars_names):
-            inputs[i, k+1, :, :, :] = dataset[id_sample].get_scalar(sn)
+            inputs[i, k+1, :, :] = dataset[id_sample].get_scalar(sn)
 
-outputs = np.empty((n_train, len(out_scalars_names)+len(out_fields_names), size, size, size))
+outputs = np.empty((n_train, len(out_scalars_names)+len(out_fields_names), size, size))
 for i, id_sample in enumerate(ids_train):
     for k, fn in enumerate(out_fields_names):
-        outputs[i, k, :, :, :] = dataset[id_sample].get_field(fn).reshape((size, size, size))
+        outputs[i, k, :, :] = dataset[id_sample].get_field(fn).reshape((size, size))
     for k, sn in enumerate(out_scalars_names):
-        outputs[i, k+len(out_fields_names), :, :, :] = dataset[id_sample].get_scalar(sn)
+        outputs[i, k+len(out_fields_names), :, :] = dataset[id_sample].get_scalar(sn)
 
 
-min_in = inputs.min(axis=(0, 2, 3, 4), keepdims=True)
-max_in = inputs.max(axis=(0, 2, 3, 4), keepdims=True)
-inputs = (inputs - min_in) / (max_in - min_in + 1e-8)
+min_in = inputs.min(axis=(0, 2, 3), keepdims=True)
+max_in = inputs.max(axis=(0, 2, 3), keepdims=True)
+inputs = (inputs - min_in) / (max_in - min_in)
 
 
-min_out = outputs.min(axis=(0, 2, 3, 4), keepdims=True)
-max_out = outputs.max(axis=(0, 2, 3, 4), keepdims=True)
-outputs = (outputs - min_out) / (max_out - min_out + 1e-8)
[…]
