+++ benchmarks/FNO/2D_MultiScHypEl/train_and_predict.py
@@ -15,6 +15,7 @@ plaid_location = # path to update
 prepared_data_dir = # path to update
 predicted_data_dir = # path to update
 
+
 pb_defpath=os.path.join(plaid_location, "problem_definition")
 
 dataset = Dataset()
