+++ benchmarks/FNO/VKI-LS59/prepare_vki.py
@@ -15,6 +15,9 @@ start = time.time()
 plaid_location = # path to update
 prepared_data_dir = # path to update
 
+
+
+
 datapath=os.path.join(plaid_location, "dataset")
 pb_defpath=os.path.join(plaid_location, "problem_definition")
 
