+++ benchmarks/FNO/Rotor37/train_and_predict.py
@@ -13,7 +13,9 @@ start = time.time()
 
 plaid_location = # path to update
 prepared_data_dir = # path to update
-predicted_data_dir= # path to update
+predicted_data_dir = # path to update
+
+
 
 pb_defpath=os.path.join(plaid_location, "problem_definition")
 
@@ -23,7 +25,7 @@ dataset._load_from_dir_(os.path.join(prepared_data_dir, "dataset"), verbose=True
 problem = ProblemDefinition()
 problem._load_from_dir_(pb_defpath)
 
-ids_train = problem.get_split('train')
+ids_train = problem.get_split('train_1000')
 ids_test  = problem.get_split('test')
 
 
@@ -31,44 +33,42 @@ n_train = len(ids_train)
 n_test  = len(ids_test)
 
 
-in_scalars_names = ['angle_in', 'mach_out']
-out_fields_names = ['mach', 'nut']
-out_scalars_names = ['Q', 'power', 'Pr', 'Tr', 'eth_is', 'angle_out']
+in_scalars_names = ['Omega', 'P']
+out_fields_names = ['Density', 'Pressure', 'Temperature']
+out_scalars_names = ['Massflow', 'Compression_ratio', 'Efficiency']
 
 
-size1 = 301
-size2 = 121
+size = 64
 
 
 # TRAIN
-
-inputs = np.empty((n_train, len(in_scalars_names)+1, size1, size2))
+inputs = np.empty((n_train, len(in_scalars_names)+1, size, size, size))
 for i, id_sample in enumerate(ids_train):
     for in_chan in range(len(in_scalars_names)+1):
-        inputs[i, in_chan, :, :] = dataset[id_sample].get_field("Signed_Distance", base_name="Base_2_2").reshape((size1, size2))
+        inputs[i, in_chan, :, :, :] = dataset[id_sample].get_field("Signed_Distance").reshape((size, size, size))
         for k, sn in enumerate(in_scalars_names):
-            inputs[i, k+1, :, :] = dataset[id_sample].get_scalar(sn)
+            inputs[i, k+1, :, :, :] = dataset[id_sample].get_scalar(sn)
 
-outputs = np.empty((n_train, len(out_scalars_names)+len(out_fields_names), size1, size2))
+outputs = np.empty((n_train, len(out_scalars_names)+len(out_fields_names), size, size, size))
 for i, id_sample in enumerate(ids_train):
     for k, fn in enumerate(out_fields_names):
-        outputs[i, k, :, :] = dataset[id_sample].get_field(fn, base_name="Base_2_2").reshape((size1, size2))
+        outputs[i, k, :, :, :] = dataset[id_sample].get_field(fn).reshape((size, size, size))
     for k, sn in enumerate(out_scalars_names):
-        outputs[i, k+len(out_fields_names), :, :] = dataset[id_sample].get_scalar(sn)
+        outputs[i, k+len(out_fields_names), :, :, :] = dataset[id_sample].get_scalar(sn)
 
 
-min_in = inputs.min(axis=(0, 2, 3), keepdims=True)
-max_in = inputs.max(axis=(0, 2, 3), keepdims=True)
-inputs = (inputs - min_in) / (max_in - min_in)
+min_in = inputs.min(axis=(0, 2, 3, 4), keepdims=True)
+max_in = inputs.max(axis=(0, 2, 3, 4), keepdims=True)
+inputs = (inputs - min_in) / (max_in - min_in + 1e-8)
 
 
-min_out = outputs.min(axis=(0, 2, 3), keepdims=True)
-max_out = outputs.max(axis=(0, 2, 3), keepdims=True)
[…]
