+++ benchmarks/FNO/2D_MultiScHypEl/prepare_2D_MultiScHypEl.py
@@ -19,6 +19,9 @@ import os, shutil
 import time
 
 
+plaid_location = # path to update
+prepared_data_dir= # path to update
+
 
 start = time.time()
 
@@ -48,8 +51,6 @@ def compute_signed_distance(mesh,eval_points):
     return sign_distance*distance
 
 
-plaid_location = # path to update
-prepared_data_dir= # path to update
 
 datapath=os.path.join(plaid_location, "dataset")
 pb_defpath=os.path.join(plaid_location, "problem_definition")
