+++ benchmarks/FNO/Rotor37/prepare_rotor37.py
@@ -17,6 +17,13 @@ from tqdm import tqdm
 
 
 
+
+plaid_location = # path to update
+prepared_data_dir = # path to update
+
+
+
+
 start = time.time()
 
 
@@ -42,9 +49,6 @@ def compute_signed_distance(mesh,eval_points):
     return sign_distance*distance
 
 
-
-plaid_location = # path to update
-prepared_data_dir = # path to update
 
 datapath=os.path.join(plaid_location, "dataset")
 pb_defpath=os.path.join(plaid_location, "problem_definition")
