### 2025-08-04 9d2cb0b
correct MMGP/2D_profile: remove FieldTransfertKokkos
