### 2025-08-04 f786967
benchmarks: readme, improve note on MMGP
