### 2025-08-05 9c10fcc
improve 2 FNO scores
