+++ benchmarks/FNO/2D_MultiScHypEl/train_and_predict.py
@@ -67,7 +67,7 @@ outputs = (outputs - min_out) / (max_out - min_out)
 
 
 import torch
-from torch.utils.data import Dataset
+from torch.utils.data import Dataset, TensorDataset
 
 class GridDataset(Dataset):
     def __init__(self, inputs, outputs):
@@ -89,11 +89,11 @@ loader = DataLoader(dataset__, batch_size=64, shuffle=True)
 model = FNO(
 in_channels=inputs.shape[1],
 out_channels=outputs.shape[1],
-decoder_layers=2,
-decoder_layer_size=32,
+decoder_layers=4,
+decoder_layer_size=64,
 dimension=2,
-latent_channels=32,
-num_fno_layers=2,
+latent_channels=64,
+num_fno_layers=4,
 padding=0,
 ).cuda()
 
@@ -101,7 +101,7 @@ padding=0,
 optimizer = torch.optim.Adam(model.parameters(), lr=1e-3)
 loss_fn = torch.nn.MSELoss()
 
-n_epoch = 500
+n_epoch = 2000
 for epoch in range(n_epoch):
     model.train()
     total_loss = 0.0
@@ -132,12 +132,22 @@ inputs = (inputs - min_in) / (max_in - min_in)
 
 
 
+test_dataset = TensorDataset(torch.tensor(inputs, dtype=torch.float32))
+test_loader = DataLoader(test_dataset, batch_size=1, shuffle=False)
+
 model.eval()
+y_preds = []
+
 with torch.no_grad():
-    x_test = torch.tensor(inputs, dtype=torch.float32).cuda()
-    y_pred = model(x_test).cpu().numpy()
+    for batch in test_loader:
+        x_batch = batch[0].cuda()
+        y_batch_pred = model(x_batch).cpu()
+        y_preds.append(y_batch_pred)
+
+y_pred = torch.cat(y_preds, dim=0).numpy()
 
 outputs_pred = y_pred * (max_out - min_out) + min_out
+
 
 
 for i, id_sample in enumerate(ids_test):
