### 2025-08-05 7b3968e
benchmarks/readme: typos
