### 2025-08-05 f54276b
benchmarks/readme: rephrasing
