### 2025-08-05 c725146
benchmarks/MARIO: remove folder and provide links in a glocal README
