+++ benchmarks/MARIO/README.md
@@ -1,0 +1,8 @@
+The codes to run the benchmark is located in the repo [MARIO](https://github.com/giovannicatalani/MARIO). Dedicated folders are:
+- `Tensile2d`
+- `2D_MultiScHypEl`
+- `2D_ElPlDynamics`
+- [`Rotor37`](https://github.com/giovannicatalani/MARIO/tree/main/Rotor37_task)
+- [`2D_profile`](https://github.com/giovannicatalani/MARIO/tree/main/2Dprofile_task)
+- [`VKI-LS59`](https://github.com/giovannicatalani/MARIO/tree/main/VKILS59_task)
+
