+++ benchmarks/MGN/main_2d_elasto_ddp.py
@@ -1,0 +1,458 @@
+import os
+import pickle
+import argparse
+import time
+
+import torch
+import torch.distributed as dist
+from torch.utils.data import Dataset
+from torch.utils.data.distributed import DistributedSampler
+
+import numpy as np
+import dgl
+from dgl.dataloading import GraphDataLoader
+
+from plaid.containers.dataset import Dataset as PlaidDataset
+from plaid.problem_definition import ProblemDefinition
+
+from data import *
+from final_features import *
+from model import MeshGraphNet
+
+
+class PlaidTransientDataset(Dataset):
+    def __init__(self, plaid_dataset, plaid_problem, split="train", fields=("U_x","U_y")):
+        self.plaid_ds   = plaid_dataset
+        self.ids        = plaid_problem.get_split(split)
+        sample0         = self.plaid_ds[self.ids[0]]
+        self.time_steps = sample0.get_all_mesh_times()
+        self.n_steps    = len(self.time_steps) - 1
+        self.fields     = fields
+
+        # Precompute mesh-level features
+        self.mesh_list = []
+        for sid in self.ids:
+            sample   = self.plaid_ds[sid]
+            pos = torch.tensor(sample.get_nodes(), dtype=torch.float32)
+            cells = sample.get_elements()["TRI_3"]
+            edge_index = tri_cells_to_edges(torch.tensor(cells, dtype=torch.long)).t().contiguous()
+
+            _, sdf = get_distances_to_borders(sample.get_nodes(), cells)
+            sdf       = torch.tensor(sdf, dtype=torch.float32)
+            sdf_sine  = sinusoidal_embedding(sdf, num_basis=4, max_coord=4, spacing=0.001)
+            angles    = angles_to_planes(pos)
+            sph       = torch.cat([
+                spherical_harmonics(angles[:,i], l_max=4)[:,1:]
+                for i in range(4)
+            ], dim=1)
+
+            src, dst  = edge_index
+            disp      = pos[src] - pos[dst]
+            sqd       = torch.exp(-0.5 * (disp**2).sum(1,keepdim=True)/(5.757861066563731*10))
+            edge_attr = torch.cat([sqd, disp], dim=-1)
+
+            self.mesh_list.append({
+                'pos': pos,
+                'edge_index': edge_index,
+                'edge_attr': edge_attr,
+                'sdf': sdf,
+                'sdf_sine': sdf_sine,
+                'sph': sph,
+            })
+
+    def __len__(self):
+        return len(self.ids) * self.n_steps
+
+    def __getitem__(self, idx):
+        sim_idx = idx // self.n_steps
+        t       = idx %  self.n_steps
+        sid     = self.ids[sim_idx]
+        sample  = self.plaid_ds[sid]
+        mesh    = self.mesh_list[sim_idx]
+
+        # Velocities at t and t+1
+        ux_t  = torch.tensor(sample.get_field(self.fields[0], time=self.time_steps[t]),   dtype=torch.float32)
+        uy_t  = torch.tensor(sample.get_field(self.fields[1], time=self.time_steps[t]),   dtype=torch.float32)
+        ux_tp = torch.tensor(sample.get_field(self.fields[0], time=self.time_steps[t+1]), dtype=torch.float32)
+        uy_tp = torch.tensor(sample.get_field(self.fields[1], time=self.time_steps[t+1]), dtype=torch.float32)
+
+        u_t  = torch.stack([ux_t,  uy_t ], dim=-1)
+        u_tp = torch.stack([ux_tp, uy_tp], dim=-1)
+
+        # Input features: [u_t, pos, sdf, sph]
[…]
