+++ benchmarks/README.md
@@ -25,7 +25,7 @@ We believe this additional work fully addresses points W1 and W2 raised in the i
 | Dataset           | MGN | MMGP | Vi-Transf. | Augur | FNO | MARIO |
 |-------------------|-----|------|------------|-------|-------|-------|
 | `Tensile2d`       | 0.0673  |  **0.0026**  |   0.0116     |  0.0154   |  0.0123  |  *0.0038*  |
-| `2D_MultiScHypEl` | 0.0437  |  ❌  |   *0.0325*     |  **0.0232**   |   0.0439  |  0.0573  |
+| `2D_MultiScHypEl` | 0.0437  |  ❌  |   0.0325     |  **0.0232**   |   *0.0302*  |  0.0573  |
 | `2D_ElPlDynamics` | 0.1202  |  ❌  |   *0.0263*     |  0.0346    |  **0.0215**  |  0.0742  |
 | `Rotor37`         | 0.0074  |  **0.0014**  |   0.0029     |  0.0033   |   0.0313  |  *0.0017*  |
 | `2D_profile`      | 0.0593  |  0.0365  |   *0.0319*     |  0.0425   |  0.0972  |  **0.0307**  |
