### 2025-08-05 7b3c865
vi-transformer benchmark code
