+++ benchmarks/Vi-Transf/src/models/vits/flatformer_cls_less.py
@@ -1,0 +1,171 @@
+from ..model import BaseModel
+import torch
+import torch.nn as nn
+from torch import Tensor
+from torch.nn import TransformerEncoder, TransformerEncoderLayer
+from typing import List, Union, Callable, Optional
+from torch_geometric.data import Data, Batch
+from .tokenization.tokenizer import Tokenizer
+from ..blocks import MLP
+from einops import rearrange
+import torch.nn.functional as F
+import random
+
+# B: Batch size
+# T: Number of tokens
+# L: Latent dimension of the tokens
+# Sin: Number of input scalars
+# Fin: Number of input fields
+# Sout: Number of output scalars
+# Fout: Number of output fields
+# Nv: Number of vertices per subdomain
+
+class FlatFormerCLSLess(BaseModel):
+    def __init__(self,
+                n_vertices_per_subdomain: int,
+                n_head: int,
+                dim_ff: int,
+                num_layers: int,
+                tokenizer: Tokenizer,
+                input_field_dim: int=None,
+                input_scalar_dim: int=None,
+                output_scalar_dim: int=None,
+                output_field_dim: int=None,
+                activation: Union[str, Callable[[Tensor], Tensor]]="relu",
+                norm_first: bool=True,
+                dropout: float=0.1,
+                latent_dim: Optional[int]=None,
+                **kwargs
+                ):
+
+        super().__init__()
+        self.device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
+        self.output_scalar_dim = output_scalar_dim
+        self.output_field_dim = output_field_dim
+        self.input_scalar_dim = input_scalar_dim
+        self.input_field_dim = input_field_dim
+
+        self.criterion = nn.MSELoss(reduction="none")
+
+        # Tokenizer
+        token_dim = n_vertices_per_subdomain * input_field_dim
+        self.tokenizer = tokenizer
+        self.n_head = n_head
+        
+        # Linear projection of the flattened tokens onto the latent space
+        self.encoder = nn.Linear(token_dim, latent_dim - input_scalar_dim, bias=False)
+
+        # CLS token for the transformer encoder
+        self.cls_token = nn.Parameter(torch.zeros(1, 1, latent_dim), requires_grad=True) # 1 1 L
+        nn.init.xavier_uniform_(self.cls_token)
+
+        # Transformer encoder
+        encoder_layers  = TransformerEncoderLayer(d_model=latent_dim,
+                                                  nhead=n_head,
+                                                  dim_feedforward=dim_ff,
+                                                  dropout=dropout,
+                                                  batch_first=True,
+                                                  norm_first=norm_first,
+                                                  dtype=torch.float32,
+                                                  activation=activation)
+        self.transformer_encoder = TransformerEncoder(encoder_layers, num_layers)
+
+        # MLP decoder
+        self.decoder     = nn.Linear(latent_dim + input_scalar_dim, n_vertices_per_subdomain*(self.output_field_dim + self.output_scalar_dim), bias=False)
+        
+    def preprocess(self, pyg_dataset: List[Data], seed: Optional[int]=None, **kwargs) -> List[Data]:
+        if seed is None: seed = random.randint(0, 2**32 - 1)
+        print(f"Using seed: {seed} for data preprocessing")
+        dataset = self.tokenizer.preprocess(pyg_dataset, seed=seed)
+        
+        return dataset
+
+    def forward(self, data_batch: Data | Batch):
+        tokens, src_key_padding_mask = self.tokenizer(data_batch) # B T (Nv * F)
+        if self.input_scalar_dim > 0:
[…]
