+++ benchmarks/Vi-Transf/src/data/loader/bridges/no_bridge.py
@@ -1,0 +1,5 @@
+from plaid.containers.sample import Sample
+from plaid.problem_definition import ProblemDefinition
+
+def no_bridge(sample: Sample, sample_id: int, problem_definition: ProblemDefinition):
+    return sample
