### 2025-08-05 1554cdf
meshgraphnet
