### 2025-08-05 1e9d42a
update readmes with dependencies
