+++ benchmarks/FNO/README.md
@@ -1,0 +1,7 @@
+### List of dependencies:
+
+- [PLAID=0.1.6](https://github.com/PLAID-lib/plaid)
+- [PhysicsNemo=1.1.1](https://github.com/NVIDIA/physicsnemo)
+- Muscat:
+  - 2D_ElPlDynamics: [Muscat=87189dab](https://git.cloud.safran/safransa/safrantech/muscat/muscat/-/tree/87189dab624f058706786ce16dcfd7736ddddef5)
+  - Other datasets: [Muscat=2.4.1](https://gitlab.com/drti/muscat)
