+++ benchmarks/MGN/README.md
@@ -83,6 +83,5 @@ Install the required Python libraries before running the scripts:
 - torch=2.6.0
 - dgl=2.4.0+cu124
 - muscat=2.4.1
-- plaid=0.1
-- numpy=1.26.4
+- plaid=0.1.6
 - scikit-learn=1.6.1
