### 2025-08-05 e385901
add MARIO links and remove code status from READEM
