### 2025-08-05 1312dfe
benchmarks: improve README
