+++ README.md
@@ -1,0 +1,165 @@
+<div align="center">
+<img src="https://plaid-lib.github.io/assets/images/PLAID-large-logo.png" width="300">
+</div>
+
+| | |
+| --- | --- |
+| Testing | [![CI Status](https://github.com/PLAID-lib/plaid/actions/workflows/testing.yml/badge.svg)](https://github.com/PLAID-lib/plaid/actions/workflows/testing.yml) [![Documentation Status](https://readthedocs.org/projects/plaid-lib/badge/?version=latest)](https://plaid-lib.readthedocs.io/en/latest/?badge=latest) [![Coverage](https://codecov.io/gh/plaid-lib/plaid/branch/main/graph/badge.svg)](https://app.codecov.io/gh/plaid-lib/plaid/tree/main?search=&displayType=list) ![Last Commit](https://img.shields.io/github/last-commit/PLAID-lib/plaid/main) |
+| Package | [![Conda Latest Release](https://anaconda.org/conda-forge/plaid/badges/version.svg)](https://anaconda.org/conda-forge/plaid) [![Conda Downloads](https://img.shields.io/conda/dn/conda-forge/plaid.svg?label=Conda%20downloads)](https://anaconda.org/conda-forge/plaid) [![PyPI Latest Release](https://img.shields.io/pypi/v/pyplaid.svg)](https://pypi.org/project/pyplaid/) [![PyPI Downloads](https://static.pepy.tech/badge/pyplaid)](https://pepy.tech/projects/pyplaid) ![Platform](https://img.shields.io/badge/platform-any-blue) ![Python Version](https://img.shields.io/pypi/pyversions/pyplaid)  |
+| Other | [![License - BSD 3-Clause](https://anaconda.org/conda-forge/plaid/badges/license.svg)](https://github.com/PLAID-lib/plaid/blob/main/LICENSE.txt) ![GitHub stars](https://img.shields.io/github/stars/PLAID-lib/plaid?style=social) [![JOSS status](https://joss.theoj.org/papers/26b2e13a9fc8e012cc997ca28a7b565e/status.svg)](https://joss.theoj.org/papers/26b2e13a9fc8e012cc997ca28a7b565e) |
+
+
+<!-- [![PyPI Downloads](https://img.shields.io/pypi/dm/pyplaid.svg?label=PyPI%20downloads)](https://pypi.org/project/pyplaid/) -->
+
+
+# Physics Learning AI Datamodel (PLAID)
+
+</div>
+
+
+- [Physics Learning AI Datamodel (PLAID)](#physics-learning-ai-datamodel-plaid)
+  - [1. Description](#1-description)
+  - [2. Getting started](#2-getting-started)
+  - [3. Call for Contributions](#3-call-for-contributions)
+  - [4. Documentation](#4-documentation)
+
+
+## 1. Description
+
+This library proposes an implementation for a datamodel tailored for AI and ML learning of physics problems.
+It has been developped at SafranTech, the research center of [Safran group](https://www.safran-group.com/).
+
+- **Documentation:** https://plaid-lib.readthedocs.io/
+- **Source code:** https://github.com/PLAID-lib/plaid
+- **Contributing:** https://github.com/PLAID-lib/plaid/blob/main/CONTRIBUTING.md
+- **License:** https://github.com/PLAID-lib/plaid/blob/main/LICENSE.txt
+- **Bug reports:** https://github.com/PLAID-lib/plaid/issues
+- **Report a security vulnerability:** https://github.com/PLAID-lib/plaid/issues
+
+
+## 2. Getting started
+
+
+### 2.1 Using the library
+
+To use the library, the simplest way is to install it as follows:
+
+```bash
[…]
