### 2025-08-05 ac4f814
🎉 Add benchmark's method code source (#121)
---------

Co-authored-by: Fabien Casenave <fabien.casenave@safrangroup.com>
Co-authored-by: williampiat3 <32038421+williampiat3@users.noreply.github.com>
Co-authored-by: Xuan Minh Vuong NGUYEN <xuan-minh-vuong.nguyen@safrangroup.com>
Co-authored-by: DaGrapix <anthony.kalaydjian@outlook.com>
