+++ benchmarks/MMGP/2D_profile/launch_morphings.py
@@ -1,0 +1,5 @@
+import os
+
+for CASE in range(0, 399):
+    os.system(f"python morphing_script.py {CASE}")
+
