+++ benchmarks/Vi-Transf/configs/vkils59.yaml
@@ -12,8 +12,8 @@ bridge:
 loader:
   _target_: src.data.loader.hub_loader.HubLoader
   bridge: ${bridge.function}
-  dataset_name: PLAID-datasets/VKILS59
-  cache_dir: cached_datasets/VKILS59
+  dataset_name: PLAID-datasets/VKI-LS59
+  cache_dir: cached_datasets/VKI-LS59
   task_split:
   - train
   - test
