### 2025-08-06 0535c90
fix(benchmarks/README.md) improve phrasing
