### 2025-08-06 fc63e64
fix(benchmarks/README.md) improve phrasing
