### 2025-08-06 f5ca156
📄 fix(benchmarks/README.md) improve phrasing (#123)
* fix(benchmarks/README.md) improve phrasing

---------

Co-authored-by: Fabien Casenave <fabien.casenave@safrangroup.com>
