### 2025-08-06 aa65277
update mgn files
