### 2025-08-06 79c3818
♻️ feat(benchmarks/MGN) remove score computation using reference
