### 2025-08-07 ae511b7
fix(*/*) add Optional typing when needed
