### 2025-08-07 69499d0
(constants.py/sample.py) add constant CGNS_FIELD_LOCATIONS + improve docstrings
