### 2025-08-07 7a9983a
(test_dataset/conftest) factorize fixtures in conftest + reorganize tests accordingly
