+++ tests/containers/test_dataset.py
@@ -21,21 +21,6 @@ from plaid.utils.base import ShapeError
 
 
 @pytest.fixture()
-def nb_scalars():
-    return 5
-
-
-@pytest.fixture()
-def tabular(nb_samples, nb_scalars):
-    return np.random.randn(nb_samples, nb_scalars)
-
-
-@pytest.fixture()
-def scalar_names(nb_scalars):
-    return [f"test_scalar_{np.random.randint(1e8, 1e9)}" for _ in range(nb_scalars)]
-
-
-@pytest.fixture()
 def sample(zone_name, base_name):
     sample = Sample()
     sample.init_base(3, 3, base_name)
@@ -49,16 +34,6 @@ def sample(zone_name, base_name):
         base_name,
     )
     return sample
-
-
-@pytest.fixture
-def empty_sample():
-    return Sample()
-
-
-@pytest.fixture()
-def empty_dataset():
-    return Dataset()
 
 
 @pytest.fixture()
@@ -200,30 +175,30 @@ class Test_Dataset:
         with pytest.raises(TypeError):
             dataset.add_sample(1)
 
-    def test_add_samples_empty(self, empty_dataset):
+    def test_add_samples_empty(self, dataset):
         with pytest.raises(ValueError):
-            empty_dataset.add_samples([])
+            dataset.add_samples([])
         with pytest.raises(ValueError):
-            empty_dataset.add_samples([], 1)
+            dataset.add_samples([], 1)
 
-    def test_add_samples_empty_with_ids(self, empty_dataset, sample):
+    def test_add_samples_empty_with_ids(self, dataset, sample):
         with pytest.raises(ValueError):
-            empty_dataset.add_samples([sample], [1, 2, 3])
+            dataset.add_samples([sample], [1, 2, 3])
 
-    def test_add_samples_bad_number_ids_inf(self, empty_dataset, sample):
+    def test_add_samples_bad_number_ids_inf(self, dataset, sample):
         with pytest.raises(ValueError):
             samples = [sample, sample, sample]
-            empty_dataset.add_samples(samples, [1, 2])
+            dataset.add_samples(samples, [1, 2])
 
-    def test_add_samples_bad_number_ids_supp(self, empty_dataset, sample):
+    def test_add_samples_bad_number_ids_supp(self, dataset, sample):
         with pytest.raises(ValueError):
             samples = [sample, sample, sample]
-            empty_dataset.add_samples(samples, [1, 2, 3, 4])
+            dataset.add_samples(samples, [1, 2, 3, 4])
 
-    def test_add_samples_with_same_ids(self, empty_dataset, sample):
+    def test_add_samples_with_same_ids(self, dataset, sample):
         with pytest.raises(ValueError):
             samples = [sample, sample, sample]
-            empty_dataset.add_samples(samples, [1, 1, 1])
+            dataset.add_samples(samples, [1, 1, 1])
 
     def test_add_samples_with_ids_good(self, dataset, sample):
         samples = [sample, sample, sample]
@@ -321,82 +296,145 @@ class Test_Dataset:
         dataset_with_samples.get_field_names(np.random.randint(2, nb_samples, size=2))
 
     # -------------------------------------------------------------------------#
-    def test_add_tabular_scalars(self, dataset, tabular, scalar_names, nb_samples):
-        dataset.add_tabular_scalars(tabular, scalar_names)
+    def test_add_tabular_scalars(
+        self, dataset, scalar_tabular, scalar_names, nb_samples
+    ):
+        dataset.add_tabular_scalars(scalar_tabular, scalar_names)
         assert len(dataset) == nb_samples
 
[…]
