### 2025-08-07 4596a9f
(sklearn wrapper) minor ruff reformating
