### 2025-08-07 7624d99
(dataset) clean methods on tabular + fix a bug
