+++ src/plaid/containers/dataset.py
@@ -409,7 +409,7 @@ class Dataset(object):
         """Add tabular scalar data to the summary.
 
         Args:
-            tabular (np.ndarray): A 2D NumPy array containing tabular scalar data.
+            tabular (np.ndarray): A 2D NumPy array containing tabular scalar data, with shape (n_samples, n_scalars)
             names (list[str], optional): A list of column names for the tabular data. Defaults to None.
 
         Raises:
@@ -419,27 +419,23 @@ class Dataset(object):
         Note:
             If no names are provided, it will automatically create names based on the pattern 'X{number}'
         """
-        nb_samples = len(tabular)
-
         if tabular.ndim != 2:
             raise ShapeError(f"{tabular.ndim=}!=2, should be == 2")
+
+        nb_samples = tabular.shape[0]
+        nb_scalars = tabular.shape[1]
+
         if names is None:
-            names = [f"X{i}" for i in range(tabular.shape[1])]
-        if tabular.shape[1] != len(names):
+            names = [f"X{i}" for i in range(nb_scalars)]
+        if nb_scalars != len(names):
             raise ShapeError(
-                f"tabular should have as many columns as there are names, but {tabular.shape[1]=} and {len(names)=}"
+                f"tabular should have as many columns as there are names, but {nb_scalars=} and {len(names)=}"
             )
 
-        # ---# For efficiency, first add values to storage
-        name_to_ids = {}
-        for col, name in zip(tabular.T, names):
-            name_to_ids[name] = col
-
-        # ---# Then add data in sample
         for i_samp in range(nb_samples):
             sample = Sample()
-            for name in names:
-                sample.add_scalar(name, name_to_ids[name][i_samp])
+            for i_col, name in enumerate(names):
+                sample.add_scalar(name, tabular[i_samp, i_col])
             self.add_sample(sample)
 
     def get_scalars_to_tabular(
@@ -505,28 +501,26 @@ class Dataset(object):
         Note:
             If no names are provided, it will automatically create names based on the pattern 'X{number}'
         """
-        nb_samples = len(tabular)
-
         if tabular.ndim != 4:
             raise ShapeError(f"{tabular.ndim=}!=4, should be == 4")
+
+        nb_samples = tabular.shape[0]
+        # max_n_timestamps = tabular.shape[1]
+        nb_features = tabular.shape[2]
+        assert tabular.shape[3] == 2
+
         if names is None:
-            names = [f"X{i}" for i in range(tabular.shape[2])]
-        if tabular.shape[2] != len(names):
+            names = [f"X{i}" for i in range(nb_features)]
+        if nb_features != len(names):
             raise ShapeError(
-                f"tabular's 3rd dimension should have same size as there are names, but {tabular.shape[2]=} and {len(names)=}"
+                f"tabular's 3rd dimension should have same size as there are names, but {nb_features=} and {len(names)=}"
             )
 
-        # ---# For efficiency, first add values to storage
-        name_to_ids = {}
-        for i_col, name in enumerate(names):
-            name_to_ids[name] = tabular[:, :, i_col]
-
-        # ---# Then add data in sample
         for i_samp in range(nb_samples):
             sample = Sample()
-            for i_name, name in enumerate(names):
-                timestamps = name_to_ids[name][i_samp, :, 0]
-                values = name_to_ids[name][i_samp, :, 1]
+            for i_col, name in enumerate(names):
+                timestamps = tabular[i_samp, :, i_col, 0]
+                values = tabular[i_samp, :, i_col, 1]
                 # detect first decreasing timestamp
                 diff_ts = timestamps[1:] - timestamps[:-1]
                 decreasing_ts_ids = (diff_ts <= 0).nonzero()[0]
@@ -619,7 +613,7 @@ class Dataset(object):
[…]
