### 2025-08-07 90102af
(constants) fix order of objects + improve docstring
