+++ src/plaid/constants.py
@@ -1,4 +1,46 @@
-# Voir https://cgns.org/standard/SIDS/grid.html#flow-solution-structure-definition-flowsolution-t
+# -*- coding: utf-8 -*-
+#
+# This file is subject to the terms and conditions defined in
+# file 'LICENSE.txt', which is part of this source code package.
+#
+#
+"""This module defines common constants used throughout the PLAID library.
+
+It includes:
+- AUTHORIZED_TASKS: List of supported task types (e.g., regression, classification).
+- AUTHORIZED_INFO_KEYS: Dictionary specifying allowed metadata keys for various information sections.
+- CGNS_FIELD_LOCATIONS: List of valid field locations as defined by the CGNS standard.
+- CGNS_ELEMENT_NAMES: List of CGNS element names representing different mesh element types.
+
+These constants help standardize metadata, task types, and mesh element references across the PLAID codebase.
+"""
+
+AUTHORIZED_TASKS = ["regression", "classification"]
+
+AUTHORIZED_INFO_KEYS = {
+    "legal": ["owner", "license"],
+    "data_production": [
+        "owner",
+        "license",
+        "type",
+        "physics",
+        "simulator",
+        "hardware",
+        "computation_duration",
+        "script",
+        "contact",
+        "location",
+    ],
+    "data_description": [
+        "number_of_samples",
+        "number_of_splits",
+        "DOE",
+        "inputs",
+        "outputs",
+    ],
+}
+
+# See https://cgns.org/standard/SIDS/grid.html#flow-solution-structure-definition-flowsolution-t
 CGNS_FIELD_LOCATIONS = [
     "Vertex",
     "CellCenter",
@@ -8,23 +50,6 @@ CGNS_FIELD_LOCATIONS = [
     "KFaceCenter",
     "EdgeCenter",
 ]
-"""Common constants used across the PLAID library.
-
-AUTHORIZED_TASKS: Tuple of strings representing the types of tasks supported by PLAID.
-
-CGNS_ELEMENT_NAMES: List of strings representing the names of CGNS elements.
-
-AUTHORIZED_INFO_KEYS: Dictionary defining the keys allowed in different sections of metadata.
-"""
-
-# -*- coding: utf-8 -*-
-#
-# This file is subject to the terms and conditions defined in
-# file 'LICENSE.txt', which is part of this source code package.
-#
-#
-
-AUTHORIZED_TASKS = ["regression", "classification"]
 
 CGNS_ELEMENT_NAMES = [
     "ElementTypeNull",
@@ -68,26 +93,3 @@ CGNS_ELEMENT_NAMES = [
     "HEXA_56",
     "HEXA_64",
 ]
-
-AUTHORIZED_INFO_KEYS = {
-    "legal": ["owner", "license"],
-    "data_production": [
-        "owner",
-        "license",
-        "type",
-        "physics",
-        "simulator",
-        "hardware",
-        "computation_duration",
-        "script",
-        "contact",
-        "location",
-    ],
-    "data_description": [
-        "number_of_samples",
-        "number_of_splits",
-        "DOE",
-        "inputs",
-        "outputs",
-    ],
-}
