### 2025-08-08 623aa64
📄 add constant CGNS_FIELD_LOCATIONS and improve doctrings accordingly (#128)
* (constants.py/sample.py) add constant CGNS_FIELD_LOCATIONS + improve docstrings

* (constants) fix order of objects + improve docstring
