+++ src/plaid/containers/sample.py
@@ -21,6 +21,7 @@ import copy
 import glob
 import logging
 import os
+import shutil
 from typing import Optional, Union
 
 import CGNS.MAP as CGM
@@ -1792,17 +1793,22 @@ class Sample(BaseModel):
             return self.get_nodes(**feature_identifier_)
 
     # -------------------------------------------------------------------------#
-    def save(self, dir_path: str) -> None:
+    def save(self, dir_path: str, overwrite: Optional[bool] = None) -> None:
         """Save the Sample in directory `dir_path`.
 
         Args:
             dir_path (str): relative or absolute directory path.
+            overwrite (str, optional): target directory overwritten if True.
         """
-        if os.path.isdir(dir_path):
-            if len(glob.glob(os.path.join(dir_path, "*"))):
+        if os.path.exists(dir_path) and os.path.isdir(dir_path):
+            if len(glob.glob(os.path.join(dir_path, "*"))) and not overwrite:
                 raise ValueError(
-                    f"directory {dir_path} already exists and is not empty"
+                    f"directory {dir_path} already exists and is not empty. Set `overwrite` to True if needed."
                 )
+            else:
+                shutil.rmtree(dir_path)
+                logger.warning(f"{dir_path} directory has been overwritten.")
+
         else:
             os.makedirs(dir_path)
 
