### 2025-08-09 5a8ac15
feat(docs) add version and rtd full git depth
