### 2025-08-09 87689a2
feat(docs) add version to docs and rtd full git depth
