+++ docs/conf.py
@@ -233,31 +233,31 @@ def setup(sphinx):
 
 # -----------------------------------------------------------------------------#
 
+def get_git_version():
+    try:
+        # Try to get exact tag first
+        version = subprocess.check_output(
+            ["git", "describe", "--tags", "--exact-match"],
+            stderr=subprocess.STDOUT,
+        ).decode().strip()
+        return version
+    except subprocess.CalledProcessError:
+        # No tag, get branch name
+        try:
+            branch = subprocess.check_output(
+                ["git", "rev-parse", "--abbrev-ref", "HEAD"]
+            ).decode().strip()
+        except Exception:
+            branch = "unknown-branch"
+        # Get short commit hash
+        try:
+            commit_hash = subprocess.check_output(
+                ["git", "rev-parse", "--short", "HEAD"]
+            ).decode().strip()
+        except Exception:
+            commit_hash = "unknown-hash"
 
-# def get_git_version():
-#     try:
-#         # Try to get exact tag first
-#         version = subprocess.check_output(
-#             ["git", "describe", "--tags", "--exact-match"],
-#             stderr=subprocess.STDOUT,
-#         ).decode().strip()
-#         return version
-#     except subprocess.CalledProcessError:
-#         # No tag, get branch name
-#         try:
-#             branch = subprocess.check_output(
-#                 ["git", "rev-parse", "--abbrev-ref", "HEAD"]
-#             ).decode().strip()
-#         except Exception:
-#             branch = "unknown-branch"
-#         # Get short commit hash
-#         try:
-#             commit_hash = subprocess.check_output(
-#                 ["git", "rev-parse", "--short", "HEAD"]
-#             ).decode().strip()
-#         except Exception:
-#             commit_hash = "unknown-hash"
+        return f"{branch}-{commit_hash}"
 
-#         return f"{branch}-{commit_hash}"
-
-# version = get_git_version()
+release = get_git_version()
+version = release
