+++ .readthedocs.yml
@@ -22,9 +22,3 @@ python:
   install:
     - method: pip
       path: .
-
-versions:
-  only:
-    - latest
-    - 0.1.6
-    - 0.1.5
