### 2025-08-09 11f5912
remove mmgp_pipelines
