+++ src/plaid/pipelines/mmgp_blocks.py
@@ -1,910 +1,0 @@
-"""Custom blocks for applying MMGP.
-
-Includes:
-- MMGPPreparer.
-- MMGPTransformer.
-"""
-
-import copy
-import time
-from typing import Callable, Optional, Tuple
-
-import Muscat.Containers.ElementsDescription as ED
-import Muscat.Containers.Filters.FilterObjects as FilterObjects
-import Muscat.Containers.MeshInspectionTools as MIT
-import Muscat.Containers.MeshModificationTools as MMT
-import networkx
-import numpy as np
-from Muscat.Bridges.CGNSBridge import CGNSToMesh, MeshToCGNS
-from Muscat.Containers import MeshFieldOperations as MFO
-from Muscat.Containers.Mesh import Mesh
-from Muscat.FE import FETools as FT
-from Muscat.FE.Fields import FEField as FF
-from scipy import sparse
-from sklearn.base import (
-    BaseEstimator,
-    TransformerMixin,
-)
-from sklearn.utils.validation import check_is_fitted
-
-from plaid.containers.dataset import Dataset
-from plaid.containers.utils import (
-    get_feature_type_and_details_from_identifier,
-)
-
-
-class MMGPPreparer(TransformerMixin, BaseEstimator):
-    """Preparer for MMGP.
-
-    Transforms
-
-    Args:
-        common_mesh_id: ...
-        morphing: ...
-    """
-
-    def __init__(
-        self, common_mesh_id: Optional[int] = None, morphing: Optional[Callable] = None
-    ):
-        self.common_mesh_id = common_mesh_id
-        self.morphing = morphing
-
-    def fit(self, dataset: Dataset, _y=None):
-        """Fits the underlying scikit-learn transformer on selected input features.
-
-        Args:
-            dataset: A `Dataset` object containing the features to transform.
-            _y: Ignored.
-
-        Returns:
-            self: The fitted transformer.
-        """
-        self.common_mesh_id_ = 0 if self.common_mesh_id is None else self.common_mesh_id
-
-        self.morphed_common_mesh_ = self.morphing(
-            CGNSToMesh(dataset[self.common_mesh_id_].get_mesh())
-        )
-        self.morphed_common_tree_ = MeshToCGNS(
-            self.morphed_common_mesh_, exportOriginalIDs=False, tagsAsFields=False
-        )
-
-        return self
-
-    def transform(self, dataset: Dataset):
-        """Applies the fitted transformer to the selected input features.
-
-        Args:
-            dataset: A `Dataset` object to transform.
-
-        Returns:
-            Dataset: Transformed features wrapped as a new `Dataset`.
-        """
-        check_is_fitted(self, "morphed_common_mesh_")
-
-        # n_dim = self.morphed_common_mesh_.GetPointsDimensionality()
-        # coord_names = ["coords_X", "coords_Y", "coords_Z"][:n_dim]
-
-        samples = []
-        for sample in dataset:
-            # morph the mesh of the current sample
-            mesh = CGNSToMesh(sample.get_mesh())
-            morphed_mesh = self.morphing(mesh)
-
-            # compute the FE interpolation operator
-            proj_operator = compute_FE_projection_operator(
-                morphed_mesh, self.morphed_common_mesh_
-            )
-            inv_proj_operator = compute_FE_projection_operator(
-                self.morphed_common_mesh_, morphed_mesh
-            )
-
-            # # initialize morph_proj_sample and add the common mesh as geometrical support of morph_proj_sample
-            # morph_proj_sample = Sample()
-            # morph_proj_sample.add_tree(copy.deepcopy(self.morphed_common_tree_))
-
-            # # update all features (except nodes and fields) of morph_proj_sample from the ones of current sample
[…]
