### 2025-08-09 96e2513
feat(docs) add version to docs and rtd full git depth
