+++ docs/source/contributing.md
@@ -6,7 +6,7 @@ Thank you for your interest in contributing to PLAID Project! We welcome contrib
 
 Everything you need to know if you want to contribute.
 
-- [Contributing to PLAID](#contributing-to-plaid)
+- [Contributing to PLAID](#how-to-contribute)
   - [Summary](#summary)
   - [1. Before contributing](#1-before-contributing)
   - [2. Reporting Issues](#2-reporting-issues)
