+++ docs/source/citation.md
@@ -1,0 +1,24 @@
+# How to Cite
+
+If you have used PLAID in your how work and fould it useful, please consider citing the following papers.
+
+## JOSS paper
+
+Currently under review:
+
+[![JOSS status](https://joss.theoj.org/papers/26b2e13a9fc8e012cc997ca28a7b565e/status.svg)](https://joss.theoj.org/papers/26b2e13a9fc8e012cc997ca28a7b565e)
+
+
+## PLAID datasets
+
+```bibtex
+@misc{casenave2025plaid,
+      title={{Physics-Learning AI Datamodel (PLAID) datasets: a collection of physics simulations for machine learning}},
+      author={Casenave, F. and Roynard, X. and Staber, B. and Piat, W. and Bucci, M. A. and Akkari, N. and Kabalan, A. and Nguyen, X. M. V. and Saverio, L. and Carpintero Perez, R. and Kalaydjian, A. and Fouch\'{e}, S. and Gonon, T. and Najjar, G. and Menier, E. and Nastorg, M. and Catalani, G. and Rey, C.},
+      year={2025},
+      archivePrefix={arXiv},
+      primaryClass={cs.LG},
+      url={https://arxiv.org/abs/2505.02974},
+}
+```
+
