### 2025-08-11 452b5ff
🐛 fix(.readthedocs.yml) remove git:depth option (#134)
