### 2025-08-11 001848f
Simplifying the wrapped regressor
