+++ tests/pipelines/test_sklearn_block_wrappers.py
@@ -2,8 +2,8 @@ import numpy as np
 import pytest
 
 from plaid.pipelines.sklearn_block_wrappers import (
-    WrappedPlaidSklearnRegressor,
-    WrappedPlaidSklearnTransformer,
+    WrappedSklearnRegressor,
+    WrappedSklearnTransformer,
     get_2Darray_from_homogeneous_identifiers,
 )
 
@@ -72,21 +72,21 @@ def test_get_2Darray_from_homogeneous_identifiers_nodes(
     assert X.shape == (4, 10)
 
 
-class Test_WrappedPlaidSklearnTransformer:
+class Test_WrappedSklearnTransformer:
     def test___init__(
         self,
         sklearn_scaler,
         dataset_with_samples_scalar1_feat_ids,
         dataset_with_samples_scalar2_feat_ids,
     ):
-        WrappedPlaidSklearnTransformer()
-        WrappedPlaidSklearnTransformer(sklearn_block=sklearn_scaler)
+        WrappedSklearnTransformer()
+        WrappedSklearnTransformer(sklearn_block=sklearn_scaler)
         # in_feat_ids = [{"type": "scalar", "name": "test_scalar"}]
-        WrappedPlaidSklearnTransformer(
+        WrappedSklearnTransformer(
             sklearn_block=sklearn_scaler,
             in_features_identifiers=dataset_with_samples_scalar1_feat_ids,
         )
-        WrappedPlaidSklearnTransformer(
+        WrappedSklearnTransformer(
             sklearn_block=sklearn_scaler,
             in_features_identifiers=dataset_with_samples_scalar1_feat_ids,
             out_features_identifiers=dataset_with_samples_scalar2_feat_ids,
@@ -134,34 +134,29 @@ class Test_WrappedPlaidSklearnTransformer:
         assert not np.allclose(in_features, tranformed_in_features)
 
 
-class Test_WrappedPlaidSklearnRegressor:
+class Test_WrappedSklearnRegressor:
     def test___init__(
         self,
         sklearn_multioutput_gp_regressor,
         dataset_with_samples_scalar1_feat_ids,
         dataset_with_samples_scalar2_feat_ids,
     ):
-        WrappedPlaidSklearnRegressor()
-        WrappedPlaidSklearnRegressor(sklearn_block=sklearn_multioutput_gp_regressor)
-        WrappedPlaidSklearnRegressor(
+        WrappedSklearnRegressor()
+        WrappedSklearnRegressor(sklearn_block=sklearn_multioutput_gp_regressor)
+        WrappedSklearnRegressor(
             sklearn_block=sklearn_multioutput_gp_regressor,
             in_features_identifiers=dataset_with_samples_scalar1_feat_ids,
         )
-        WrappedPlaidSklearnRegressor(
+        WrappedSklearnRegressor(
             sklearn_block=sklearn_multioutput_gp_regressor,
             in_features_identifiers=dataset_with_samples_scalar1_feat_ids,
             out_features_identifiers=dataset_with_samples_scalar2_feat_ids,
         )
 
-        def length_scale_init(X):
-            return np.ones(X.shape[1])
-
-        dynamics_params_factory = {"estimator__kernel__length_scale": length_scale_init}
-        WrappedPlaidSklearnRegressor(
+        WrappedSklearnRegressor(
             sklearn_block=sklearn_multioutput_gp_regressor,
             in_features_identifiers=dataset_with_samples_scalar1_feat_ids,
             out_features_identifiers=dataset_with_samples_scalar2_feat_ids,
-            dynamics_params_factory=dynamics_params_factory,
         )
 
[…]
