### 2025-08-12 77076a0
updates: rename block and correct block sk reg simplification
