+++ examples/pipelines/pipeline.py
@@ -2,20 +2,19 @@
 # coding: utf-8
 
 # # Pipeline Examples
+# 
+# This notebook demonstrates the end-to-end process of building a machine learning pipeline using PLAID datasets and PLAID’s scikit-learn-compatible blocks.
 
-# ## PCA-GP on Field Data
+# ## PCA-GP for `mach` field prediction of `VKI-LS59` dataset
 # 
-# This notebook demonstrates the end-to-end process of building a machine learning pipeline for predicting high-dimensional flow fields from scalar input parameters using the PLAID dataset. The pipeline leverages PLAID’s scikit-learn-compatible blocks and standard dataset structures to enable modular, reusable workflows.
+# Key steps covered:
 # 
-# Key steps covered in this notebook:
-# 
-# - **Loading and preparing the PLAID dataset** using Hugging Face integration and PLAID’s dataset classes  
-# - **Standardizing features** with PLAID-wrapped scikit-learn transformers for scalars and fields  
+# - **Loading the PLAID dataset** using Hugging Face integration and PLAID’s dataset classes  
+# - **Standardizing features** with PLAID-wrapped scikit-learn transformers for scalars  
 # - **Dimensionality reduction** of flow fields via Principal Component Analysis (PCA) to reduce output complexity  
 # - **Regression modeling** of PCA coefficients from scalar inputs using Gaussian Process regression  
 # - **Pipeline assembly** combining transformations and regressors into a single scikit-learn-compatible workflow  
 # - **Hyperparameter tuning** using Optuna and scikit-learn’s `GridSearchCV`
-# - **Model evaluation** using cross-validation and appropriate metrics  
 # - **Best practices** for working with PLAID datasets and pipelines in a reproducible and modular manner
 
 # ### 📦 Imports
@@ -49,14 +48,14 @@ from sklearn.multioutput import MultiOutputRegressor
 from sklearn.model_selection import KFold, GridSearchCV
 
 from plaid.bridges.huggingface_bridge import huggingface_dataset_to_plaid, huggingface_description_to_problem_definition
-from plaid.pipelines.sklearn_block_wrappers import WrappedPlaidSklearnTransformer, WrappedPlaidSklearnRegressor
-from plaid.pipelines.plaid_blocks import PlaidTransformedTargetRegressor, PlaidColumnTransformer
+from plaid.pipelines.sklearn_block_wrappers import WrappedSklearnTransformer, WrappedSklearnRegressor
+from plaid.pipelines.plaid_blocks import TransformedTargetRegressor, ColumnTransformer
 
 disable_progress_bar()
 n_processes = min(max(1, os.cpu_count()), 6)
 
 
-# ## 📥 Load Dataset
+# ### 📥 Load Dataset
 # 
 # We load the `VKI-LS59` dataset from Hugging Face and restrict ourselves to the first 24 samples of the training set.
 
@@ -75,7 +74,7 @@ dataset_train, _ = huggingface_dataset_to_plaid(hf_dataset, processes_number = n
 print(dataset_train)
 
 
-# ## ⚙️ Pipeline Configuration
+# ### ⚙️ Pipeline Configuration
 # 
 # For convenience, the `in_features_identifiers` and `out_features_identifiers` for each pipeline block are defined in a `.yml` file. Here's an example of how the configuration might look:
 
@@ -117,17 +116,17 @@ print("field names =", dataset_train.get_field_names())
 
 # We notive that only the 2 scalars and the field of interest are kept after restriction.
 
-# ### 1. Preprocessor
+# #### 1. Preprocessor
 # 
[…]
