### 2025-08-12 42e8a02
(path) replace all uses of os by pathlib
