+++ src/plaid/containers/sample.py
@@ -18,10 +18,9 @@ else:  # pragma: no cover
     Self = TypeVar("Self")
 
 import copy
-import glob
 import logging
-import os
 import shutil
+from pathlib import Path
 from typing import Optional, Union
 
 import CGNS.MAP as CGM
@@ -137,7 +136,7 @@ class Sample(BaseModel):
 
     def __init__(
         self,
-        directory_path: str = None,
+        directory_path: Union[str, Path] = None,
         mesh_base_name: str = "Base",
         mesh_zone_name: str = "Zone",
         meshes: dict[float, CGNSTree] = None,
@@ -149,7 +148,7 @@ class Sample(BaseModel):
         """Initialize an empty :class:`Sample <plaid.containers.sample.Sample>`.
 
         Args:
-            directory_path (str, optional): The path from which to load PLAID sample files.
+            directory_path (Union[str, Path], optional): The path from which to load PLAID sample files.
             mesh_base_name (str, optional): The base name for the mesh. Defaults to 'Base'.
             mesh_zone_name (str, optional): The zone name for the mesh. Defaults to 'Zone'.
             meshes (dict[float, CGNSTree], optional): A dictionary mapping time steps to CGNSTrees. Defaults to None.
@@ -189,7 +188,8 @@ class Sample(BaseModel):
         self._paths: dict[float, list[PathType]] = paths
 
         if directory_path is not None:
-            self.load(str(directory_path))
+            directory_path = Path(directory_path)
+            self.load(directory_path)
 
         self._defaults: dict = {
             "active_base": None,
@@ -504,9 +504,7 @@ class Sample(BaseModel):
         tree = copy.deepcopy(tree)
         for link in links:
             if not in_memory:
-                subtree, _, _ = CGM.load(
-                    os.path.join(link[0], link[1]), subtree=link[2]
-                )
+                subtree, _, _ = CGM.load(str(Path(link[0]) / link[1]), subtree=link[2])
             else:
                 linked_timestep = int(link[1].split(".cgns")[0].split("_")[1])
                 linked_timestamp = list(self._meshes.keys())[linked_timestep]
@@ -636,7 +634,7 @@ class Sample(BaseModel):
 
     def link_tree(
         self,
-        path_linked_sample: str,
+        path_linked_sample: Union[str, Path],
         linked_sample: Self,
         linked_time: float,
         time: float,
@@ -644,7 +642,7 @@ class Sample(BaseModel):
         """Link the geometrical features of the CGNS tree of the current sample at a given time, to the ones of another sample.
 
         Args:
-            path_linked_sample (str): The absolute path of the folder containing the linked CGNS
+            path_linked_sample (Union[str, Path]): The absolute path of the folder containing the linked CGNS
             linked_sample (Sample): The linked sample
             linked_time (float): The time step of the linked CGNS in the linked sample
             time (float): The time step the current sample to which the CGNS tree is linked.
@@ -657,6 +655,8 @@ class Sample(BaseModel):
 
         # https://pycgns.github.io/MAP/examples.html#save-with-links
         # When you load a file all the linked-to files are resolved to produce a full CGNS/Python tree with actual node data.
+
+        path_linked_sample = Path(path_linked_sample)
 
         if linked_time not in linked_sample._meshes:  # pragma: no cover
             raise KeyError(
@@ -723,7 +723,7 @@ class Sample(BaseModel):
                 ]
                 grid[2].append(zone_family)
 
-        def find_feature_roots(sample, time, Type_t):
+        def find_feature_roots(sample: Sample, time: float, Type_t: str):
             Types_t = CGU.getAllNodesByTypeSet(sample.get_mesh(time), Type_t)
[…]
