### 2025-08-12 2edbd72
(examples) replace all uses of os by pathlib
