### 2025-08-12 f818434
(docs) replace all uses of os by pathlib
