### 2025-08-12 a549e22
🎉 feat(sample.py) add option to overwrite folder in save (#130)
* feat(sample.py) add option to overwrite folder in save

* feat(sample.py) add option to overwrite folder in save

* (path) replace all uses of os by pathlib

* (examples) replace all uses of os by pathlib

* (docs) replace all uses of os by pathlib

---------

Co-authored-by: Fabien Casenave <fabien.casenave@safrangroup.com>
Co-authored-by: Xavier Roynard <xavier.roynard@safrangroup.com>
