### 2025-08-12 63feb47
(dataset) add tests for add_to_dir method
