### 2025-08-12 e843b86
feat(pipelines/typing) remove '=None' for various pipeline block parameters and factorize FeatureIdentifier typing
