+++ src/plaid/types/__init__.py
@@ -10,22 +10,27 @@
 from plaid.types.cgns_types import (
     CGNSNode,
     CGNSTree,
-    FeatureType,
-    FieldType,
     LinkType,
     NodeLabel,
     NodeName,
     NodeValue,
     PathType,
+)
+from plaid.types.common import Array, ArrayDType, IndexType
+from plaid.types.feature_types import (
+    FeatureIdentifier,
+    FeatureType,
+    FieldType,
     ScalarType,
     TimeSequenceType,
     TimeSeriesType,
 )
-from plaid.types.common import Array, ArrayDType, IndexType
+from plaid.types.sklearn_types import SklearnBlock
 
 __all__ = [
     "Array",
     "ArrayDType",
+    "IndexType",
     "CGNSNode",
     "CGNSTree",
     "LinkType",
@@ -38,5 +43,6 @@ __all__ = [
     "TimeSequenceType",
     "TimeSeriesType",
     "FeatureType",
-    "IndexType",
+    "FeatureIdentifier",
+    "SklearnBlock",
 ]
