### 2025-08-12 7418410
feat(plaid_blocks.py) simplify ColumnTransformer init
