### 2025-08-13 be96591
fix(pipelines) correct import error for typing Self for python 3.9
