### 2025-08-13 a1bcdd0
fix(docs) update config and huggingface_bridge for doc compatibility
