+++ docs/conf.py
@@ -14,7 +14,7 @@
 # -- Path setup --------------------------------------------------------------
 
 import datetime
-import sys
+import os, sys
 import subprocess
 from pathlib import Path
 
@@ -267,7 +267,6 @@ def get_git_info():
         ).strip()
 
         return f"{branch}-{commit}"
-
 
     except Exception:
         return "unknown"
