### 2025-08-13 e7aa81d
fix(huggingface_bridge.py) ruff formatting
