### 2025-08-13 3b40625
(problem_definition) define inputs and outputs with feature identifiers instead of names
