+++ src/plaid/problem_definition.py
@@ -8,9 +8,14 @@
 #
 
 # %% Imports
-
+# Standard library
+import csv
+import logging
 import sys
+from pathlib import Path
+from typing import Optional, Union
 
+# Typing support
 if sys.version_info >= (3, 11):
     from typing import Self
 else:  # pragma: no cover
@@ -18,15 +23,20 @@ else:  # pragma: no cover
 
     Self = TypeVar("Self")
 
-import csv
-import logging
-from pathlib import Path
-from typing import Union
-
+# Third party imports
 import yaml
 
-from plaid.constants import AUTHORIZED_TASKS
+# Local imports
+from plaid.constants import (
+    AUTHORIZED_FEATURE_TYPES,
+    AUTHORIZED_TASKS,
+    CGNS_FIELD_LOCATIONS,
+)
 from plaid.types import IndexType
+from plaid.types.feature_types import (
+    FeatureIdentifier,
+    FeatureIdentifierSequence,
+)
 
 # %% Globals
 
@@ -42,12 +52,26 @@ logging.basicConfig(
 
 
 class ProblemDefinition(object):
-    """Gathers all necessary informations to define a learning problem."""
+    """Gathers all necessary informations to define a learning problem.
 
-    def __init__(self, directory_path: Union[str, Path] = None) -> None:
+    This class defines a machine learning problem by specifying:
+    1. Input and output feature identifiers
+    2. Training task (regression, classification, etc.)
+    3. Train/test split configuration
+
+    Features are identified by dictionaries containing:
+    - type: Feature type (scalar, field, nodes, time_series)
+    - name: Feature name (required for all except nodes)
+    - base_name: Base name (required for field and nodes)
+    - zone_name: Zone name (required for field and nodes)
+    - location: Field location (required for field)
+    - time: Timestamp (optional for field and nodes)
+    """
+
+    def __init__(self, directory_path: Optional[Union[str, Path]] = None) -> None:
         """Initialize an empty :class:`ProblemDefinition <plaid.problem_definition.ProblemDefinition>`.
 
-        Use :meth:`add_inputs <plaid.problem_definition.ProblemDefinition.add_inputs>` or :meth:`add_output_scalars_names <plaid.problem_definition.ProblemDefinition.add_output_scalars_names>` to feed the :class:`ProblemDefinition`
+        Use :meth:`add_inputs <plaid.problem_definition.ProblemDefinition.add_inputs>` or :meth:`add_outputs <plaid.problem_definition.ProblemDefinition.add_outputs>` to feed the :class:`ProblemDefinition`
 
         Args:
             directory_path (Union[str, Path], optional): The path from which to load PLAID problem definition files.
@@ -65,29 +89,25 @@ class ProblemDefinition(object):
                 # 2. Load problem definition and create ProblemDefinition instance
                 problem_definition = ProblemDefinition("path_to_plaid_prob_def")
                 print(problem_definition)
-                >>> ProblemDefinition(input_scalars_names=['s_1'], output_scalars_names=['s_2'], input_meshes_names=['mesh'], task='regression')
+                >>> ProblemDefinition(inputs=[{'type': 'scalar', 'name': 's_1'}], outputs=[{'type': 'scalar', 'name': 's_2'}], task='regression')
         """
-        self._task: str = None  # list[task name]
-        self.in_scalars_names: list[str] = []
-        self.out_scalars_names: list[str] = []
-        self.in_timeseries_names: list[str] = []
-        self.out_timeseries_names: list[str] = []
-        self.in_fields_names: list[str] = []
-        self.out_fields_names: list[str] = []
-        self.in_meshes_names: list[str] = []
-        self.out_meshes_names: list[str] = []
-        self._split: dict[str, IndexType] = None
+        # Core attributes
[…]
