### 2025-08-13 6360fea
(dataset) remove one #pragma no cover + add Optional type hint where necessary
