### 2025-08-14 159533e
feat(split.py) add mmd_subsample_fn function
