### 2025-08-14 ee3d1d4
feat(split.py) add testing for mmd_subsample_fn
