### 2025-08-14 350b802
🐛 fix(docs) correct doc compilation (#135)
* fix(.readthedocs.yml) config for git fetch --unshallow for unlimited list of commits retrieved

* fix(docs) update config and huggingface_bridge for doc compatibility

* fix(huggingface_bridge.py) ruff formatting

* feat(CHANGELOG.md) update CHANGELOG

---------

Co-authored-by: Fabien Casenave <fabien.casenave@safrangroup.com>
