### 2025-08-14 5c87f50
fix(split.py) correct typing
