### 2025-08-14 fd89e73
🎉 Online saving with appropriate arborescence (#18)
* Update src/plaid/containers/dataset.py

* (dataset) add tests for add_to_dir method

* (dataset) remove one #pragma no cover + add Optional type hint where necessary

* update changelog

---------

Co-authored-by: Xavier Roynard <xavier.roynard@safrangroup.com>
