### 2025-08-14 eb2dcad
:tada: Mmd split of input scalars (#21)
* feat(split.py) add mmd_subsample_fn function

* feat(split.py) add testing for mmd_subsample_fn

* feat(CHANGELOG.md) update CHANGELOG

* fix(split.py) correct typing
