### 2025-08-14 ccecd72
fix(huggingface_bridge.py) enforce ruff convention
