+++ src/plaid/bridges/huggingface_bridge.py
@@ -1,13 +1,13 @@
 """Huggingface bridge for PLAID datasets."""
 
 import os
+
 # -*- coding: utf-8 -*-
 #
 # This file is subject to the terms and conditions defined in
 # file 'LICENSE.txt', which is part of this source code package.
 #
 #
-
 import pickle
 import shutil
 import sys
@@ -299,7 +299,6 @@ def huggingface_dataset_to_plaid(
                 disable=not verbose,
             ):
                 dataset.add_sample(sample)
-
 
     infos = {}
     if "legal" in ds.description:
