### 2025-08-14 2243239
🎉 Pipeline formalization, including scikit-learn block wrappers (#101)
* wip

* update

* wip

* wip

* wip

* wip: a cleaned version of scikit-learn pipelines is provided in examples/pipelines

* fix(ruff formatting)

* feat(tests) reduce samples list in autotests to speedup pytest runs

* feat(tests) improve coverage of huggingface_bridge

* feat(huggingface_bridge) improve coverage

* feat(examples/pipelines) remove wip files

* fix(huggingface_bridge) coverage: ignore line not reported with multiprocessing

* fix(examples/pipelines) remove comment

* fix(examples/pipelines) remove unused i
[…]
