+++ tests/containers/test_sample.py
@@ -179,6 +179,9 @@ class Test_Sample:
         with pytest.raises(FileExistsError):
             Sample(dataset_path)
 
+    def test_copy(self, sample_with_tree_and_scalar_and_time_series):
+        sample_with_tree_and_scalar_and_time_series.copy()
+
     # -------------------------------------------------------------------------#
     def test_set_default_base(self, sample, topological_dim, physical_dim):
         sample.init_base(topological_dim, physical_dim, time=0.5)
@@ -633,6 +636,12 @@ class Test_Sample:
         assert scalar is not None
         assert isinstance(scalar, np.ndarray)
 
+    def test__add_feature(self, sample_with_scalar):
+        sample_with_scalar._add_feature(
+            feature_identifier={"type": "scalar", "name": "test_scalar_2"},
+            feature=[3.1415],
+        )
+
     # -------------------------------------------------------------------------#
     def test_get_time_series_names_empty(self, sample):
         assert sample.get_time_series_names() == []
@@ -877,6 +886,9 @@ class Test_Sample:
         fields = new_sample.get_field_names(zone_name, base_name, location="CellCenter")
         assert len(fields) == 0
 
+    def test_del_all_fields(self, sample_with_tree):
+        sample_with_tree.del_all_fields()
+
     # -------------------------------------------------------------------------#
     def test_get_feature_from_string_identifier(
         self, sample_with_tree_and_scalar_and_time_series
@@ -1001,6 +1013,327 @@ class Test_Sample:
         sample_with_tree_and_scalar_and_time_series.get_feature_from_identifier(
             {"type": "nodes", "time": 0.0}
         )
+
+    def test_get_features_from_identifiers(
+        self, sample_with_tree_and_scalar_and_time_series
+    ):
+        sample_with_tree_and_scalar_and_time_series.get_features_from_identifiers(
+            [{"type": "scalar", "name": "test_scalar_1"}]
+        )
+        sample_with_tree_and_scalar_and_time_series.get_features_from_identifiers(
+            [
+                {"type": "scalar", "name": "test_scalar_1"},
+                {"type": "time_series", "name": "test_time_series_1"},
+            ]
+        )
+
+        sample_with_tree_and_scalar_and_time_series.get_features_from_identifiers(
+            [
+                {
+                    "type": "field",
+                    "name": "test_node_field_1",
+                    "base_name": "Base_2_2",
+                    "zone_name": "Zone",
+                    "location": "Vertex",
+                    "time": 0.0,
+                },
+                {"type": "scalar", "name": "test_scalar_1"},
+                {"type": "nodes"},
+            ]
+        )
+
+    def test_update_features_from_identifier(
+        self, sample_with_tree_and_scalar_and_time_series
+    ):
+        before = sample_with_tree_and_scalar_and_time_series.get_scalar("test_scalar_1")
+        sample_ = (
+            sample_with_tree_and_scalar_and_time_series.update_features_from_identifier(
+                feature_identifiers={"type": "scalar", "name": "test_scalar_1"},
+                features=3.1415,
+                in_place=False,
+            )
+        )
+        after = sample_.get_scalar("test_scalar_1")
+        assert after != before
+
+        before = sample_with_tree_and_scalar_and_time_series.get_time_series(
+            "test_time_series_1"
+        )
+        sample_ = (
+            sample_with_tree_and_scalar_and_time_series.update_features_from_identifier(
+                feature_identifiers={
+                    "type": "time_series",
+                    "name": "test_time_series_1",
+                },
+                features=(np.array([0, 1]), np.array([3.14, 3.15])),
+                in_place=False,
+            )
[…]
