+++ examples/pipelines/pipeline.py
@@ -1,0 +1,352 @@
+#!/usr/bin/env python
+# coding: utf-8
+
+# # Pipeline Examples
+# 
+# This notebook demonstrates the end-to-end process of building a machine learning pipeline using PLAID datasets and PLAID’s scikit-learn-compatible blocks.
+
+# ## PCA-GP for `mach` field prediction of `VKI-LS59` dataset
+# 
+# Key steps covered:
+# 
+# - **Loading the PLAID dataset** using Hugging Face integration and PLAID’s dataset classes  
+# - **Standardizing features** with PLAID-wrapped scikit-learn transformers for scalars  
+# - **Dimensionality reduction** of flow fields via Principal Component Analysis (PCA) to reduce output complexity  
+# - **Regression modeling** of PCA coefficients from scalar inputs using Gaussian Process regression  
+# - **Pipeline assembly** combining transformations and regressors into a single scikit-learn-compatible workflow  
+# - **Hyperparameter tuning** using Optuna and scikit-learn’s `GridSearchCV`
+# - **Best practices** for working with PLAID datasets and pipelines in a reproducible and modular manner
+
+# ### 📦 Imports
+
+# In[ ]:
+
+
+import warnings
+warnings.filterwarnings('ignore', module='sklearn')
+warnings.filterwarnings("ignore", message=".*IProgress not found.*")
+
+import os
+from pathlib import Path
+
+import yaml
+import numpy as np
+import optuna
+
+from datasets.utils.logging import disable_progress_bar
+from datasets import load_dataset
+
+from sklearn.base import clone
+from sklearn.pipeline import Pipeline
+
+from sklearn.decomposition import PCA
+from sklearn.preprocessing import MinMaxScaler
+from sklearn.gaussian_process import GaussianProcessRegressor
+from sklearn.gaussian_process.kernels import Matern
+from sklearn.multioutput import MultiOutputRegressor
+
+from sklearn.model_selection import KFold, GridSearchCV
+
+from plaid.bridges.huggingface_bridge import huggingface_dataset_to_plaid, huggingface_description_to_problem_definition
+from plaid.pipelines.sklearn_block_wrappers import WrappedSklearnTransformer, WrappedSklearnRegressor
+from plaid.pipelines.plaid_blocks import TransformedTargetRegressor, ColumnTransformer
+
+disable_progress_bar()
+n_processes = min(max(1, os.cpu_count()), 6)
+
+
+# ### 📥 Load Dataset
+# 
+# We load the `VKI-LS59` dataset from Hugging Face and restrict ourselves to the first 24 samples of the training set.
+
+# In[ ]:
+
+
+hf_dataset = load_dataset("PLAID-datasets/VKI-LS59", split="all_samples[:24]")
+dataset_train, _ = huggingface_dataset_to_plaid(hf_dataset, processes_number = n_processes, verbose = False)
+
+
+# We print the summary of dataset_train, which contains 24 samples, with 8 scalars and 8 fields, which is consistent with the `VKI-LS59` dataset:
+
+# In[ ]:
+
+
+print(dataset_train)
+
+
+# ### ⚙️ Pipeline Configuration
+# 
+# For convenience, the `in_features_identifiers` and `out_features_identifiers` for each pipeline block are defined in a `.yml` file. Here's an example of how the configuration might look:
+
+# ```yaml
+# pca_nodes:
+#   in_features_identifiers:
+#     - type: nodes
+#       base_name: Base_2_2
+#   out_features_identifiers:
+#     - type: scalar
+#       name: reduced_nodes_*
+# ```
+
+# In[ ]:
+
+
+try:
[…]
