### 2025-08-14 bcf291c
feat(CHANGELOG.md) update for 0.1.7 release
