### 2025-08-14 650d884
📦 update packaging: bundle HDF5 in the wheel
