+++ pyproject.toml
@@ -1,7 +1,9 @@
 [build-system]
 requires = [
-    "setuptools>=60",
-    "setuptools-scm>=8.0"]
+    "setuptools>=61",
+    "wheel",
+    "setuptools-scm>=8.0"
+]
 build-backend = "setuptools.build_meta"
 
 [project]
