+++ .github/workflows/publish-pypi.yml
@@ -14,11 +14,6 @@ jobs:
     steps:
       - uses: actions/checkout@v4
 
-      - name: Install HDF5 dev libraries
-        run: |
-          sudo apt-get update
-          sudo apt-get install -y libhdf5-dev
-
       - name: Set up Python
         uses: actions/setup-python@v4
         with:
@@ -27,7 +22,7 @@ jobs:
       - name: Install build tools
         run: |
           python -m pip install --upgrade pip
-          pip install build
+          pip install --upgrade setuptools wheel build
 
       - name: Check version consistency
         run: |
