+++ .github/workflows/testing.yml
@@ -5,10 +5,10 @@ permissions:
 
 on:
   push:
-    branches: 
+    branches:
       - '**'
   pull_request:
-    branches: 
+    branches:
       - '**'
   schedule:
     - cron: '0 0 * * *'  # Every day at 00:00 UTC
@@ -50,7 +50,7 @@ jobs:
       - name: Create environment
         run: |
           conda env create -n plaid-dev -f environment.yml
-          conda run -n plaid-dev pip install -e .
+          conda run -n plaid-dev pip install -e .[dev]
 
       - name: Run tests on Linux
         if: runner.os == 'Linux'
