+++ .github/workflows/doc.yml
@@ -2,10 +2,10 @@ name: Documentation
 
 on:
   push:
-    branches: 
+    branches:
       - '**'
   pull_request:
-    branches: 
+    branches:
       - '**'
   schedule:
     - cron: '0 0 * * *'  # Every day at 00:00 UTC
@@ -32,7 +32,7 @@ jobs:
       - name: Create environment
         run: |
           mamba env create -f environment.yml
-          mamba run -n plaid pip install -e .
+          mamba run -n plaid pip install -e .[dev]
 
       - name: Compile documentation
         run: |
