### 2025-08-14 afde3e0
:bug: Test build with uv
