+++ .github/workflows/publish-pypi.yml
@@ -6,67 +6,44 @@ on:
 
 jobs:
   build:
-    name: Build wheels
+    name: Build wheels with uv
+    runs-on: ubuntu-latest
     permissions:
       contents: read
-    runs-on: ubuntu-latest
 
     steps:
       - uses: actions/checkout@v4
 
-      - name: Install HDF5 dev libraries
-        run: |
-          sudo apt-get update
-          sudo apt-get install -y libhdf5-dev
+      - name: Set up uv
+        uses: astral-sh/setup-uv@v5
 
-      - name: Set up Python
-        uses: actions/setup-python@v4
-        with:
-          python-version: "3.11"
+      - name: Install the project
+        run: uv sync --locked --dev
 
-      - name: Install build tools
-        run: |
-          python -m pip install --upgrade pip
-          pip install build
-
-      - name: Check version consistency
-        run: |
-          pip install .
-          RELEASE_VER=${GITHUB_REF#refs/*/}
-          PACKAGE_VER="$(python -c 'import plaid; print(plaid.__version__)')"
-          if [ "$RELEASE_VER" != "$PACKAGE_VER" ]; then
-            echo "Package version ($PACKAGE_VER) doesn't match tag ($RELEASE_VER)"
-            exit 1
-          fi
-
-      - name: Build package
-        run: |
-          python -m build --wheel --outdir dist/
-
-      - uses: actions/upload-artifact@v4
-        with:
-          name: wheels-3.11
-          path: dist/*.whl
+      - name: Build distribution
+        run: uv build
 
   publish:
-    name: Publish to PyPI
     needs: build
+    if: github.event_name == 'release'
     runs-on: ubuntu-latest
     environment: pypi
     permissions:
       id-token: write
 
     steps:
-      - uses: actions/checkout@v4
+      - name: sanity check
+        run: echo "This job is only for publishing to PyPI."      
+      # - uses: actions/checkout@v4
 
-      - name: Download all wheel artifacts
-        uses: actions/download-artifact@v4
-        with:
-          pattern: wheels-*
-          path: dist
-          merge-multiple: true
+      # - name: Download all wheel artifacts
+      #   uses: actions/download-artifact@v4
+      #   with:
+      #     pattern: wheels-*
+      #     path: dist
+      #     merge-multiple: true
 
-      - name: Publish to PyPI
-        uses: pypa/gh-action-pypi-publish@release/v1
-        with:
-          packages-dir: dist
+      # - name: Publish to PyPI
+      #   uses: pypa/gh-action-pypi-publish@release/v1
+      #   with:
+      #     packages-dir: dist
