### 2025-08-14 fe15d04
numpy<2.0.0
