+++ pyproject.toml
@@ -30,19 +30,20 @@ classifiers = [
     "Topic :: Scientific/Engineering :: Artificial Intelligence",
     "Topic :: Scientific/Engineering :: Mathematics",
 ]
-dependencies = [
-    "tqdm",
-    "pyyaml",
-    "pycgns",
-    "numpy",
-    "matplotlib",
-    "rich",
-    "scikit-learn",
-    "datasets >=3.0",
-    "pydantic",
-]
+
 
 dynamic = ["version"]
+dependencies = [
+    "datasets>=4.0.0",
+    "matplotlib>=3.9.4",
+    "numpy<2.0.0",
+    "pycgns>=6.2.2",
+    "pydantic>=2.11.7",
+    "pyyaml>=6.0.2",
+    "rich>=14.1.0",
+    "scikit-learn>=1.6.1",
+    "tqdm>=4.67.1",
+]
 
 [tool.setuptools_scm]
 write_to = "src/plaid/_version.py"
@@ -80,10 +81,8 @@ dev = [
     "sphinx",
     "myst-nb",
     "myst-parser",
-    "nbconvert",
     "furo",
     "sphinx-autoapi",
     "sphinx-tabs",
     "sphinxcontrib.bibtex",
-    "optuna",
 ]
