+++ pyproject.toml
@@ -35,7 +35,6 @@ classifiers = [
 dependencies = [
     "tqdm",
     "pyyaml",
-    "h5py>=3.8",
     "pycgns",
     "scikit-learn",
     "datasets",
