+++ .github/workflows/doc.yml
@@ -31,10 +31,10 @@ jobs:
 
       - name: Create environment
         run: |
-          mamba env create -f environment.yml
-          mamba run -n plaid pip install -e .[dev]
+          conda create -n plaid-dev python=3.11 muscat=2.4.1 -c conda-forge
+          conda run -n plaid-dev pip install -e .[dev]
 
       - name: Compile documentation
         run: |
           cd docs
-          mamba run -n plaid bash generate_doc.sh
+          conda run -n plaid-dev bash generate_doc.sh
