+++ .github/workflows/publish-pypi.yml
@@ -17,6 +17,11 @@ jobs:
     steps:
       - uses: actions/checkout@v4
 
+      - name: Install HDF5 dev libraries
+        run: |
+          sudo apt-get update
+          sudo apt-get install -y libhdf5-dev
+
       - name: Set up Python
         uses: actions/setup-python@v4
         with:
