### 2025-08-15 7d6ecea
update build action for version check on release only
