### 2025-08-15 e6d1582
revert inital choices in bulding action and pyproject.toml
