+++ pyproject.toml
@@ -1,9 +1,7 @@
 [build-system]
 requires = [
-    "setuptools>=61",
-    "wheel",
-    "setuptools-scm>=8.0"
-]
+    "setuptools>=60",
+    "setuptools-scm>=8.0"]
 build-backend = "setuptools.build_meta"
 
 [project]
