+++ .github/workflows/publish-pypi.yml
@@ -30,7 +30,7 @@ jobs:
       - name: Install build tools
         run: |
           python -m pip install --upgrade pip
-          pip install --upgrade setuptools wheel build
+          pip install build
 
       - name: Check version consistency
         if: github.event_name == 'release'
