### 2025-08-15 007d0f9
build action: try to correct package version built
