### 2025-08-15 ac03d19
📦 feat(github actions) prepare full pip: rely an all but muscat on pypi  (#141)
* 📦 update packaging: bundle HDF5 in the wheel

* test

* test build

* update action

* update

* test2

* test

* revert publish-pypi.yml action

* update build action for version check on release only

* delete notebook outputs

* clean notebook [skip ci]

* revert inital choices in bulding action and pyproject.toml

* build action: try to correct package version built
