+++ .github/workflows/doc.yml
@@ -2,10 +2,10 @@ name: Documentation
 
 on:
   push:
-    branches: 
+    branches:
       - '**'
   pull_request:
-    branches: 
+    branches:
       - '**'
   schedule:
     - cron: '0 0 * * *'  # Every day at 00:00 UTC
@@ -31,10 +31,10 @@ jobs:
 
       - name: Create environment
         run: |
-          mamba env create -f environment.yml
-          mamba run -n plaid pip install -e .
+          conda create -n plaid-dev python=3.11 muscat=2.4.1 -c conda-forge
+          conda run -n plaid-dev pip install -e .[dev]
 
       - name: Compile documentation
         run: |
           cd docs
-          mamba run -n plaid bash generate_doc.sh
+          conda run -n plaid-dev bash generate_doc.sh
