+++ .github/workflows/testing.yml
@@ -25,10 +25,10 @@ jobs:
 
     steps:
       - name: Checkout code
-        uses: actions/checkout@v3
+        uses: actions/checkout@v5
 
       - name: Set up Conda
-        uses: conda-incubator/setup-miniconda@v2
+        uses: conda-incubator/setup-miniconda@v3
         with:
           use-mamba: false
           python-version: 3.11
@@ -69,9 +69,9 @@ jobs:
           cd examples
           conda run -n plaid-dev run_examples.bat
 
-      - name: Upload coverage to Codecov (Linux only)
-        if: runner.os == 'Linux'
-        uses: codecov/codecov-action@v4
+      - name: Upload coverage to Codecov (Linux and python 3.11 only)
+        if: runner.os == 'Linux' && matrix.python-version == '3.11'
+        uses: codecov/codecov-action@v5
         with:
           files: coverage.xml
           token: ${{ secrets.CODECOV_TOKEN }}
