### 2025-08-15 3acdb69
update readtheforge config
