+++ .readthedocs.yml
@@ -11,9 +11,9 @@ conda:
   environment: environment.yml
 
 build:
-  os: "ubuntu-22.04"
+  os: "ubuntu-24.04"
   tools:
-    python: "mambaforge-4.10"
+    python: "mambaforge-latest"
   jobs:
     pre_build:
       - git fetch origin --unshallow
