+++ src/plaid/pipelines/plaid_blocks.py
@@ -19,7 +19,7 @@ from typing import Union
 
 try:
     from typing import Self  # Python 3.10+
-except ImportError:
+except ImportError:  # pragma: no cover
     from typing_extensions import Self
 
 import numpy as np
