### 2025-08-15 bb94593
feat(*/*) move to muscat=2.5.0
