+++ .github/workflows/testing.yml
@@ -37,7 +37,7 @@ jobs:
 
       - name: Create environment
         run: |
-          conda create -n plaid-dev python=${{ matrix.python-version }} muscat=2.4.1 -c conda-forge
+          conda create -n plaid-dev python=${{ matrix.python-version }} muscat=2.5.0 -c conda-forge
           conda run -n plaid-dev pip install -e .[dev]
 
       - name: Run tests on Linux
