### 2025-08-15 5892a2e
fix(benchmarks) revert muscat=2.5.0 modification, since dependency is set to muscat=2.4.1
