### 2025-08-16 aebf356
feat(examples) add simple example retrieval capabilities
