### 2025-08-16 a0af3cf
fix(huggingface_bridge) correct indices handles in conversion
