### 2025-08-16 e43a6cd
fix(huggingface_bridge) correct indices handles in conversion
