### 2025-08-17 be6634f
fix(sample) from_features_identifier: stop adding None features
