### 2025-08-18 6b259c0
fix(README/docs) correct instructions for using packages
