### 2025-08-18 #8830 @editorialbot
[REVIEW]: PLAID: Physics-Learning AI Datamodel
**Submitting author:** @casenave (<a href="http://orcid.org/0000-0002-8810-9128">Fabien Casenave</a>)
**Repository:** https://github.com/PLAID-lib/plaid
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.5
