### 2025-08-19 c76d8fe
🎉 Add practical mechanisms to retrieve advanced datasets and examples (#149)
* feat(examples) add simple example retrieval capabilities

* update

* delete notebook outputs
