+++ src/plaid/bridges/huggingface_bridge.py
@@ -1,4 +1,4 @@
-"""Huggingface bridge for PLAID datasets."""
+"""Hugging Face bridge for PLAID datasets."""
 
 import os
 
@@ -15,7 +15,7 @@ from multiprocessing import Pool
 from pathlib import Path
 from typing import Callable, Optional
 
-from datasets import load_from_disk
+from datasets import load_dataset, load_from_disk
 from tqdm import tqdm
 
 if sys.version_info >= (3, 11):
@@ -32,9 +32,9 @@ from plaid.containers.sample import Sample
 from plaid.problem_definition import ProblemDefinition
 
 """
-Convention with hf (huggingface) datasets:
-- hf-datasets contains a single huggingface split, named 'all_samples'.
-- samples contains a single huggingface feature, named called "sample".
+Convention with hf (Hugging Face) datasets:
+- hf-datasets contains a single Hugging Face split, named 'all_samples'.
+- samples contains a single Hugging Face feature, named called "sample".
 - Samples are instances of plaid.containers.sample.Sample.
 - Mesh objects included in samples follow the CGNS standard, and can be converted in Muscat.Containers.Mesh.Mesh.
 - problem_definition info is stored in hf-datasets "description" parameter
@@ -44,16 +44,16 @@ Convention with hf (huggingface) datasets:
 def generate_huggingface_description(
     infos: dict, problem_definition: ProblemDefinition
 ) -> dict[str]:
-    """Generates a huggingface dataset description field from a plaid dataset infos and problem definition.
+    """Generates a Hugging Face dataset description field from a plaid dataset infos and problem definition.
 
     The conventions chosen here ensure working conversion to and from huggingset datasets.
 
     Args:
-        infos (dict): infos entry of the plaid dataset from which the huggingface description is to be generated
-        problem_definition (ProblemDefinition): of which the huggingface description is to be generated
+        infos (dict): infos entry of the plaid dataset from which the Hugging Face description is to be generated
+        problem_definition (ProblemDefinition): of which the Hugging Face description is to be generated
 
     Returns:
-        dict[str]: huggingface dataset description
+        dict[str]: Hugging Face dataset description
     """
     description = {}
 
@@ -76,17 +76,17 @@ def generate_huggingface_description(
 def plaid_dataset_to_huggingface(
     dataset: Dataset, problem_definition: ProblemDefinition, processes_number: int = 1
 ) -> datasets.Dataset:
-    """Use this function for converting a huggingface dataset from a plaid dataset.
+    """Use this function for converting a Hugging Face dataset from a plaid dataset.
 
-    The dataset can then be saved to disk, or pushed to the huggingface hub.
+    The dataset can then be saved to disk, or pushed to the Hugging Face hub.
 
     Args:
-        dataset (Dataset): the plaid dataset to be converted in huggingface format
-        problem_definition (ProblemDefinition): from which the huggingface dataset is to be generated
-        processes_number (int, optional): The number of processes used to generate the huggingface dataset
+        dataset (Dataset): the plaid dataset to be converted in Hugging Face format
+        problem_definition (ProblemDefinition): from which the Hugging Face dataset is to be generated
+        processes_number (int, optional): The number of processes used to generate the Hugging Face dataset
 
     Returns:
-        datasets.Dataset: dataset in huggingface format
+        datasets.Dataset: dataset in Hugging Face format
 
     Example:
         .. code-block:: python
@@ -113,20 +113,20 @@ def plaid_generator_to_huggingface(
     problem_definition: ProblemDefinition,
     processes_number: int = 1,
 ) -> datasets.Dataset:
-    """Use this function for creating a huggingface dataset from a sample generator function.
[…]
