### 2025-08-19 e6fa9c1
🎉 feat(*/*) move to muscat=2.5.0 (#146)
---------

Co-authored-by: Fabien Casenave <fabien.casenave@safrangroup.com>
