+++ .github/workflows/doc.yml
@@ -31,7 +31,7 @@ jobs:
 
       - name: Create environment
         run: |
-          conda create -n plaid-dev python=3.11 muscat=2.4.1 -c conda-forge
+          conda create -n plaid-dev python=3.11 muscat=2.5.0 -c conda-forge
           conda run -n plaid-dev pip install -e .[dev]
 
       - name: Compile documentation
