### 2025-08-19 692330b
fix(huggingface_bridge.py) correct merge
