### 2025-08-19 d9b1946
fix(huggingface_bridge.py) missing imports
