### 2025-08-19 b6d67d1
♻️ fix(docs) tricks to correct sphinx warning - some remain, but the aut… (#147)
---------

Co-authored-by: Fabien Casenave <fabien.casenave@safrangroup.com>
