+++ benchmarks/Vi-Transf/src/data/loader/loader.py
@@ -1,6 +1,6 @@
 from plaid.containers.dataset import Dataset as PlaidDataset
 from plaid.problem_definition import ProblemDefinition
-from typing import List, Tuple, Union
+from typing import List, Union
 from torch_geometric.data import Data
 import os
 from tqdm import tqdm
@@ -13,7 +13,7 @@ class Loader():
                  processes_number: int=1):
         """
         Initializes the Loader with the given parameters.
-        
+
         :param bridge: A callable that bridges the dataset loading process.
         :param task_split: The split(s) of the dataset to load.
         """
@@ -22,12 +22,12 @@ class Loader():
         self.task_split: Union[List[str], str, List[Union[List[str], str]]] = task_split
         self.processes_number = processes_number
 
-    def load_plaid(self, verbose: bool) -> Tuple[ProblemDefinition, List[PlaidDataset]]:
+    def load_plaid(self, verbose: bool) -> tuple[ProblemDefinition, List[PlaidDataset]]:
         raise NotImplementedError("This method should be implemented in the subclass")
 
     def get_dataset_split(self,
                           problem_definition: ProblemDefinition,
-                          plaid_dataset: PlaidDataset) -> Union[PlaidDataset, Tuple[PlaidDataset]]:
+                          plaid_dataset: PlaidDataset) -> Union[PlaidDataset, tuple[PlaidDataset]]:
         if not type(self.task_split)==str:
             datasets = []
             for split in self.task_split:
@@ -38,20 +38,20 @@ class Loader():
                 datasets.append(dataset)
 
             return tuple(datasets)
-        
+
         ids = problem_definition.get_split(self.task_split)
         dataset = PlaidDataset()
         dataset.set_samples(plaid_dataset.get_samples(ids=ids))
         return (dataset, )
 
     def load(self,
-             verbose=False) -> Tuple[ProblemDefinition, Tuple[List[Data], ...]]:
+             verbose=False) -> tuple[ProblemDefinition, tuple[List[Data], ...]]:
         """
         Load and converts a plaid dataset to torch geometric format.
 
         Returns:
-            Tuple[ProblemDefinition, Union[List[Data], Tuple[List[Data], ...]]]: 
-            A tuple containing the problem definition and either a single list of Data objects 
+            tuple[ProblemDefinition, Union[List[Data], tuple[List[Data], ...]]]:
+            A tuple containing the problem definition and either a single list of Data objects
             or a tuple of multiple lists of Data objects.
         """
 
@@ -59,7 +59,7 @@ class Loader():
         problem_definition = buffer[0]
         dataset_list = buffer[1:]
 
-        processed_list = []        
+        processed_list = []
         for dataset in dataset_list:
             processed_list.append(self.plaid_to_bridge(dataset, problem_definition=problem_definition, verbose=verbose))
 
@@ -104,4 +104,3 @@ class Loader():
                 data_list.append(new_data)
 
         return data_list
-
