### 2025-08-22 e3b15cb
(docstrings) revert useless changes
