### 2025-08-22 9723bfc
(benchmarks) apply Ruff
