+++ benchmarks/Vi-Transf/src/models/vits/tokenization/partitioners/metis_partitioner.py
@@ -10,12 +10,16 @@ from torch.multiprocessing import Pool, cpu_count
 import torch
 from functools import partial
 
-def npoints_to_nparts(npoints: int, n_sim_points: int, absolute_tol: int, relative_tol: float) -> int:
+
+def npoints_to_nparts(
+    npoints: int, n_sim_points: int, absolute_tol: int, relative_tol: float
+) -> int:
     """computes the number of subdomains in a simulation, with a tolerance given the number of points per subdomains we want and the number of points in the simulation"""
     nparts = n_sim_points // npoints + 1
-    nparts = int(np.ceil((1 + relative_tol)*nparts + absolute_tol))
+    nparts = int(np.ceil((1 + relative_tol) * nparts + absolute_tol))
 
     return nparts
+
 
 def torch_geometric_to_metis_format(data: Data):
     """
@@ -33,50 +37,91 @@ def torch_geometric_to_metis_format(data: Data):
     num_nodes = data.num_nodes
     adjacency = [[] for _ in range(num_nodes)]
 
-    for source, target in zip(data.edge_index[0, :].numpy(), data.edge_index[1, :].numpy()):
+    for source, target in zip(
+        data.edge_index[0, :].numpy(), data.edge_index[1, :].numpy()
+    ):
         adjacency[source].append(target)
 
     return adjacency
 
 
 class MetisPartitioner(Partitioner):
-    def __init__(self, n_vertices_per_subdomain: int=None, processes_number: int=1, absolute_tol: int=1, relative_tol: float=0.05):
+    def __init__(
+        self,
+        n_vertices_per_subdomain: int = None,
+        processes_number: int = 1,
+        absolute_tol: int = 1,
+        relative_tol: float = 0.05,
+    ):
         super().__init__(n_vertices_per_subdomain)
         self.processes_number = processes_number
         self.absolute_tol = absolute_tol
         self.relative_tol = relative_tol
 
-    def partition(self, dataset: list[Data], seed: Optional[int]=None) -> np.ndarray:
-        if seed is None: seed = random.randint(0, 2**32 - 1)
+    def partition(self, dataset: list[Data], seed: Optional[int] = None) -> np.ndarray:
+        if seed is None:
+            seed = random.randint(0, 2**32 - 1)
         if seed is not None:
             rng_generator = random.Random(seed)
-            seed_vector = [rng_generator.randint(0, 2**30 - 1) for _ in range(len(dataset))] # metis doesn't accept seeds that are too large
+            seed_vector = [
+                rng_generator.randint(0, 2**30 - 1) for _ in range(len(dataset))
+            ]  # metis doesn't accept seeds that are too large
 
-        if self.processes_number==-1:
+        if self.processes_number == -1:
             self.processes_number = cpu_count()
 
         partitioned_dataset = []
 
         if self.processes_number > 1:
             with Pool(processes=self.processes_number) as pool:
-                partitioned_dataset = list(tqdm(pool.starmap(
-                    _partition_single,
-                    zip(dataset, [self.n_vertices_per_subdomain]*len(dataset), seed_vector, [self.absolute_tol]*len(dataset), [self.relative_tol]*len(dataset))
-                ), desc="Partitioning dataset with METIS", total=len(dataset)))
+                partitioned_dataset = list(
+                    tqdm(
+                        pool.starmap(
+                            _partition_single,
+                            zip(
+                                dataset,
+                                [self.n_vertices_per_subdomain] * len(dataset),
+                                seed_vector,
+                                [self.absolute_tol] * len(dataset),
+                                [self.relative_tol] * len(dataset),
+                            ),
+                        ),
+                        desc="Partitioning dataset with METIS",
[…]
