### 2025-08-22 bb9cda5
(examples) apply Ruff
