+++ examples/containers/dataset_example.py
@@ -14,7 +14,6 @@
 
 # %%
 # Import required libraries
-import os
 from pathlib import Path
 
 # %%
