### 2025-08-22 0364f0d
(benchmarks) apply Ruff
