+++ benchmarks/Vi-Transf/src/data/utils.py
@@ -1,6 +1,8 @@
-from plaid.containers.dataset import Dataset as PlaidDataset
 import copy
+
 from torch_geometric.data import Data
+
+from plaid.containers.dataset import Dataset as PlaidDataset
 
 
 def extract_plaid_dataset(plaid_dataset: PlaidDataset, ids: list[int]) -> PlaidDataset:
