### 2025-08-22 46a3b57
♻️ Fix typing + Ruff on benchmarks and exemples (#156)
* (typing) replace Dict by dict

* (typing) replace Tuple by tuple

* (typing) replace List by list

* (benchmark/examples) apply Ruff

* (benchmarks) apply Ruff

* (examples) apply Ruff

* (benchmarks) apply Ruff
