+++ benchmarks/Vi-Transf/src/models/vits/tokenization/tokenizer.py
@@ -1,14 +1,16 @@
+from typing import Optional
+
 import torch.nn as nn
-from torch_geometric.data import Data, Batch
-from typing import List, Optional
+from torch_geometric.data import Batch, Data
+
 
 class Tokenizer(nn.Module):
     def __init__(self):
         super().__init__()
-    
-    def preprocess(self, dataset: List[Data], seed: Optional[int]=None) -> List[Data]:
+
+    def preprocess(self, dataset: list[Data], seed: Optional[int] = None) -> list[Data]:
         pass
-        
+
     def forward(self, data: Batch):
         pass
 
