+++ examples/containers/sample_example.py
@@ -22,6 +22,7 @@ import CGNS.PAT.cgnskeywords as CGK
 import numpy as np
 from Muscat.Bridges.CGNSBridge import MeshToCGNS
 from Muscat.MeshTools import MeshCreationTools as MCT
+
 from plaid.containers.sample import Sample
 from plaid.utils import cgns_helper as CGH
 

[…]
