### 2025-08-22 70b6c07
fix(doc/env) set muscat=2.5.0 dependency. Remove useless files
