+++ docs/source/notebooks.rst
@@ -4,7 +4,7 @@ Notebooks
 You can find here detailed examples for different parts of plaid, explained in jupyter notebooks.
 
 .. note::
-   The `Muscat <https://anaconda.org/conda-forge/muscat>`_ dependency is required on some of the notebooks,
+   The `Muscat=2.5.0 <https://anaconda.org/conda-forge/muscat>`_ dependency is required on some of the notebooks,
    and is not distributed with the `plaid <https://anaconda.org/conda-forge/plaid>`_
    conda-forge package.
 
