### 2025-08-22 28d7ddb
(tests) add wrappers tests
