### 2025-08-22 9aa4831
(utils) add a deprecation decorator for old methods using feature names instead of identifiers
