### 2025-08-22 c1da4d0
📄 fix(doc/env) set muscat=2.5.0 dependency. Remove useless files (#158)
