### 2025-08-23 19641b4
feat(Hugging Face bridge) add datasetdict conversion, and simple function for plaid sample init from hf sample
