+++ src/plaid/bridges/huggingface_bridge.py
@@ -78,7 +78,10 @@ def generate_huggingface_description(
 
 
 def plaid_dataset_to_huggingface(
-    dataset: Dataset, problem_definition: ProblemDefinition, processes_number: int = 1
+    dataset: Dataset,
+    problem_definition: ProblemDefinition,
+    split: Optional[str] = "all_samples",
+    processes_number: Optional[int] = 1,
 ) -> datasets.Dataset:
     """Use this function for converting a Hugging Face dataset from a plaid dataset.
 
@@ -86,7 +89,8 @@ def plaid_dataset_to_huggingface(
 
     Args:
         dataset (Dataset): the plaid dataset to be converted in Hugging Face format
-        problem_definition (ProblemDefinition): from which the Hugging Face dataset is to be generated
+        problem_definition (ProblemDefinition): the problem definition is used to generate the description of the Hugging Face dataset.
+        split (str, optional): The name of the split.
         processes_number (int, optional): The number of processes used to generate the Hugging Face dataset
 
     Returns:
@@ -95,27 +99,75 @@ def plaid_dataset_to_huggingface(
     Example:
         .. code-block:: python
 
-            dataset = plaid_dataset_to_huggingface(dataset, problem_definition)
+            dataset = plaid_dataset_to_huggingface(dataset, problem_definition, split)
             dataset.save_to_disk("path/to/dir)
             dataset.push_to_hub("chanel/dataset")
     """
+    if split == "all_samples":
+        ids = dataset.get_sample_ids()
+    else:
+        ids = problem_definition.get_split(split)
 
     def generator():
-        for sample in dataset:
+        for sample in dataset[ids]:
             yield {
                 "sample": pickle.dumps(sample.model_dump()),
             }
 
     return plaid_generator_to_huggingface(
-        generator, dataset.get_infos(), problem_definition, processes_number
+        generator=generator,
+        infos=dataset.get_infos(),
+        problem_definition=problem_definition,
+        split=split,
+        processes_number=processes_number,
     )
+
+
+def plaid_dataset_to_huggingface_datasetdict(
+    dataset: Dataset,
+    problem_definition: ProblemDefinition,
+    main_splits: list,
+    processes_number: Optional[int] = 1,
+) -> datasets.Dataset:
+    """Use this function for converting a Hugging Face dataset dict from a plaid dataset.
+
+    The dataset can then be saved to disk, or pushed to the Hugging Face hub.
+
+    Args:
+        dataset (Dataset): the plaid dataset to be converted in Hugging Face format
+        problem_definition (ProblemDefinition): the problem definition is used to generate the description of the Hugging Face dataset.
+        main_splits (str): The name of the main splits: defining a partitioning of the sample ids.
+        processes_number (int, optional): The number of processes used to generate the Hugging Face dataset
+
+    Returns:
+        datasets.Dataset: dataset in Hugging Face format
+
+    Example:
+        .. code-block:: python
+
+            dataset = plaid_dataset_to_huggingface(dataset, problem_definition, split)
+            dataset.save_to_disk("path/to/dir)
+            dataset.push_to_hub("chanel/dataset")
+    """
+    _dict = {}
+    for _, split in enumerate(main_splits):
+        ds = plaid_dataset_to_huggingface(
+            dataset=dataset,
+            problem_definition=problem_definition,
+            split=split,
+            processes_number=processes_number,
+        )
+        _dict[split] = ds
+
+    return datasets.DatasetDict(_dict)
 
 
 def plaid_generator_to_huggingface(
     generator: Callable,
     infos: dict,
     problem_definition: ProblemDefinition,
-    processes_number: int = 1,
+    split: Optional[str] = "all_samples",
[…]
