### 2025-08-25 48dd182
(hf bridge) update to_plaid_sample docstring
