### 2025-08-26 a910968
♻️ rename types to remove Type from name of types(#164)
