+++ src/plaid/containers/sample.py
@@ -42,14 +42,14 @@ from plaid.containers.utils import get_feature_type_and_details_from
 from plaid.types import (
     CGNSNode,
     CGNSTree,
+    Feature,
     FeatureIdentifier,
-    FeatureType,
-    FieldType,
+    Field,
     LinkType,
     PathType,
-    ScalarType,
-    TimeSequenceType,
-    TimeSeriesType,
+    Scalar,
+    TimeSequence,
+    TimeSeries,
 )
 from plaid.utils import cgns_helper as CGH
 from plaid.utils.base import safe_len
@@ -167,8 +167,8 @@ class Sample(BaseModel):
         mesh_base_name: str = "Base",
         mesh_zone_name: str = "Zone",
         meshes: dict[float, CGNSTree] = None,
-        scalars: dict[str, ScalarType] = None,
-        time_series: dict[str, TimeSeriesType] = None,
+        scalars: dict[str, Scalar] = None,
+        time_series: dict[str, TimeSeries] = None,
         links: dict[float, list[LinkType]] = None,
         paths: dict[float, list[PathType]] = None,
     ) -> None:
@@ -179,8 +179,8 @@ class Sample(BaseModel):
             mesh_base_name (str, optional): The base name for the mesh. Defaults to 'Base'.
             mesh_zone_name (str, optional): The zone name for the mesh. Defaults to 'Zone'.
             meshes (dict[float, CGNSTree], optional): A dictionary mapping time steps to CGNSTrees. Defaults to None.
-            scalars (dict[str, ScalarType], optional): A dictionary mapping scalar names to their values. Defaults to None.
-            time_series (dict[str, TimeSeriesType], optional): A dictionary mapping time series names to their values. Defaults to None.
+            scalars (dict[str, Scalar], optional): A dictionary mapping scalar names to their values. Defaults to None.
+            time_series (dict[str, TimeSeries], optional): A dictionary mapping time series names to their values. Defaults to None.
             links (dict[float, list[LinkType]], optional): A dictionary mapping time steps to lists of links. Defaults to None.
             paths (dict[float, list[PathType]], optional): A dictionary mapping time steps to lists of paths. Defaults to None.
 
@@ -208,8 +208,8 @@ class Sample(BaseModel):
         self._mesh_zone_name: str = mesh_zone_name
 
         self._meshes: dict[float, CGNSTree] = meshes
-        self._scalars: dict[str, ScalarType] = scalars
-        self._time_series: dict[str, TimeSeriesType] = time_series
+        self._scalars: dict[str, Scalar] = scalars
+        self._time_series: dict[str, TimeSeries] = time_series
 
         self._links: dict[float, list[LinkType]] = links
         self._paths: dict[float, list[PathType]] = paths
@@ -1163,26 +1163,26 @@ class Sample(BaseModel):
             res = sorted(self._scalars.keys())
             return res
 
-    def get_scalar(self, name: str) -> ScalarType:
+    def get_scalar(self, name: str) -> Scalar:
         """Retrieve a scalar value associated with the given name.
 
         Args:
             name (str): The name of the scalar value to retrieve.
 
         Returns:
-            ScalarType or None: The scalar value associated with the given name, or None if the name is not found.
+            Scalar or None: The scalar value associated with the given name, or None if the name is not found.
         """
         if (self._scalars is None) or (name not in self._scalars):
             return None
         else:
             return self._scalars[name]
 
-    def add_scalar(self, name: str, value: ScalarType) -> None:
+    def add_scalar(self, name: str, value: Scalar) -> None:
         """Add a scalar value to a dictionary.
 
         Args:
             name (str): The name of the scalar value.
[…]
