### 2025-08-26 f42bf46
Defining collections
