+++ src/plaid/containers/collections.py
@@ -1,0 +1,582 @@
+"""Module for implementing collections."""
+
+import logging
+from typing import Optional, Union
+
+import CGNS.PAT.cgnskeywords as CGK
+import CGNS.PAT.cgnslib as CGL
+import CGNS.PAT.cgnsutils as CGU
+import numpy as np
+
+from plaid.types import CGNSNode, CGNSTree, Field, Scalar
+from plaid.utils import cgns_helper as CGH
+
+logger = logging.getLogger(__name__)
+logging.basicConfig(
+    format="[%(asctime)s:%(levelname)s:%(filename)s:%(funcName)s(%(lineno)d)]:%(message)s",
+    level=logging.INFO,
+)
+
+
+def _check_names(names: Union[str, list[str]]):
+    """Check that names do not contain invalid character ``/``.
+
+    Args:
+        names (Union[str, list[str]]): The names to check.
+
+    Raises:
+        ValueError: If any name contains the invalid character ``/``.
+    """
+    if isinstance(names, str):
+        names = [names]
+    for name in names:
+        if (name is not None) and ("/" in name):
+            raise ValueError(
+                f"feature_names containing `/` are not allowed, but {name=}, you should first replace any occurence of `/` with something else, for example: `name.replace('/','__')`"
+            )
+
+
+class ScalarCollection:
+    """Manager object for scalars."""
+
+    def __init__(self):
+        self.features: dict[str, Scalar] = {}
+
+    def add(self, name: str, value: Scalar) -> None:
+        """Add a scalar value to a dictionary.
+
+        Args:
+            name (str): The name of the scalar value.
+            value (Scalar): The scalar value to add or update in the dictionary.
+        """
+        _check_names([name])
+        if self._scalars is None:
+            self._scalars = {name: value}
+        else:
+            self._scalars[name] = value
+
+    def remove(self, name: str) -> Scalar:
+        """Delete a scalar value from the dictionary.
+
+        Args:
+            name (str): The name of the scalar value to be deleted.
+
+        Raises:
+            KeyError: Raised when there is no scalar / there is no scalar with the provided name.
+
+        Returns:
+            Scalar: The value of the deleted scalar.
+        """
+        if self._scalars is None:
+            raise KeyError("There is no scalar inside this sample.")
+
+        if name not in self._scalars:
+            raise KeyError(f"There is no scalar value with name {name}.")
+
+        return self._scalars.pop(name)
+
+    def get(self, name: str) -> Scalar:
+        """Retrieve a scalar value associated with the given name.
+
+        Args:
+            name (str): The name of the scalar value to retrieve.
+
+        Returns:
+            Scalar or None: The scalar value associated with the given name, or None if the name is not found.
+        """
+        if (self._scalars is None) or (name not in self._scalars):
+            return None
+        else:
+            return self._scalars[name]
+
+    def get_names(self) -> list[str]:
+        """Get a set of scalar names available in the object.
+
+        Returns:
+            set[str]: A set containing the names of the available scalars.
+        """
+        if self._scalars is None:
+            return []
+        else:
+            res = sorted(self._scalars.keys())
+            return res
+
+
+class FieldCollection:
+    """Manager object for fields."""
+
+    def __init__(self):
+        self.features: dict[str, Field] = {}
+        self._defaults: dict = {
+            "active_base": None,
+            "active_zone": None,
+            "active_time": None,
+        }
+
+    def add(
+        self,
+        name: str,
+        field: Field,
+        zone_name: Optional[str] = None,
+        base_name: Optional[str] = None,
+        location: str = "Vertex",
+        time: Optional[float] = None,
+        warning_overwrite=True,
+    ) -> None:
+        """Add a field to a specified zone in the grid.
+
+        Args:
[…]
