+++ src/plaid/containers/sample_container/__init__.py
@@ -1,1 +1,0 @@
-"""Implements the `Sample` container and features managers."""
