### 2025-08-26 e6b3035
:recycle: Introduce SampleScalars
