+++ examples/utils/init_with_tabular_example.py
@@ -62,12 +62,12 @@ print(f"{sample_1 = }")
 
 # %%
 # Access and display the value of a particular scalar within a sample
-scalar_value = sample_1.get_scalar("scalar_0")
+scalar_value = sample_1.scalars.get("scalar_0")
 print("Scalar 'scalar_0' in Sample 1:", scalar_value)
 
 # %%
 # Retrieve tabular data from the dataset based on scalar names
 scalar_names = ["scalar_1", "scalar_3", "scalar_5"]
-tabular_data_subset = dataset.get_scalars_to_tabular(scalar_names)
+tabular_data_subset = dataset.scalars.gets_to_tabular(scalar_names)
 print("Tabular Data Subset for Scalars 1, 3, and 5:")
 dprint("tabular_data_subset", tabular_data_subset)
