+++ benchmarks/FNO/VKI-LS59/prepare_vki.py
@@ -1,11 +1,8 @@
 from plaid.problem_definition import ProblemDefinition
 from plaid.containers.sample import Sample
-import numpy as np
 from Muscat.Bridges.CGNSBridge import MeshToCGNS
-import Muscat.Containers.ElementsDescription as ED
 from Muscat.Containers.ConstantRectilinearMeshTools import CreateConstantRectilinearMesh
 from Muscat.Containers.MeshTetrahedrization import Tetrahedrization
-from Muscat.Containers import MeshCreationTools as MCT
 import os, time, shutil
 from tqdm import tqdm
 
@@ -64,7 +61,7 @@ for sample_index in tqdm(range(nSamples)):
         raise("unkown sample_index")
 
     for sn in scalar_names:
-        new_sample.add_scalar(sn, sample.get_scalar(sn))
+        new_sample.scalars.add(sn, sample.scalars.get(sn))
 
     new_sample.add_field("Signed_Distance", sample.get_field("sdf", base_name="Base_2_2"))
 
