+++ src/plaid/containers/sample.py
@@ -38,6 +38,7 @@ from plaid.constants import (
     CGNS_ELEMENT_NAMES,
     CGNS_FIELD_LOCATIONS,
 )
+from plaid.containers.collections import SampleScalars, _check_names
 from plaid.containers.utils import get_feature_type_and_details_from
 from plaid.types import (
     CGNSNode,
@@ -59,24 +60,6 @@ logging.basicConfig(
     format="[%(asctime)s:%(levelname)s:%(filename)s:%(funcName)s(%(lineno)d)]:%(message)s",
     level=logging.INFO,
 )
-
-
-def _check_names(names: Union[str, list[str]]):
-    """Check that names do not contain invalid character ``/``.
-
-    Args:
-        names (Union[str, list[str]]): The names to check.
-
-    Raises:
-        ValueError: If any name contains the invalid character ``/``.
-    """
-    if isinstance(names, str):
-        names = [names]
-    for name in names:
-        if (name is not None) and ("/" in name):
-            raise ValueError(
-                f"feature_names containing `/` are not allowed, but {name=}, you should first replace any occurence of `/` with something else, for example: `name.replace('/','__')`"
-            )
 
 
 def _read_index(pyTree: list, dim: list[int]):
@@ -161,8 +144,11 @@ class Sample(BaseModel):
         directory_path: Optional[Union[str, Path]] = None,
         mesh_base_name: str = "Base",
         mesh_zone_name: str = "Zone",
-        links: dict[float, list[LinkType]] = None,
-        paths: dict[float, list[PathType]] = None,
+        meshes: Optional[dict[float, CGNSTree]] = None,
+        scalars: Optional[dict[str, Scalar]] = None,
+        time_series: Optional[dict[str, TimeSeries]] = None,
+        links: Optional[dict[float, list[LinkType]]] = None,
+        paths: Optional[dict[float, list[PathType]]] = None,
     ) -> None:
         """Initialize an empty :class:`Sample <plaid.containers.sample.Sample>`.
 
@@ -199,24 +185,28 @@ class Sample(BaseModel):
         self._mesh_base_name: str = mesh_base_name
         self._mesh_zone_name: str = mesh_zone_name
 
-        self._meshes: dict[float, CGNSTree] = {}
-        self._scalars: dict[str, Scalar] = {}
-        self._time_series: dict[str, TimeSeries] = {}
+        self._meshes: dict[float, CGNSTree] | None = meshes
+        self._scalars = SampleScalars(scalars)
+        self._time_series: dict[str, TimeSeries] | None = time_series
 
-        self._links: dict[float, list[LinkType]] = links
-        self._paths: dict[float, list[PathType]] = paths
+        self._links: dict[float, list[LinkType]] | None = links
+        self._paths: dict[float, list[PathType]] | None = paths
 
         if directory_path is not None:
             directory_path = Path(directory_path)
             self.load(directory_path)
 
-        self._defaults: dict = {
+        self._defaults: dict[str, Optional[str]] = {
             "active_base": None,
             "active_zone": None,
             "active_time": None,
         }
 
         self._extra_data = None
+
+    @property
+    def scalars(self) -> SampleScalars:
+        return self._scalars
 
     def copy(self) -> Self:
         """Create a deep copy of the sample.
@@ -1143,66 +1133,6 @@ class Sample(BaseModel):
         return CGU.getValueByPath(zone_node, "ZoneType").tobytes().decode()
 
     # -------------------------------------------------------------------------#
-    def get_scalar_names(self) -> set[str]:
-        """Get a set of scalar names available in the object.
-
-        Returns:
-            set[str]: A set containing the names of the available scalars.
-        """
-        if self._scalars is None:
-            return []
[…]
