### 2025-08-27 f0256ca
Not breaking current API [skip ci]
