+++ examples/bridges/huggingface_bridge_example.py
@@ -91,10 +91,10 @@ for cgns_tree in tqdm(CGNS_meshes):
 
     # Add random scalar values to the sample
     for sname in in_scalars_names:
-        sample.scalars.add(sname, np.random.randn())
+        sample.add_scalar(sname, np.random.randn())
 
     for sname in out_scalars_names:
-        sample.scalars.add(sname, np.random.randn())
+        sample.add_scalar(sname, np.random.randn())
 
     # Add random field values to the sample
     for j, sname in enumerate(out_fields_names):
