+++ src/plaid/utils/stats.py
@@ -376,7 +376,7 @@ class Stats:
             sample (Sample): Sample containing scalar data
             data_dict (dict[str, list]): Dictionary to store processed data
         """
-        for name in sample.get_scalar_names( ):
+        for name in sample.get_scalar_names():
             if name not in data_dict:
                 data_dict[name] = []
             value = sample.get_scalar(name)
