### 2025-08-27 ea34cd7
(typing) replace LinkType and PathType by CGNSLink and CGNSPath respectively
