### 2025-08-27 848a208
(typing) add CGNS to names of CGNSNodes related types
