### 2025-08-28 266d727
remove dirty space between brackets
