### 2025-08-28 555bed7
🎉  feat(Hugging Face bridge) add datasetdict conversion, and simple func… (#159)
Co-authored-by: Xavier Roynard <xavroynard@hotmail.com>
