+++ src/plaid/containers/sample.py
@@ -38,6 +38,7 @@ from plaid.constants import (
     CGNS_ELEMENT_NAMES,
     CGNS_FIELD_LOCATIONS,
 )
+from plaid.containers.collections import SampleScalars, _check_names
 from plaid.containers.utils import get_feature_type_and_details_from
 from plaid.types import (
     CGNSLink,
@@ -60,31 +61,8 @@ logging.basicConfig(
     level=logging.INFO,
 )
 
-# %% Globals
 
-
-# %% Classes
-
-
-def _check_names(names: Union[str, list[str]]):
-    """Check that names do not contain invalid character ``/``.
-
-    Args:
-        names (Union[str, list[str]]): The names to check.
-
-    Raises:
-        ValueError: If any name contains the invalid character ``/``.
-    """
-    if isinstance(names, str):
-        names = [names]
-    for name in names:
-        if (name is not None) and ("/" in name):
-            raise ValueError(
-                f"feature_names containing `/` are not allowed, but {name=}, you should first replace any occurence of `/` with something else, for example: `name.replace('/','__')`"
-            )
-
-
-def read_index(pyTree: list, dim: list[int]):
+def _read_index(pyTree: list, dim: list[int]):
     """Read Index Array or Index Range from CGNS.
 
     Args:
@@ -94,12 +72,12 @@ def read_index(pyTree: list, dim: list[int]):
     Returns:
         indices
     """
-    a = read_index_array(pyTree)
-    b = read_index_range(pyTree, dim)
+    a = _read_index_array(pyTree)
+    b = _read_index_range(pyTree, dim)
     return np.hstack((a, b))
 
 
-def read_index_array(pyTree: list):
+def _read_index_array(pyTree: list):
     """Read Index Array from CGNS.
 
     Args:
@@ -119,7 +97,7 @@ def read_index_array(pyTree: list):
     return np.array(res, dtype=int).ravel()
 
 
-def read_index_range(pyTree: list, dim: list[int]):
+def _read_index_range(pyTree: list, dim: list[int]):
     """Read Index Range from CGNS.
 
     Args:
@@ -163,14 +141,14 @@ class Sample(BaseModel):
 
     def __init__(
         self,
-        directory_path: Union[str, Path] = None,
+        directory_path: Optional[Union[str, Path]] = None,
         mesh_base_name: str = "Base",
         mesh_zone_name: str = "Zone",
-        meshes: dict[float, CGNSTree] = None,
-        scalars: dict[str, Scalar] = None,
-        time_series: dict[str, TimeSeries] = None,
-        links: dict[float, list[CGNSLink]] = None,
-        paths: dict[float, list[CGNSPath]] = None,
+        meshes: Optional[dict[float, CGNSTree]] = None,
+        scalars: Optional[dict[str, Scalar]] = None,
+        time_series: Optional[dict[str, TimeSeries]] = None,
+        links: Optional[dict[float, list[CGNSLink]]] = None,
+        paths: Optional[dict[float, list[CGNSPath]]] = None,
     ) -> None:
         """Initialize an empty :class:`Sample <plaid.containers.sample.Sample>`.
 
@@ -207,24 +185,66 @@ class Sample(BaseModel):
         self._mesh_base_name: str = mesh_base_name
         self._mesh_zone_name: str = mesh_zone_name
 
-        self._meshes: dict[float, CGNSTree] = meshes
-        self._scalars: dict[str, Scalar] = scalars
-        self._time_series: dict[str, TimeSeries] = time_series
+        self._meshes: Optional[dict[float, CGNSTree]] = meshes
+        self._scalars = SampleScalars(scalars)
+        self._time_series: Optional[dict[str, TimeSeries]] = time_series
 
[…]
