### 2025-08-28 e9f66d5
(utils.interpolation) fix docstrings style
