+++ src/plaid/utils/interpolation.py
@@ -26,19 +26,12 @@ def binary_search(
         - 0 if item <= ordered_list[0]
         - the rank of the largest element smaller or equal than item otherwise
 
-    Parameters
-    ----------
-    ordered_list: List or one-dimensional np.ndarray
-        the data sorted in increasing order from which the previous rank is\
-        searched.
-    item : float or int
-        the item for which the previous rank is searched.
+    Args:
+        ordered_list (Union[list, np.ndarray]): The data sorted in increasing order from which the previous rank is searched.
+        item (Union[float, int]): The item for which the previous rank is searched.
 
     Returns:
-    -------
-    int
-        0 or the rank of the largest element smaller or equal than item in\
-        "ordered_list".
+        int: 0 or the rank of the largest element smaller or equal than item in "ordered_list".
     """
     return max(bisect.bisect_right(ordered_list, item) - 1, 0)
 
@@ -49,8 +42,8 @@ def binary_search_vectorized(
     """Vectorized binary search for multiple items in a sorted list (items is now a list or one-dimensional np.ndarray).
 
     Args:
-        ordered_list (list or one-dimensional np.ndarray): The data sorted in increasing order.
-        items (list or one-dimensional np.ndarray): The items for which ranks are searched.
+        ordered_list (Union[list, np.ndarray]): The data sorted in increasing order.
+        items (Union[list, np.ndarray]): The items for which ranks are searched.
 
     Returns:
         np.ndarray: An array containing the ranks of the largest elements smaller or equal to each item.
@@ -68,21 +61,14 @@ def piece_wise_linear_interpolation(
 ) -> np.ndarray:
     """Computes a item interpolation for temporal vectors defined either by item_indices  and vectors at these indices.
 
-    Parameters
-    ----------
-    item : float
-        the input item at which the interpolation is required.
-    item_indices : np.ndarray
-        the items where the available data is defined, of size (numberOfTimeIndices).
-    vectors : np.ndarray or dict
-        the available data, of size (numberOfVectors, numberOfDofs).
-    tolerance : float
-        tolerance for deciding when using the closest timestep value instead of carrying out the linear interpolation, default to 1e-4.
+    Args:
+        item (float): The input item at which the interpolation is required.
+        item_indices (np.ndarray): The items where the available data is defined, of size (numberOfTimeIndices).
+        vectors (Union[np.ndarray, dict]): The available data, of size (numberOfVectors, numberOfDofs).
+        tolerance (float): Tolerance for deciding when using the closest timestep value instead of carrying out the linear interpolation, default to 1e-4.
 
     Returns:
-    -------
-    np.ndarray
-        interpolated vector, of size (numberOfDofs).
+        np.ndarray: Interpolated vector, of size (numberOfDofs).
     """
     if item <= item_indices[0]:
         return vectors[0]
@@ -104,26 +90,18 @@ def piece_wise_linear_interpolation_with_map(
     vectors: Union[np.ndarray, dict],
     vectors_map: list = None,
     tolerance: float = 1e-4,
-) -> float:
+) -> np.ndarray:
     """Computes a item interpolation for temporal vectors defined either by item_indices, some tags at these item indices (vectors_map), and vectors at those tags.
 
-    Parameters
-    ----------
-    item : float
-        the input item at which the interpolation is required.
-    item_indices : np.ndarray
-        the items where the available data is defined, of size (numberOfTimeIndices).
-    vectors : np.ndarray or dict
-        the available data, of size (numberOfVectors, numberOfDofs).
-    vectors_map : list
[…]
