### 2025-08-28 0b8d6f0
:recycle: Refactoring Sample
