+++ src/plaid/containers/collections.py
@@ -1,15 +1,44 @@
 """Module for implementing collections of features within a Sample."""
 
+import copy
 import logging
+from pathlib import Path
 from typing import Optional, Union
 
-from plaid.types import Scalar
+import CGNS.MAP as CGM
+import CGNS.PAT.cgnskeywords as CGK
+import CGNS.PAT.cgnslib as CGL
+import CGNS.PAT.cgnsutils as CGU
+import numpy as np
+from CGNS.PAT.cgnsutils import __CHILDREN__, __NAME__
+
+from plaid.constants import (
+    CGNS_ELEMENT_NAMES,
+)
+from plaid.containers.sample import Sample
+from plaid.types import CGNSLink, CGNSNode, CGNSTree, Field, Scalar
+from plaid.utils import cgns_helper as CGH
 
 logger = logging.getLogger(__name__)
 logging.basicConfig(
     format="[%(asctime)s:%(levelname)s:%(filename)s:%(funcName)s(%(lineno)d)]:%(message)s",
     level=logging.INFO,
 )
+
+
+def _read_index(pyTree: list, dim: list[int]):
+    """Read Index Array or Index Range from CGNS.
+
+    Args:
+        pyTree (list): CGNS node which has a child Index to read
+        dim (list): dimensions of the coordinates
+
+    Returns:
+        indices
+    """
+    a = _read_index_array(pyTree)
+    b = _read_index_range(pyTree, dim)
+    return np.hstack((a, b))
 
 
 def _check_names(names: Union[str, list[str]]):
@@ -30,11 +59,74 @@ def _check_names(names: Union[str, list[str]]):
             )
 
 
+def _read_index_array(pyTree: list):
+    """Read Index Array from CGNS.
+
+    Args:
+        pyTree (list): CGNS node which has a child of type IndexArray_t to read
+
+    Returns:
+        indices
+    """
+    indexArrayPaths = CGU.getPathsByTypeSet(pyTree, ["IndexArray_t"])
+    res = []
+    for indexArrayPath in indexArrayPaths:
+        data = CGU.getNodeByPath(pyTree, indexArrayPath)
+        if data[1] is None:  # pragma: no cover
+            continue
+        else:
+            res.extend(data[1].ravel())
+    return np.array(res, dtype=int).ravel()
+
+
+def _read_index_range(pyTree: list, dim: list[int]):
+    """Read Index Range from CGNS.
+
+    Args:
+        pyTree (list): CGNS node which has a child of type IndexRange_t to read
+        dim (list[str]): dimensions of the coordinates
+
+    Returns:
+        indices
+    """
+    indexRangePaths = CGU.getPathsByTypeSet(pyTree, ["IndexRange_t"])
+    res = []
+
+    for indexRangePath in indexRangePaths:  # Is it possible there are several ?
+        indexRange = CGU.getValueByPath(pyTree, indexRangePath)
+
+        if indexRange.shape == (3, 2):  # 3D  # pragma: no cover
+            for k in range(indexRange[:, 0][2], indexRange[:, 1][2] + 1):
+                for j in range(indexRange[:, 0][1], indexRange[:, 1][1] + 1):
+                    global_id = (
+                        np.arange(indexRange[:, 0][0], indexRange[:, 1][0] + 1)
+                        + dim[0] * (j - 1)
+                        + dim[0] * dim[1] * (k - 1)
+                    )
+                    res.extend(global_id)
+
+        elif indexRange.shape == (2, 2):  # 2D  # pragma: no cover
+            for j in range(indexRange[:, 0][1], indexRange[:, 1][1]):
+                for i in range(indexRange[:, 0][0], indexRange[:, 1][0]):
+                    global_id = i + dim[0] * (j - 1)
+                    res.append(global_id)
[…]
