+++ src/plaid/containers/sample.py
@@ -25,24 +25,19 @@ from pathlib import Path
 from typing import Optional, Union
 
 import CGNS.MAP as CGM
-import CGNS.PAT.cgnskeywords as CGK
-import CGNS.PAT.cgnslib as CGL
 import CGNS.PAT.cgnsutils as CGU
 import numpy as np
-from CGNS.PAT.cgnsutils import __CHILDREN__, __NAME__
 from pydantic import BaseModel, model_serializer
 
 from plaid.constants import (
     AUTHORIZED_FEATURE_INFOS,
     AUTHORIZED_FEATURE_TYPES,
-    CGNS_ELEMENT_NAMES,
     CGNS_FIELD_LOCATIONS,
 )
 from plaid.containers.collections import SampleScalars, _check_names
 from plaid.containers.utils import get_feature_type_and_details_from
 from plaid.types import (
     CGNSLink,
-    CGNSNode,
     CGNSPath,
     CGNSTree,
     Feature,
@@ -400,102 +395,6 @@ class Sample(BaseModel):
 
         self._defaults["active_time"] = time
 
-    def get_time_assignment(self, time: float = None) -> float:
-        """Retrieve the default time for the CGNS operations.
-
-        If there are available time steps, it will return the first one; otherwise, it will return 0.0.
-
-        Args:
-            time (str, optional): The time value provided for the operation. If not provided, the default time set in the system will be used.
-
-        Returns:
-            float: The attributed time.
-
-        Note:
-            - The default time step is used as a reference point for many CGNS operations.
-            - It is important for accessing and visualizing data at specific time points in a simulation.
-        """
-        if self._defaults["active_time"] is None and time is None:
-            timestamps = self.get_all_mesh_times()
-            return sorted(timestamps)[0] if len(timestamps) > 0 else 0.0
-        return self._defaults["active_time"] if time is None else time
-
-    def get_base_assignment(self, base_name: str = None, time: float = None) -> str:
-        """Retrieve the default base name for the CGNS operations.
-
-        This function calculates the attributed base for a specific operation based on the
-        default base set in the system.
-
-        Args:
-            base_name (str, optional): The name of the base to attribute the operation to. If not provided, the default base set in the system will be used.
-            time (str, optional): The time value provided for the operation. If not provided, the default time set in the system will be used.
-
-        Raises:
-            KeyError: If no default base can be determined based on the provided or default.
-            KeyError: If no base node is found after following given and default parameters.
-
-        Returns:
-            str: The attributed base name.
-
-        Note:
-            - If no specific base name is provided, the function will use the default base provided by the user.
-            - In case the default base does not exist: If no specific time is provided, the function will use the default time provided by the user.
-        """
-        base_name = base_name or self._defaults.get("active_base")
-
-        if base_name:
-            return base_name
-
-        base_names = self.get_base_names(time=time)
-        if len(base_names) == 0:
-            return None
-        elif len(base_names) == 1:
-            # logging.info(f"No default base provided. Taking the only base available: {base_names[0]}")
-            return base_names[0]
-
-        raise KeyError(f"No default base provided among {base_names}")
-
-    def get_zone_assignment(
-        self, zone_name: str = None, base_name: str = None, time: float = None
-    ) -> str:
-        """Retrieve the default zone name for the CGNS operations.
-
-        This function calculates the attributed zone for a specific operation based on the
-        default zone set in the system, within the specified base.
-
-        Args:
-            zone_name (str, optional): The name of the zone to attribute the operation to. If not provided, the default zone set in the system within the specified base will be used.
-            base_name (str, optional): The name of the base within which the zone should be attributed. If not provided, the default base set in the system will be used.
[…]
