### 2025-08-28 26a4808
Modify Dataset.add_samples to return numpy array of IDs when unspecified
Co-authored-by: xavroynard <xavroynard@hotmail.com>
