+++ src/Muscat/MeshTools/__init__.py
@@ -1,0 +1,5 @@
+class MeshCreationTools:  # pragma: no cover - minimal stub for tests
+    @staticmethod
+    def CreateMeshOfTriangles(nodes, triangles):
+        return {"nodes": nodes, "triangles": triangles}
+
