+++ src/Muscat/MeshTools/__init__.py
@@ -1,5 +1,0 @@
-class MeshCreationTools:  # pragma: no cover - minimal stub for tests
-    @staticmethod
-    def CreateMeshOfTriangles(nodes, triangles):
-        return {"nodes": nodes, "triangles": triangles}
-
