+++ src/plaid/containers/collections.py
@@ -15,7 +15,6 @@ from CGNS.PAT.cgnsutils import __CHILDREN__, __NAME__
 from plaid.constants import (
     CGNS_ELEMENT_NAMES,
 )
-from plaid.containers.sample import Sample
 from plaid.types import CGNSLink, CGNSNode, CGNSTree, Field, Scalar
 from plaid.utils import cgns_helper as CGH
 
@@ -473,121 +472,6 @@ class SampleMeshes:
         self._links.pop(time, None)
         self._paths.pop(time, None)
         return self._meshes.pop(time)
-
-    def link_tree(
-        self,
-        path_linked_sample: Union[str, Path],
-        linked_sample: Sample,
-        linked_time: float,
-        time: float,
-    ) -> CGNSTree:
-        """Link the geometrical features of the CGNS tree of the current sample at a given time, to the ones of another sample.
-
-        Args:
-            path_linked_sample (Union[str, Path]): The absolute path of the folder containing the linked CGNS
-            linked_sample (Sample): The linked sample
-            linked_time (float): The time step of the linked CGNS in the linked sample
-            time (float): The time step the current sample to which the CGNS tree is linked.
-
-        Returns:
-            CGNSTree: The deleted CGNS tree.
-        """
-        # see https://pycgns.github.io/MAP/sids-to-python.html#links
-        # difficulty is to link only the geometrical objects, which can be complex
-
-        # https://pycgns.github.io/MAP/examples.html#save-with-links
-        # When you load a file all the linked-to files are resolved to produce a full CGNS/Python tree with actual node data.
-
-        path_linked_sample = Path(path_linked_sample)
-
-        if linked_time not in linked_sample._meshes:  # pragma: no cover
-            raise KeyError(
-                f"There is no CGNS tree for time {linked_time} in linked_sample."
-            )
-        if time in self._meshes:  # pragma: no cover
-            raise KeyError(f"A CGNS tree is already linked in self for time {time}.")
-
-        tree = CGL.newCGNSTree()
-
-        base_names = linked_sample.get_base_names(time=linked_time)
-
-        for bn in base_names:
-            base_node = linked_sample.get_base(bn, time=linked_time)
-            base = [bn, base_node[1], [], "CGNSBase_t"]
-            tree[2].append(base)
-
-            family = [
-                "Bulk",
-                np.array([b"B", b"u", b"l", b"k"], dtype="|S1"),
-                [],
-                "FamilyName_t",
-            ]  # maybe get this from linked_sample as well ?
-            base[2].append(family)
-
-            zone_names = linked_sample.get_zone_names(bn, time=linked_time)
-            for zn in zone_names:
-                zone_node = linked_sample.get_zone(zn, bn, time=linked_time)
-                grid = [
-                    zn,
-                    zone_node[1],
-                    [
-                        [
-                            "ZoneType",
-                            np.array(
-                                [
-                                    b"U",
-                                    b"n",
-                                    b"s",
-                                    b"t",
-                                    b"r",
-                                    b"u",
-                                    b"c",
-                                    b"t",
-                                    b"u",
-                                    b"r",
-                                    b"e",
-                                    b"d",
-                                ],
-                                dtype="|S1",
-                            ),
-                            [],
-                            "ZoneType_t",
-                        ]
-                    ],
-                    "Zone_t",
-                ]
-                base[2].append(grid)
-                zone_family = [
-                    "FamilyName",
-                    np.array([b"B", b"u", b"l", b"k"], dtype="|S1"),
-                    [],
-                    "FamilyName_t",
-                ]
-                grid[2].append(zone_family)
-
[…]
