### 2025-08-29 c18e5c1
feat(dataset.py) improve coverage of summarize_features and check_feature_completeness
