+++ src/plaid/containers/dataset.py
@@ -1029,167 +1029,179 @@ class Dataset(object):
         shutil.rmtree(save_dir)
 
     def summarize_features(self) -> str:
-      """Show the name of each feature and the number of samples containing it.
-    
-      Returns:
-          str: A summary of features across the dataset.
-      """
-      summary = "Dataset Feature Summary:\n"
-      summary += "=" * 50 + "\n"
-    
-      if len(self._samples) == 0:
-          return summary + "No samples in dataset.\n"
-    
-      # Collect all feature names across all samples
-      all_scalar_names = set()
-      all_ts_names = set()
-      all_field_names = set()
-    
-      # Count occurrences of each feature
-      scalar_counts = {}
-      ts_counts = {}
-      field_counts = {}
-    
-      for sample_id, sample in self._samples.items():
-          # Scalars
-          scalar_names = sample.get_scalar_names()
-          all_scalar_names.update(scalar_names)
-          for name in scalar_names:
-              scalar_counts[name] = scalar_counts.get(name, 0) + 1
-        
-          # Time series
-          ts_names = sample.get_time_series_names()
-          all_ts_names.update(ts_names)
-          for name in ts_names:
-              ts_counts[name] = ts_counts.get(name, 0) + 1
-        
-          # Fields
-          times = sample.get_all_mesh_times()
-          for time in times:
-              base_names = sample.get_base_names(time=time)
-              for base_name in base_names:
-                  zone_names = sample.get_zone_names(base_name=base_name, time=time)
-                  for zone_name in zone_names:
-                      field_names = sample.get_field_names(zone_name=zone_name, base_name=base_name, time=time)
-                      all_field_names.update(field_names)
-                      for name in field_names:
-                          field_counts[name] = field_counts.get(name, 0) + 1
-    
-      total_samples = len(self._samples)
-    
-      # Scalars summary
-      summary += f"Scalars ({len(all_scalar_names)} unique):\n"
-      if all_scalar_names:
-          for name in sorted(all_scalar_names):
-              count = scalar_counts.get(name, 0)
-              summary += f"  - {name}: {count}/{total_samples} samples ({count/total_samples*100:.1f}%)\n"
-      else:
-          summary += "  None\n"
-      summary += "\n"
-    
-      # Time series summary
-      summary += f"Time Series ({len(all_ts_names)} unique):\n"
-      if all_ts_names:
-          for name in sorted(all_ts_names):
-              count = ts_counts.get(name, 0)
-              summary += f"  - {name}: {count}/{total_samples} samples ({count/total_samples*100:.1f}%)\n"
-      else:
-          summary += "  None\n"
-      summary += "\n"
-    
-      # Fields summary
-      summary += f"Fields ({len(all_field_names)} unique):\n"
-      if all_field_names:
-          for name in sorted(all_field_names):
-              count = field_counts.get(name, 0)
-              summary += f"  - {name}: {count}/{total_samples} samples ({count/total_samples*100:.1f}%)\n"
-      else:
-          summary += "  None\n"
-    
-      return summary
+        """Show the name of each feature and the number of samples containing it.
+
+        Returns:
+            str: A summary of features across the dataset.
+        """
+        summary = "Dataset Feature Summary:\n"
+        summary += "=" * 50 + "\n"
+
+        if len(self._samples) == 0:
+            return summary + "No samples in dataset.\n"
+
+        # Collect all feature names across all samples
+        all_scalar_names = set()
+        all_ts_names = set()
+        all_field_names = set()
+
+        # Count occurrences of each feature
+        scalar_counts = {}
+        ts_counts = {}
+        field_counts = {}
+
+        for sample_id, sample in self._samples.items():
+            # Scalars
[…]
