### 2025-08-30 f06aae9
feat(sample) add summarize and check_completeness functions
