### 2025-08-30 bd9ae4a
fix(sample) correct comment
