### 2025-08-30 c9a34d6
feat(dataset.py) add examples in docstring and incresae coverage to 100%
