### 2025-08-30 c984c8d
feat(sample.py) add examples in docstring
