### 2025-08-31 714700b
feat(dataset.py) from_list_of_samples: add optional ids argument
