+++ src/plaid/containers/dataset.py
@@ -1029,17 +1029,20 @@ class Dataset(object):
         shutil.rmtree(save_dir)
 
     @classmethod
-    def from_list_of_samples(cls, list_of_samples: list[Sample]) -> Self:
+    def from_list_of_samples(
+        cls, list_of_samples: list[Sample], ids: Optional[list[int]] = None
+    ) -> Self:
         """Initialise a dataset from a list of samples.
 
         Args:
             list_of_samples (list[Sample]): The list of samples.
+            ids (list[int], optional): An optional list of IDs for the new samples. If not provided, the IDs will be automatically generated based on the current number of samples in the dataset.
 
         Returns:
             Self: The intialized dataset (Dataset).
         """
         instance = cls()
-        instance.add_samples(list_of_samples)
+        instance.add_samples(list_of_samples, ids)
         return instance
 
     @classmethod
