+++ tests/conftest.py
@@ -146,7 +146,7 @@ def tree(nodes, triangles, vertex_field, cell_center_field, nodal_tags):
 def sample_with_tree(tree: CGNSTree) -> Sample:
     """Generate a Sample objects with a tree."""
     sample = Sample()
-    sample.add_tree(tree)
+    sample._meshes.add_tree(tree)
     return sample
 
 
