+++ src/plaid/utils/stats.py
@@ -434,9 +434,11 @@ class Stats:
             sample (Sample): Sample containing field data
             data_dict (dict[str, list]): Dictionary to store processed data
         """
-        for time in sample.get_all_mesh_times():
-            for base_name in sample.get_base_names(time=time):
-                for zone_name in sample.get_zone_names(base_name=base_name, time=time):
+        for time in sample._meshes.get_all_mesh_times():
+            for base_name in sample._meshes.get_base_names(time=time):
+                for zone_name in sample._meshes.get_zone_names(
+                    base_name=base_name, time=time
+                ):
                     for field_name in sample.get_field_names(
                         zone_name=zone_name, base_name=base_name, time=time
                     ):
