+++ examples/convert_users_data_example.py
@@ -121,7 +121,7 @@ for cgns_tree in tqdm(CGNS_meshes):
     # Add CGNS Meshe to samples with specific time steps
     sample = Sample()
 
-    sample.add_tree(cgns_tree)
+    sample._meshes.add_tree(cgns_tree)
 
     # Add random scalar values to the sample
     for sname in in_scalars_names:
