### 2025-09-02 bb47eac
Fix some examples
