### 2025-09-09 #8830 @ekourlit
Hi @Nelly-Barret and thank you for your review!

JOSS do not have a page limit for the paper but a _soft_ 1000 words limit. I would suggest adding just a sentence, referencing the examples found at the software documentation (https://plaid-lib.readthedocs.io/en/stable/source/notebooks.html). Would this be enough @Nelly-Barret ?
