### 2026-01-22 #8830 @casenave
Dear @ekourlit,

The first reviewer has accepted the paper as-is, and no changes were requested.
Could you please let us know whether this is sufficient to proceed, or if any further action is required from our side?

Many thanks!
