### 2026-01-27 #8830 @karanprime
Dear @casenave and authors,

This work presents Physics-Learning AI Datamodel (PLAID), a framework for representing physics simulation datasets, with a focus towards benchmarking surrogate (especially ML) models.
I went through the software paper, codebase in the repo (https://github.com/PLAID-lib/plaid) and documentation (https://plaid-lib.readthedocs.io/en/latest/).

The software paper is well-written, and the gap (lack of standard format for ML-PDE datasets) is clearly identified.
This datamodel would be useful to physics-informed ML (PIML) practictioners because of its flexible structure supporting heterogeneous meshes and time-dependent data. Utilities such as interpolation, online statistics and easy hugging face support are useful additions to the core datamodel. While not necessary at this stage, the paper could be made stronger with a more comprehensive state-of-the-field paragraph, mentioning current ad-hoc workflows using formats like HDF5 or VTK, and the presence of PDE benchmarks such as the Well, CoDBench, Mechanical MNIST etc. These benchmarks provide standardized metrics for new PIML algorithms but do not provide a standard format, which is where PLAID would be useful.

The repo and documentation are well-organized. I installed the library and played around with the examples. All (399) `pytest` tests pass too.

---

I am a bit confused about these two points in the checklist:

Data sharing: If the paper contains original data, data are accessible to the reviewers. If the paper contains no original data, please check this item.
- checking this because the Hugging Face and Zenodo pages consist of some datasets that the authors have uploaded.

Reproducibility:  If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.
- I am checking this because the original results/datasets are not the focus of this paper. The datamodel itself is the focus of this software paper.

Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)
- There are no performance claims for this datamodel.

---

Reproduction notes (on my local Macbook Pro):

On macos apple silicon (osx-arm64), `conda install -c conda-forge plaid` currently fails due to missing platform builds of dependencies (e.g., pycgns is not available for osx-arm64). The installation instructions would be stronger if they documented this limitation and provided a supported workaround (e.g., an osx-64/Rosetta conda environment or a fully specified pip + system dependency path).

For the development engvironment, I installed `hdf5` through homebrew and ran `pip install -e .`

To get around the apple silicon conda issues, I ran conda commands prepending `CONDA_SUBDIR=osx-64` (eg `CONDA_SUBDIR=osx-64 conda install -c conda-forge muscat-core`)

---

Overall, this work would make a great addition to JOSS. I look forward to further support for Pytorch, and maybe the authors could even interoperate  with popular frameworks such as Physics Nemo or Neural Operators.

Thanks,
Karan
