### 2026-01-29 #8830 @casenave
Dear @karanprime,

Thank you very much for your review.

We have updated the paper and the installation instructions based on your feedback.

The PDF of the paper can be downloaded from the previous comment, and all modifications are visible in this [PR](https://github.com/PLAID-lib/plaid/pull/307/changes).

Best regards,
Fabien Casenave
